% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmCharts_plot.R
\name{add_dataloader_dependency}
\alias{add_dataloader_dependency}
\title{Add dataloader dependency}
\usage{
add_dataloader_dependency(widget)
}
\arguments{
\item{widget}{An htmlwidget}
}
\value{
Return the updated htmlwidget.
}
\description{
Add the 'dataloader' dependency to an htmlwidget.
You can only manipulate the htmlwidget if you call the method 'plot' with an rAmChart.
}

