% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_AmCharts.R
\name{add_responsive_dependency}
\alias{add_responsive_dependency}
\title{Add responsive dependency}
\usage{
add_responsive_dependency(widget)
}
\arguments{
\item{widget}{An htmlwidget.}
}
\value{
Return an updated htmlwidget with the dependency.
}
\description{
Add the 'responsive' dependency to an htmlwidget.
You can only manipulate the htmlwidget if you call the method 'plot' with an rAmChart.
}

