% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_AmCharts.R
\name{add_theme_dependency}
\alias{add_theme_dependency}
\title{Add theme dependency}
\usage{
add_theme_dependency(widget, theme_js = c("light.js", "patterns.js",
  "dark.js", "chalk.js"))
}
\arguments{
\item{widget}{An htmlwidget.}

\item{theme_js}{A character indicating the JS file dependency.}
}
\value{
Return the updated htmlwidget.
}
\description{
Add the 'theme' dependency to an htmlwidget.
You can only manipulate the htmlwidget if you call the method 'plot' with an rAmChart.
}
\examples{
\donttest{
library(pipeR)
amPlot(1:10, theme = "dark") \%>>\% plot() \%>>\% add_theme_dependency("light.js")
}

}
