% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addArrow}
\alias{addArrow}
\alias{addArrow,AmChart-method}
\title{Setter}
\usage{
addArrow(.Object, arrow = NULL, ...)

\S4method{addArrow}{AmChart}(.Object, arrow = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}}

\item{arrow}{(optional) Object of class \code{\linkS4class{GaugeArrow}}.}

\item{...}{properties of \code{\linkS4class{GaugeArrow}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
print(addArrow(.Object = amAngularGaugeChart(), alpha = 1))
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{GaugeArrow}} S4 class
}

