% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addAxe}
\alias{addAxe}
\alias{addAxe,AmChart-method}
\title{Setter}
\usage{
addAxe(.Object, axe = NULL, ...)

\S4method{addAxe}{AmChart}(.Object, axe = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}}

\item{axe}{object of class \code{\linkS4class{GaugeAxis}}.}

\item{...}{properties of \code{\linkS4class{GaugeAxis}}}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
addAxe(.Object = amChart(), bandAlpha = 1)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{GaugeAxis}} S4 class
}

