% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addGraph}
\alias{addGraph}
\alias{addGraph,AmChart-method}
\title{Setter}
\usage{
addGraph(.Object, amGraph = NULL, ...)

\S4method{addGraph}{AmChart}(.Object, amGraph = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{amGraph}{object of class \code{\linkS4class{AmGraph}}, default \code{NULL}.}

\item{...}{properties of \code{\linkS4class{AmGraph}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
addGraph(.Object = amChart(), amGraph = amGraph(balloonText = "balloonText", "type" = "column"))
addGraph(.Object = amChart(), balloonText = "balloonText", "type" = "column")
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{AmGraph}} S4 class
}

