% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addGuide,AmChart-method}
\alias{addGuide,AmChart-method}
\title{Setter}
\usage{
\S4method{addGuide}{AmChart}(.Object, guide = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{guide}{object of class \code{\linkS4class{Guide}}, default \code{NULL}}

\item{...}{properties of \code{\linkS4class{Guide}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\examples{
addGuide(.Object = amChart(), fillAlpha = .1)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{Guide}} S4 class
}

