% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addLabel}
\alias{addLabel}
\alias{addLabel,AmChart-method}
\title{Setter}
\usage{
addLabel(.Object, label = NULL, ...)

\S4method{addLabel}{AmChart}(.Object, label = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}}

\item{label}{Object of class \code{\linkS4class{Label}}.}

\item{...}{properties of class \code{\linkS4class{Label}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
addLabel(.Object = amChart(), text = "balloonText")
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{Label}} S4 class
}

