% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmObject.R
\docType{methods}
\name{addListener}
\alias{addListener}
\title{Setter for Listener}
\usage{
\S4method{addListener}{AmObject,character,character}(.Object, name, expression)
}
\arguments{
\item{\code{.Object}:}{Object of of inherited class \code{\linkS4class{AmObject}}.}

\item{\code{name}:}{Object of class \code{character} containing the name of the listener.}

\item{\code{expression}:}{Object of class \code{character} containing the associated function event.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter for Listener
}
\examples{
library(pipeR)
amChart() \%>>\% addListener("select", "function onSelect (properties) {
     alert('selected nodes: ' + properties.nodes);}")
amLegend() \%>>\% addListener("select", "function onSelect (properties) {
     alert('selected nodes: ' + properties.nodes);}")
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{AmLegend}} S4 class

Other AmChart methods: \code{\link{addArrow}};
  \code{\link{addAxe}}; \code{\link{addGraph}};
  \code{\link{addGuide}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}}; \code{\link{addSegment}};
  \code{\link{addSubData}}; \code{\link{addTitle}};
  \code{\link{addTrendLine}}; \code{\link{addValueAxes}};
  \code{\link{initialize,AmChart-method}};
  \code{\link{initialize,AmStockChart-method}};
  \code{\link{setAllLabels}}; \code{\link{setArrows}};
  \code{\link{setAxes}}; \code{\link{setBalloon}};
  \code{\link{setCategoryAxis}};
  \code{\link{setCategoryField}};
  \code{\link{setChartCursor}};
  \code{\link{setChartScrollbar}};
  \code{\link{setCreditsPosition}};
  \code{\link{setDataProvider}},
  \code{\link{setDataProvider}}; \code{\link{setExport}};
  \code{\link{setGraphs}}; \code{\link{setGraph}},
  \code{\link{setGraph}}; \code{\link{setGuides}};
  \code{\link{setLegend}};
  \code{\link{setSubChartProperties}};
  \code{\link{setTitles}}; \code{\link{setTrendLines}};
  \code{\link{setValueAxes}}; \code{\link{setValueAxis}}

Other AmChart setters: \code{\link{addArrow}};
  \code{\link{addAxe}}; \code{\link{addGraph}};
  \code{\link{addGuide}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}}; \code{\link{addTitle}};
  \code{\link{addTrendLine}}; \code{\link{addValueAxes}};
  \code{\link{setAllLabels}}; \code{\link{setArrows}};
  \code{\link{setAxes}}; \code{\link{setBalloon}};
  \code{\link{setCategoryAxis}};
  \code{\link{setCategoryField}};
  \code{\link{setChartCursor}};
  \code{\link{setChartScrollbar}};
  \code{\link{setCreditsPosition}};
  \code{\link{setDataProvider}},
  \code{\link{setDataProvider}}; \code{\link{setExport}};
  \code{\link{setGraphs}}; \code{\link{setGraph}},
  \code{\link{setGraph}}; \code{\link{setGuides}};
  \code{\link{setLegend}};
  \code{\link{setSubChartProperties}};
  \code{\link{setTitles}}; \code{\link{setTrendLines}};
  \code{\link{setValueAxes}}; \code{\link{setValueAxis}}
}

