% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PeriodSelector.R
\docType{methods}
\name{addPeriod}
\alias{addPeriod}
\title{Add a period}
\usage{
\S4method{addPeriod}{PeriodSelector}(.Object, ...)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{PeriodSelector}}.}

\item{\code{...}:}{Properties of period.}
}
\value{
The updated object of class \code{\linkS4class{PeriodSelector}}.
}
\description{
Add a period
}
\examples{
library(pipeR)
periodSelector() \%>>\% addPeriod(  period = "MM", selected = TRUE, count = 1, label= "1 month" )
}
\seealso{
\code{\linkS4class{PeriodSelector}} S4 class
}

