% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addSegment}
\alias{addSegment}
\alias{addSegment,AmChart,numeric-method}
\title{Add a segment to a category of AmChart}
\usage{
addSegment(.Object, categoryIDs, sgts)

\S4method{addSegment}{AmChart,numeric}(.Object, categoryIDs, sgts)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}}

\item{categoryIDs}{Object of class \code{numeric}.}

\item{sgts}{Object of class \code{data.frame}
( or \code{list} of \code{data.frame} for multiple add ).}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Add a segment to a category of AmChart
}
\details{
Use this methode in case of an AmChart.
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,categoryIDs = numeric}: 
}}
\examples{
library(pipeR)
amGanttChart(segmentsField = "segments"
) \%>>\% setDataProvider(data.frame(category = c( "John", "Julia"))
) \%>>\% addSegment(1, data.frame(start = 7, duration = 2:3, task = c("Task #1", "Task #2"))
) \%>>\% addSegment(2, data.frame(start = 10, duration = 2:3, task = c("Task #1", "Task #2")))

ls <- list( data.frame(start = 7, duration = 2:3, task = c("Task #1", "Task #2")),
data.frame(start = 10, duration = 2:3, task = c("Task #1", "Task #2")))
amGanttChart(segmentsField = "segments"
) \%>>\% setDataProvider(data.frame(category = c( "John", "Julia"))
) \%>>\% addSegment( 1:2,  ls)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class
}

