% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{addStockEvent,DataSet-method}
\alias{addStockEvent,DataSet-method}
\title{Add stockEvent}
\usage{
\S4method{addStockEvent}{DataSet}(.Object, stockEvent = NULL, ...)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{DataSet}}.}

\item{stockEvent}{(optional) Object of class \code{\linkS4class{StockEvent}}}

\item{...}{Properties of \code{\linkS4class{StockEvent}}
See \url{http://docs.amcharts.com/3/javascriptstockchart/StockEvent}}
}
\description{
Add stockEvent
}
\examples{
dataSet()
}

