% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StockPanel.R
\docType{methods}
\name{addStockGraph}
\alias{addStockGraph}
\title{Setter}
\usage{
\S4method{addStockGraph}{StockPanel}(.Object, stockGraph = NULL, ...)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{\code{stockGraph}:}{(optionnal) Object of class \code{\linkS4class{amGraph}}.
Use stockGraph(*) constructor}

\item{\code{...}:}{Properties of the \code{\linkS4class{amGraph}} to add.}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter
}
\examples{
library(pipeR)
stockPanel() \%>>\% addStockGraph(comparable = FALSE)
stockPanel() \%>>\% addStockGraph( stockGraph(comparable = FALSE) )
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

\code{\linkS4class{AmGraph}} S4 class
}

