% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addSubData}
\alias{addSubData}
\alias{addSubData,AmChart,numeric,data.frame-method}
\title{Add subData for drilldrown}
\usage{
addSubData(.Object, categoryIDs, data)

\S4method{addSubData}{AmChart,numeric,data.frame}(.Object, categoryIDs, data)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{categoryIDs}{object of class \code{numeric} (vector or simple value).
Indicates corresponding indice(s) of the \code{dataProvider} where to add the data.}

\item{data}{object of class \code{data.frame}. Data to draw at the second level,
after clicking on the serial / column.}
}
\description{
This method allows to add subdata for a chart with drilldown.
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,categoryIDs = numeric,data = data.frame}: 
}}
\examples{
library(pipeR)
amChart(dataProvider = data.frame(a = 1:5, b = 6:10)) \%>>\%
addSubData(3, data.frame(a = 1:10, b = 11:20))
}

