% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addTrendLine}
\alias{addTrendLine}
\alias{addTrendLine,AmChart-method}
\title{Add a trendLine}
\usage{
addTrendLine(.Object, trendLine = NULL, ...)

\S4method{addTrendLine}{AmChart}(.Object, trendLine = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{trendLine}{object of class \code{\linkS4class{TrendLine}}, default \code{NULL}.}

\item{...}{properties of class \code{\linkS4class{TrendLine}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Add a trendLine
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
addTrendLine(.Object = amChart(), initialValue = 1, finalValue = 11)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{TrendLine}} S4 class
}

