% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addValueAxes}
\alias{addValueAxes}
\alias{addValueAxes,AmChart-method}
\title{Add a ValueAxes}
\usage{
addValueAxes(.Object, valueAxis = NULL, ...)

\S4method{addValueAxes}{AmChart}(.Object, valueAxis = NULL, ...)
}
\arguments{
\item{.Object}{object of class \linkS4class{AmChart}.}

\item{valueAxis}{object of class \linkS4class{ValueAxis}, default \code{NULL}.}

\item{...}{properties of \linkS4class{ValueAxis}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Add a ValueAxes
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
addValueAxes(.Object = amChart(), axisTitleOffset = 12, tickLength = 10)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{ValueAxis}} S4 class
}

