% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amBar.R
\name{amBar}
\alias{amBar}
\title{Plotting bar chart using rAmCharts}
\usage{
amBar(x, y, data, main = "", mainSize = 15, xlab = "", ylab = "",
  groups_color = NULL, horiz = FALSE, stack_type = "none",
  layered = FALSE, legend = FALSE, legend_side = "right",
  show_values = FALSE, third_dim = FALSE)
}
\arguments{
\item{x}{\code{character} column name for x-axis or \code{numeric} 
the number of the corresponding column.}

\item{y}{\code{character} column name for y-axis or \code{numeric} 
the number of the corresponding column. If you want to display a grouped barchart
or a stacked barchart, y is a vector of characters or numerics.}

\item{data}{\code{data.frame} dataframe with values to display.
You can add a column "color" (character, colors in hexadecimal). You can
also add a column "description" (character) containing the text you want to
display when mouse is on the graphic ('<br>' for a new line).}

\item{main}{\code{character} title of the graph.}

\item{mainSize}{\code{numeric} size of the title of the graph.}

\item{xlab}{\code{character} label for x-axis.}

\item{ylab}{\code{character} label for y-axis.}

\item{groups_color}{\code{character} vector of colors in hexadecimal, 
same length as y.}

\item{horiz}{\code{boolean} TRUE for an horizontal chart, FALSE for a vertical one}

\item{stack_type}{\code{character}, "regular" if you wish stacked bars, "100" if
you want 100 percent stacked bars. Default is set to "none".}

\item{layered}{\code{boolean} TRUE for layered. If TRUE, stack_type must be set
to "none".}

\item{legend}{\code{boolean} TRUE to display legend.}

\item{legend_side}{\code{character} either "left" or "right" if legend is set to TRUE.}

\item{show_values}{\code{boolean} TRUE to display values.}

\item{third_dim}{\code{boolean} if TRUE, chart is displayed in 3D}
}
\description{
amBar computes a pie chart of the given value.
}
\examples{


#Basic Example : column chart
data_bar <- data.frame(country = c("USA", "China", "Japan", "Germany", 
                                    "UK", "France", "India", "Spain",
                                    "Netherlands", "Russia", "South Korea",
                                    "Canada"),
                        visits = c(3025, 1882, 1809, 1322, 1122, 1114, 
                                   984, 711, 665, 580, 443, 441),
                        color = c("#FF0F00", "#FF6600", "#FF9E01", "#FCD202",
                                  "#F8FF01", "#B0DE09", "#04D215", "#0D8ECF",
                                  "#0D52D1", "#2A0CD0", "#8A0CCF", "#CD0D74"),
                        stringsAsFactors = FALSE)
amBar(x = "country", y = "visits", data = data_bar)

#horizontal bar
amBar(x = "country", y = "visits", data = data_bar, horiz = TRUE)

#3D bar
amBar(x = "country", y = "visits", data = data_bar, third_dim = TRUE)

#display values
amBar(x = "country", y = "visits", data = data_bar, show_values = TRUE)

#grouped columns
data_gbar <- data.frame(year = c("2005", "2006", "2007", "2008", "2009"),
                        income = c(23.5, 26.2, 30.1, 29.5, 24.6),
                        expenses = c(18.1, 22.8, 23.9, 25.1, 25),
                        stringsAsFactors = FALSE)

amBar(x = "year", y = c("income", "expenses"), data = data_gbar)

#add legend
amBar(x = "year", y = c("income", "expenses"), data = data_gbar, legend = TRUE)

#change groups colors
amBar(x = "year", y = c("income", "expenses"), data = data_gbar, 
      groups_color = c("#87cefa", "#c7158"), legend = TRUE)

#stacked bars
amBar(x = "year", y = c("income", "expenses"), data = data_gbar, stack_type = "regular")

#100\% stacked bars
amBar(x = "year", y = c("income", "expenses"), data = data_gbar, stack_type = "100")

#layered bars
amBar(x = "year", y = c("income", "expenses"), data = data_gbar, layered = TRUE)


}

