% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_shinyUtils.R
\docType{methods}
\name{amChartOutput}
\alias{amChartOutput}
\title{SHINY}
\usage{
\S4method{amChartOutput}{character,character}(outputId, type, width, height)
}
\arguments{
\item{amChart:}{an \code{\linkS4class{AmChart}} object}
}
\description{
Widget output function for use in Shiny
}

