% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_AmCharts.R
\docType{methods}
\name{setExport}
\alias{setExport}
\alias{setExport,AmCharts,logicalOrMissing-method}
\alias{setResponsive}
\alias{setResponsive,AmCharts,logicalOrMissing-method}
\title{Setters for AmChart and AmStockChart.}
\usage{
setExport(.Object, enabled = TRUE, ...)

\S4method{setExport}{AmCharts,logicalOrMissing}(.Object, enabled = TRUE, ...)

setResponsive(.Object, enabled = TRUE, ...)

\S4method{setResponsive}{AmCharts,logicalOrMissing}(.Object, enabled = TRUE,
  ...)
}
\arguments{
\item{.Object}{\linkS4class{AmChart} or \linkS4class{AmStockChart}.}

\item{enabled}{\code{logical}, TRUE to display the export button.}

\item{...}{Other properties that can be used depending on the setter.}
}
\description{
These methods can be used both for AmChart and AmStockChart.
There are general for some first-level properties.
}
\examples{
\donttest{
# Dummy examples
setExport(amPlot(1:10))
setExport(amStockChart())
}
\donttest{
# Dummy examples
setResponsive(amSerialChart())
setResponsive(amStockChart())
}
}

