% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amChartsAPI.R
\name{amChartsAPI}
\alias{amChartsAPI}
\title{View full API of amcharts.js}
\usage{
amChartsAPI(viewer = getOption("viewer"))
}
\arguments{
\item{viewer}{: Set to NULL to open in a browser}
}
\description{
View full API of amcharts.js
}
\examples{

\dontrun{
amChartsAPI()
amChartsAPI(NULL)
}


}
\references{
See online documentation \url{http://datastorm-open.github.io/introduction_ramcharts/}
}
