% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amFunnel.R
\name{amFunnel}
\alias{amFunnel}
\title{Plotting funnel chart using rAmCharts}
\usage{
amFunnel(data, main = "Funnel Chart", mainSize = 15, inverse = FALSE,
  neck_height = NULL, neck_width = NULL, third_dim = FALSE,
  label_side = "right", margin_right = 160, margin_left = 15)
}
\arguments{
\item{data}{a data frame of at least 2 columns : value (numeric, positive),
and description (character). Your can add a third column "color" (character,
colors in hexadecimal)}

\item{main}{\code{character}, title of the graph.}

\item{mainSize}{\code{numeric}, size of the title of the graph.}

\item{inverse}{\code{boolean}, if TRUE, the funnel chart will be inversed.}

\item{neck_height}{\code{numeric} value between 0 and 100 : if a bottleneck
is desired, this value determines its heigh. Default to NULL}

\item{neck_width}{\code{numeric} value between 0 and 100 : if a bottleneck
is desired, this value determines its witdh. Default to NULL}

\item{third_dim}{\code{boolean} if TRUE, chart is displayed in 3D, only for
pyramid chart (without a bottleneck)}

\item{label_side}{\code{character} label position : "right" or "left"}

\item{margin_right}{\code{numeric} margin at the right side}

\item{margin_left}{\code{numeric} margin at the left side}
}
\description{
amFunnel computes a funnel chart of the given value.
}
\examples{
amFunnel(data = data.frame(description = c("Website visits", "Downloads",
                                                 "Requested price list",
                                                 "Contaced for more info",
                                                 "Purchased", "Contacted for support",
                                                 "Purchased additional products"),
                                value = c(300, 123, 98, 72, 80, 15, 8),
                                stringsAsFactors = FALSE),
               main = "Pyramid", mainSize = 15, inverse = TRUE)

amFunnel(data = data.frame(description = c("Website visits", "Downloads",
                                                 "Requested price list",
                                                 "Contaced for more info",
                                                 "Purchased", "Contacted for support",
                                                 "Purchased additional products"),
                                value = c(300, 123, 98, 72, 80, 15, 8),
                                stringsAsFactors = FALSE),
               main = "Reversed Pyramid", mainSize = 15, inverse = FALSE,
               label_side = "left", margin_right = 15, margin_left = 160)

amFunnel(data = data.frame(description = c("Website visits", "Downloads",
                                                 "Requested price list",
                                                 "Contaced for more info",
                                                 "Purchased", "Contacted for support",
                                                 "Purchased additional products"),
                                value = c(300, 123, 98, 72, 80, 15, 8),
                                stringsAsFactors = FALSE),
               neck_height = 30, neck_width = 40,
               main = "Funnel", mainSize = 15, inverse = FALSE)

amFunnel(data = data.frame(description = c("Website visits", "Downloads",
                                                 "Requested price list",
                                                 "Contaced for more info",
                                                 "Purchased", "Contacted for support",
                                                 "Purchased additional products"),
                                value = c(300, 123, 98, 72, 80, 15, 8),
                                stringsAsFactors = FALSE),
               neck_height = 30, neck_width = 40,
               main = "Reversed Funnel", mainSize = 15, inverse = TRUE,
               label_side = "left", margin_right = 15, margin_left = 160)

amFunnel(data = data.frame(description = c("Website visits", "Downloads",
                                                 "Requested price list",
                                                 "Contaced for more info",
                                                 "Purchased", "Contacted for support",
                                                 "Purchased additional products"),
                                value = c(300, 123, 98, 72, 80, 15, 8),
                                stringsAsFactors = FALSE),
               third_dim = TRUE,
               main = "Pyramid", mainSize = 15, inverse = TRUE)
}

