% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmGraph.R
\docType{class}
\name{AmGraph-class}
\alias{AmGraph-class}
\title{AmGraph class}
\description{
This class represents data for a serial chart
}
\section{Slots}{

\describe{
\item{\code{\code{balloonText}:}}{{Object of class \code{character}.
Balloon text. You can use tags like [[value]], [[description]], [[percents]], [[open]], [[category]]
or any other field name from your data provider. HTML tags can also be used.}}

\item{\code{\code{title}:}}{{Object of class \code{character}. Graph title.}}

\item{\code{\code{type}:}}{{Object of class \code{character}.
Type of the graph. Possible values are: "line", "column", "step", "smoothedLine", "candlestick", "ohlc".
XY and Radar charts can only display "line" otherArguments graphs.}}

\item{\code{\code{valueField}:}}{{Object of class \code{character}.
Name of the value field in your dataProvider.}}
}}
\section{Slots}{

}
\examples{
\dontrun{
new("AmGraph")
unclass(new("AmGraph"))
}
}
\author{
DataKnowledge
}
\seealso{
Other rAmChart classes: \code{\link{AmBalloon-class}};
  \code{\link{AmChart-class}}; \code{\link{AmObject-class}}
}

