% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmLegend.R
\docType{class}
\name{AmLegend-class}
\alias{AmLegend-class}
\title{AmLegend class}
\description{
Creates the legend for the chart, automatically adapts the color settings of the graphs.
}
\details{
Run \code{api("AmLegend")} for more information and all avalaible properties.
}
\section{Slots}{

\describe{
\item{\code{useGraphSettings}}{\code{logical}.
If TRUE, border color will be changed when user rolls-over the slice, graph, 
etc, instead of background color.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must be a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list} 
containing other avalaible properties not yet implemented in the package.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

