% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amMekko.R
\name{amMekko}
\alias{amMekko}
\title{Plotting mekko chart (quali vs quali) using rAmCharts}
\usage{
amMekko(x, y, data, main = "", mainSize = 15, xlab = "", ylab = "",
  groups_color = NULL, horiz = FALSE, legend = FALSE,
  legend_side = "right", show_values = FALSE)
}
\arguments{
\item{x}{\code{character} column name for x-axis or \code{numeric} 
the number of the corresponding column.}

\item{y}{\code{character} column name for y-axis or \code{numeric} 
the number of the corresponding column.}

\item{data}{\code{data.frame} dataframe with values to display.}

\item{main}{\code{character} title of the graph.}

\item{mainSize}{\code{numeric} size of the title of the graph.}

\item{xlab}{\code{character} label for x-axis.}

\item{ylab}{\code{character} label for y-axis.}

\item{groups_color}{\code{character} vector of colors in hexadecimal, 
same length as the number of y modalities.}

\item{horiz}{\code{boolean} TRUE for an horizontal chart, FALSE for a vertical one}

\item{legend}{\code{boolean} TRUE to display legend.}

\item{legend_side}{\code{character} either "bottom", "top", "left" or "right" if 
legend is set to TRUE.}

\item{show_values}{\code{boolean} TRUE to display values.}
}
\description{
amMekko computes a mekko chart of the given values.
}
\examples{


#Basic Example : 
data_mekko <- data.frame(var1 = c(rep("A1", 150), rep("A2", 350), rep("A3", 500)),
                         var2 = sample(c("B1", "B2", "B3", "B4", "B5", "B6"), 
                                       1000, replace = TRUE))
                                       
amMekko(x = "var1", y = "var2", data = data_mekko)

#Horizontal :
amMekko(x = "var1", y = "var2", data = data_mekko, horiz = TRUE)

#display values :
amMekko(x = "var1", y = "var2", data = data_mekko, show_values = TRUE)

#add legend :
amMekko(x = "var1", y = "var2", data = data_mekko, legend = TRUE)

}

