% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmObject.R
\docType{class}
\name{AmObject-class}
\alias{AmObject-class}
\title{AmObject class}
\description{
This is a virtual class for representing any Am** class
}
\section{Slots}{

\describe{
\item{\code{\code{otherProperties}:}}{Object of class \code{"list"},
containing other avalaible properties non coded in the package yet}

\item{\code{\code{otherProperties}:}}{Object of class \code{"numeric"}}
}}
\section{Slots}{

}
\author{
DataKnowledge
}
\seealso{
Other rAmChart classes: \code{\link{AmBalloon-class}};
  \code{\link{AmChart-class}}; \code{\link{AmGraph-class}}
}

