% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amPie.R
\name{amPie}
\alias{amPie}
\title{Plotting pie chart using rAmCharts}
\usage{
amPie(data, main = "", mainSize = 15, legend = FALSE,
  legend_side = NULL, show_values = TRUE)
}
\arguments{
\item{data}{\code{data.frame} dataframe with at least 2 columns :
label (character), value (numeric).
Your can add a third column "color" (character, colors in hexadecimal).}

\item{main}{\code{character}, title of the graph.}

\item{mainSize}{\code{numeric}, size of the title of the graph.}

\item{legend}{\code{boolean} TRUE to display legend.}

\item{legend_side}{\code{character} either "left" or "right" if legend is set to TRUE}

\item{show_values}{\code{boolean} TRUE to display values.}
}
\description{
amPie computes a pie chart of the given value.
}
\examples{
amPie(data = data.frame(label = c("Facebook", "Twitter", "LinkedIn", "Google+",
                                       "Pinterest"),
                              value = c(38, 25, 15, 14, 8),
                              stringsAsFactors = FALSE),
            main = "Favourite social media", show_values = TRUE)
}

