% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_amStockAggreg.R
\name{amStockAggreg}
\alias{amStockAggreg}
\title{Plotting times series which aggregation}
\usage{
amStockAggreg(data, col_date, col_series, main = "", ylab = "",
  color = c("#2E2EFE", "#31B404", "#FF4000", "#AEB404"), bullet = NULL,
  aggregation = c("Average", "Low", "High", "Sum"), maxSeries = 300,
  groupToPeriods = c("ss", "mm", "hh", "DD", "MM", "YYYY"),
  ZoomButton = NULL, precision = 1, ...)
}
\arguments{
\item{data}{\code{data.frame}, data of graph.}

\item{col_date}{\code{character} name of date column}

\item{col_series}{\code{character} names of series columns}

\item{main}{\code{character}, title.}

\item{ylab}{\code{character}, value axis label.}

\item{color}{\code{character}, color of series (in hexadecimal).}

\item{bullet}{\code{character}, point shape. Possible values are : "diamond", "square", 
"bubble",  "yError", "xError", "round", "triangleLeft", "triangleRight", "triangleUp",}

\item{aggregation}{\code{character}, aggregation type. Possible values are : 
"Low", "High", "Average" and "Sum"}

\item{maxSeries}{\code{numeric} Maximum series shown at a time.
In case there are more data points in the selection than maxSeries,
the chart will group data to longer periods,
for example - you have 250 days in the selection,
and maxSeries is 150 - the chart will group data to weeks.}

\item{groupToPeriods}{\code{character}, Periods to which data will
be gruoped in case there are more data items in the selected
period than specified in maxSeries property. Possible value are :
'ss', 'mm', 'hh', 'DD', 'MM', 'YYYY'. It's also possible to add multiple like "30mm".}

\item{ZoomButton}{\code{data.frame}, 3 columns : 
Unit, times unit
multiple : multiple*unit 
label : button's label}

\item{precision}{\code{numeric}, digits precision}

\item{...}{other first level attributes}
}
\description{
amStockAggreg computes a stock chart.
}
\examples{
data("data_stock_2")
amStockAggreg(data_stock_2, "date", c("ts1", "ts2"))

\donttest{
amStockAggreg(data_stock_2, "date", c("ts1", "ts2"), bullet = "round")
amStockAggreg(data_stock_2, "date", c("ts1", "ts2"), bullet = "round",
              groupToPeriods = c('hh', 'DD', '10DD'),)

amStockAggreg(data_stock_2, "date", c("ts1", "ts2"), bullet = "round",
              groupToPeriods = c('12hh', 'DD', '10DD'),
              maxSeries = 50)



amStockAggreg(data_stock_2, "date", c("ts1", "ts2"), bullet = "round",
              groupToPeriods = c('12hh', 'DD', '10DD'),
              maxSeries = 50, precision = 5)

amStockAggreg(data_stock_2, "date", c("ts1", "ts2"), bullet =  c("diamond", "square"),
             groupToPeriods = c('12hh', 'DD', '10DD'),
             maxSeries = 50, aggregation = "Sum")


ZoomButton <- data.frame(Unit = c("DD", "DD", "MAX"), multiple = c(1, 2 ,1),
                        label = c("Day","2 days", "MAX"))
amStockAggreg(data_stock_2, "date", c("ts1", "ts2"), bullet = "round",
             ZoomButton = ZoomButton, main = "My title", ylab = "Interest")
             
}
}
\seealso{
\itemize{
\item{\url{https://datastorm-open.github.io/introduction_ramcharts/}}
}
}

