% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\name{amStockChart}
\alias{amStockChart}
\title{amStockChart is a shortcut constructor
for instantiating AmChart of type \code{stock}}
\usage{
amStockChart(categoryAxesSettings, chartCursorSettings, chartScrollbarSettings,
  comparedDataSets, dataSets, dataSetSelector, legendSettings, mainDataSet,
  panels, panelsSettings, periodSelector, stockEventsSettings, theme,
  valueAxesSettings, ...)
}
\description{
amStockChart is a shortcut constructor
for instantiating AmChart of type \code{stock}
}
\examples{
amStockChart()
}

