% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amWaterfall.R
\name{amWaterfall}
\alias{amWaterfall}
\title{Plotting waterfall chart using rAmCharts}
\usage{
amWaterfall(data, start = 0, horiz = FALSE, show_values = FALSE, ...)
}
\arguments{
\item{data}{\code{data.frame} dataframe with at least 3 columns : 
label (character), value (numeric), operation (character : "plus", "minus", "total").
You can add a third column "color" (character, colors in hexadecimal). You can
also add a column "description" (character) containing the text you want to
display when mouse is on the graphic ('<br>' for a new line).}

\item{start}{\code{numeric} value from which to start}

\item{horiz}{\code{boolean} TRUE for an horizontal chart, FALSE for a vertical one}

\item{show_values}{\code{boolean} TRUE to display values on the chart.}

\item{...}{see \code{\link{amOptions}} for more options}
}
\description{
amWaterfall computes a waterfall chart of the given value.
}
\examples{
# Basic example
data_waterfall <- data.frame(label = c("Income 1", "Income 2", "Income 3", "Total 1", 
                                       "Expenses 1", "Expenses 2", "Total 2", "Income 4", 
                                       "Income 5", "Income 6", "Expenses 3","Total 3", 
                                       "Expenses 4", "Expenses 5", "Total 4"),
                             value = c(5, 10, 15, 30, 10, 5, 15, 50, 10, 35, 10, 100, 
                                        15, 60, 25),
                             operation = c(rep("plus", 3), "total", rep("minus", 2),
                                            "total", "plus", "minus", rep("plus", 2), 
                                            "total", rep("minus", 2), "total"), 
                                            stringsAsFactors = FALSE)

#simple exemple
amWaterfall(data = data_waterfall, show_values = TRUE)



#change the orientation :
amWaterfall(data = data_waterfall, horiz = TRUE)                                  
}

