% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amWind.R
\name{amWind}
\alias{amWind}
\title{Plotting wind using rAmCharts}
\usage{
amWind(data, col = NULL, backTransparency = 0.5, main = "",
  legend = TRUE, export = FALSE, fontSize = 15, pch = "round")
}
\arguments{
\item{data}{: columns are series of values, from week wind (first column) to strong wind (last column)}

\item{col}{: color of series}

\item{backTransparency}{: background transparency}

\item{main}{: title of graph}

\item{legend}{: add legend, TRUE or FALSE}

\item{export}{: add export, TRUE or FALSE}

\item{fontSize}{: font size}

\item{pch}{: symbols}
}
\description{
wind computes a windplot of the given data values.
}
\examples{

data <- data.frame(Week = c(1, 2, 3, 4,1, 2,1, 2),
                   Middle = c(2, 8, 1, 1,2, 8,1, 2),
                   Strong = c(1, 1, 2, 2,1, 1 ,1, 2))
amWind(data, main = "", export = TRUE,
       col = c("#0404B4","#01DF01","#FFBF00"),
       backTransparency = 1, pch="round")
       
}
\seealso{
\code{\link{amRadar}}
}

