% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api}
\alias{api}
\title{See AmCharts API}
\usage{
api(class = NULL)
}
\arguments{
\item{class}{Object of class \code{character}.
Name of the class to see documentation.
Please respect lower and upper case.}
}
\description{
Open a window in your browser at the referenced documentation
under \url{http://docs.amcharts.com/3/javascriptstockchart/}.
}
\examples{
api()
api("AmChart")
}
