% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AxisBase.R
\docType{class}
\name{AxisBase-class}
\alias{AxisBase-class}
\title{AxisBase class}
\description{
AxisBase class
}
\section{Slots}{

\describe{
\item{\code{guides}}{Object of class \code{list}.}

\item{\code{listeners}}{Object of class \code{"list"} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string. See examples for details.}

\item{\code{otherProperties}}{Object of class \code{"list"},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{Object of class \code{numeric}.
Guides belonging to this axis. Use addGuide method}
}}
\author{
DataKnowledge
}

