% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amBulletChart.R
\name{bulletChart}
\alias{bulletChart}
\title{Plotting bullet chart using rAmCharts}
\usage{
bulletChart(value, min = 0, max = 100, val_color = "#000000",
  limit = 85, limit_color = "#000000", steps = TRUE,
  main = "Bullet Chart", mainSize = 15, legend = "", horiz = TRUE,
  rates)
}
\arguments{
\item{value}{\code{numeric}}

\item{min}{\code{numeric} minimum value allowed}

\item{max}{\code{numeric} maximum value allowed}

\item{val_color}{\code{character} color of the bar value, in hexadecimal}

\item{limit}{\code{numeric} target value}

\item{limit_color}{\code{character} color of the line limit}

\item{steps}{\code{boolean} default set to TRUE}

\item{main}{\code{character}, title of the graph}

\item{mainSize}{\code{numeric}, size of the title of the graph.}

\item{legend}{\code{character} text lengend}

\item{horiz}{\code{boolean} TRUE for an horizontal bullet chart, FALSE for a vertical one}

\item{rates}{a data frame with 4 columns : name (character), min (numeric), max (numeric),
and color (character, color in hexadecimal)}
}
\description{
bulletChart computes a bullet chart of the given value.
}
\examples{
### basic example
# keep the default parameter values:
bulletChart(value = 65)

# equivalent to:
bulletChart(value = 65, min = 0, max = 100, val_color = "#000000",
             limit = 85, limit_color = "#000000",
             steps = TRUE, main = "Bullet chart", mainSize = 15,
             legend = "", horiz = TRUE,
             rates = data.frame(name = c("excelent", "good", "average", "poor", "bad"),
                                min = c(0, 20, 40, 60, 80),
                                max = c(20, 40, 60, 80, 100),
                                color = c("#fb7116", "#b4dd1e", "#f4fb16",
                                          "#19d228", "#f6d32b")))

### Remove steps for background:
bulletChart(value = 65, steps = FALSE)

### Tune the colors with name or HTML code:
bulletChart(value = 65, val_color = "purple", limit_color = "#3c8dbc")

### Change the orientation:
bulletChart(value = 65, steps = FALSE, horiz = FALSE)

### Add title and legend:
bulletChart(value = 65, legend = "Evaluation",
            main = "Bullet chart 1", mainSize = 15)

### Change min and max values:
bulletChart(value = 65, min = 20, max = 90)
}

