% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CategoryAxis.R
\docType{class}
\name{CategoryAxis-class}
\alias{CategoryAxis-class}
\title{CategoryAxis class}
\description{
Children class of AxisBase.
Gets automatically populated if none has been specified.
}
\details{
Run \code{api("CategoryAxis")} for more information and all avalaible properties.
}
\section{Slots}{

\describe{
\item{\code{gridPosition}}{\code{character}.
Specifies if a grid line is placed on the center of a cell or on the beginning of a cell.
Possible values are: "start" and "middle"
This setting doesn't work if parseDates is set to true.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

