% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CategoryAxis.R
\name{categoryAxis}
\alias{categoryAxis}
\title{Constructor for an AmGraph}
\usage{
categoryAxis(gridPosition, ...)
}
\arguments{
\item{...}{Properties of CategoryAxis.
See \url{http://docs.amcharts.com/3/javascriptcharts/CategoryAxis}}
}
\value{
An \code{\linkS4class{CategoryAxis}} object
}
\description{
Constructor for an AmGraph
}
\examples{
new("CategoryAxis", gridPosition = "start")
categoryAxis(gridPosition = "start", adjustBorderColor = TRUE)
}

