% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartCursor.R
\docType{class}
\name{ChartCursor-class}
\alias{ChartCursor-class}
\title{ChartCursor class}
\description{
ChartCursor class
}
\section{Slots}{

\describe{
\item{\code{oneBalloonOnly}}{Object of class \code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}

\item{\code{valueLineAxis}}{Object of class \code{list}.
Properties of Axis of value line. If you set valueLineBalloonEnabled to true,
but you have more than one axis, you can use this property
to indicate which axis should display balloon.}

\item{\code{listeners}}{Object of class \code{"list"} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{Object of class \code{"list"},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{Object of class \code{numeric}.}
}}
\author{
DataKnowledge
}

