% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartScrollbar.R
\docType{class}
\name{ChartScrollbar-class}
\alias{ChartScrollbar-class}
\title{ChartScrollbar class}
\description{
ChartScrollbar class
}
\section{Slots}{

\describe{
\item{\code{updateOnReleaseOnly}}{Object of class \code{logical}.
Specifies if the chart should be updated while dragging/resizing the scrollbar
or only at the moment when user releases mouse button.}

\item{\code{graph}}{Object of class \code{list}.
Specifies which graph properties will be displayed in the scrollbar.
Only Serial chart's scrollbar can display a graph.}

\item{\code{listeners}}{Object of class \code{"list"} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{Object of class \code{"list"},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{Object of class \code{numeric}.}
}}
\author{
DataKnowledge
}

