% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartScrollbar.R
\name{chartScrollbar}
\alias{chartScrollbar}
\title{#’ Constructor.}
\usage{
chartScrollbar(graph, updateOnReleaseOnly,
  pathToImages = "http://www.amcharts.com/lib/3/images/", ...)
}
\arguments{
\item{\code{...}:}{{Properties of ChartScrollbar.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/ChartScrollbar}}}}
}
\value{
An \code{\linkS4class{ChartScrollbar}} object
}
\description{
#’ Constructor.
}
\examples{
chartScrollbar()
chartScrollbar(updateOnReleaseOnly = TRUE)
}

