% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_AirPassengers}
\alias{data_AirPassengers}
\title{Air passengers for example}
\format{2 column, 144 rows :
\describe{
  \item{AirPassengers}{\code{numeric}}
  \item{Period}{\code{character}, MM/YYYY}
}}
\usage{
data_AirPassengers
}
\description{
Based on the dataset 'AirPassengers' from the package 'datasets'.
}
\keyword{datasets}
