% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_bar}
\alias{data_bar}
\title{Random data for plotting bar chart examples}
\format{Dataset of 3 columns and 12 rows
\describe{
  \item{country}{\code{character}}
  \item{visits}{\code{numeric}}
  \item{color}{\code{character}}
}}
\usage{
data_bar
}
\description{
This dataset is used in the examples,
moreover it can be used as a reference if you notice a bug
}
\keyword{datasets}

