% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_fbar}
\alias{data_fbar}
\title{Random data for plotting floating bar chart examples}
\format{Dataset of 2 columns and 1000 rows
\describe{
  \item{country}{\code{character}}
  \item{visits_inf}{\code{numeric}}
  \item{visits_sup}{\code{numeric}}
  \item{color}{\code{character}}
}}
\usage{
data_fbar
}
\description{
This dataset is used in the examples,
moreover it can be used as a reference if you notice a bug
}
\keyword{datasets}

