% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_funnel}
\alias{data_funnel}
\title{Random data for plotting funnel chart examples}
\format{
Dataset of 2 columns and 7 rows
\describe{
  \item{description}{\code{character}}
  \item{value}{\code{numeric}}
}
}
\usage{
data_funnel
}
\description{
This dataset is used in the examples,
moreover it can be used as a reference if you notice a bug.
}
\keyword{datasets}
