% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_gantt}
\alias{data_gantt}
\title{Random data for plotting gantt chart examples}
\format{
Dataset of 5 columns and 4 rows
\describe{
  \item{category}{\code{character}}
  \item{begin}{\code{date}}
  \item{end}{\code{date}}
  \item{color}{\code{character}}
}
}
\usage{
data_gantt
}
\description{
This dataset is used in the examples,
moreover it can be used as a reference if you notice a bug.
}
\keyword{datasets}
