% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_gbar}
\alias{data_gbar}
\title{Random data for plotting bar chart examples}
\format{Dataset of 5 columns and 12 rows
\describe{
  \item{year}{\code{character}}
  \item{day}{\code{character}}
  \item{month}{\code{character}}
  \item{income}{\code{numeric}}
  \item{expenses}{\code{numeric}}
}}
\usage{
data_gbar
}
\description{
This dataset is used in the examples,
moreover it can be used as a reference if you notice a bug.
}
\keyword{datasets}
