% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_gdp}
\alias{data_gdp}
\title{10 Richest Countries in the World by 2015 GDP}
\format{Dataset of 2 columns and 10 rows
\describe{
  \item{country}{\code{chracter}}
  \item{gdp}{\code{numeric}}
}}
\source{
\url{http://www.insidermonkey.com/blog/10-richest-countries-in-the-world-by-2015-gdp-344692/}
}
\usage{
data_gdp
}
\description{
Value in $ trillion
}
\keyword{datasets}

