% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_mekko}
\alias{data_mekko}
\title{Random data for plotting mekko chart examples}
\format{Dataset of 2 columns and 1000 rows
\describe{
  \item{var1}{\code{character}}
  \item{var2}{\code{numeric}}
}}
\usage{
data_mekko
}
\description{
This dataset is used in the examples,
moreover it can be used as a reference if you notice a bug.
}
\keyword{datasets}

