% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_pie}
\alias{data_pie}
\title{Random data for plotting pie chart examples}
\format{Dataset of 2 columns and 5 rows
\describe{
  \item{label}{\code{character}}
  \item{value}{\code{numeric}}
}}
\usage{
data_pie
}
\description{
This dataset is used in the examples,
moreover it can be used as a reference if you notice a bug.
}
\keyword{datasets}

