% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_radar}
\alias{data_radar}
\title{Random data for plotting radar chart examples}
\format{Dataset of 4 columns and 5 rows
\describe{
  \item{label}{\code{character}}
  \item{Product1}{\code{numeric}}
  \item{Product2}{\code{numeric}}
  \item{Product3}{\code{numeric}}
}}
\usage{
data_radar
}
\description{
This dataset is used in the examples,
moreover it can be used as a reference if you notice a bug
}
\keyword{datasets}

