% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_stock1}
\alias{data_stock1}
\title{Random data for example}
\format{Each datasetis a data.table with 21 rows and 4 variables:
\describe{
  \item{date}{vector of dates}
  \item{a}{random vector of data}
  \item{b}{random vector of data}
}}
\usage{
data_stock1
}
\description{
A list containing 4 datasets
}
\keyword{datasets}

