% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_stock_3}
\alias{data_stock_3}
\title{Random data for example}
\format{List of 4 datasets, 4 variables in each
\describe{
  \item{date}{vector of datesTimes}
  \item{value}{random vector of data}
  \item{volume}{random vector of data}
  \item{value2}{random vector of data}
  \item{value3}{random vector of data}
}}
\usage{
data_stock_3
}
\description{
Times Series on 2017, by months
}
\keyword{datasets}

