% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_waterfall}
\alias{data_waterfall}
\title{Random data for plotting candlestick chart examples}
\format{Dataset of 3 columns and 15 rows
\describe{
  \item{label}{\code{character}, can be parsed as a date}
  \item{value}{\code{numeric}}
  \item{operation}{\code{numeric}}
}}
\usage{
data_waterfall
}
\description{
This dataset is used in the examples,
moreover it can be used as a reference if you notice a bug
}
\keyword{datasets}

