% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_data.R
\docType{data}
\name{data_wind}
\alias{data_wind}
\title{Random data for plotting wind chart examples}
\format{Dataset of 3 columns and 8 rows
\describe{
  \item{weak}{\code{numeric}}
  \item{middle}{\code{numeric}}
  \item{strong}{\code{numeric}}
}}
\usage{
data_wind
}
\description{
This dataset is used in the examples,
moreover it can be used as a reference if you notice a bug.
}
\keyword{datasets}

