% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataSet.R
\name{dataSet}
\alias{dataSet}
\title{#’ Constructor.}
\usage{
dataSet(compared = FALSE, dataProvider, fieldMappings, stockEvents, ...)
}
\arguments{
\item{\code{...}:}{Properties of DataSet.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/DataSet}}}
}
\value{
An \code{\linkS4class{DataSet}} object
}
\description{
#’ Constructor.
}
\examples{
dataSet( categoryField = "categoryField" )
}

