% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DataSetSelector.R
\docType{class}
\name{DataSetSelector-class}
\alias{DataSetSelector-class}
\title{DataSetSelector class}
\description{
DataSetSelector is a tool for selecting data set's
as main and for comparing with main data set.
}
\details{
Run \code{api("DataSetSelector")} for more information and all avalaible properties.
}
\section{Slots}{

\describe{
\item{\code{position}}{\code{character}.
Possible values: "right", "left", "top", "bottom".
"top" and "bottom" positions has a limitation -
only one data set can be selected for comparing.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must be a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list}
containing other avalaible properties not yet coded in the package.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

