% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DataSetSelector.R
\docType{methods}
\name{initialize,DataSetSelector-method}
\alias{initialize,DataSetSelector-method}
\alias{dataSetSelector}
\alias{setPosition}
\alias{setPosition,DataSetSelector,character-method}
\title{Creates or updates a DataSetSelector}
\usage{
\S4method{initialize}{DataSetSelector}(.Object, position, ...)

dataSetSelector(position, ...)

setPosition(.Object, position)

\S4method{setPosition}{DataSetSelector,character}(.Object, position)
}
\arguments{
\item{.Object}{\linkS4class{DataSetSelector}.}

\item{position}{\code{character}.
Possible values: "right", "left", "top", "bottom".
"top" and "bottom" positions has a limitation -
only one data set can be selected for comparison.}

\item{...}{other properties of DataSetSelector.}
}
\value{
(updated) \linkS4class{DataSetSelector}.
}
\description{
Use the constructors to create the object with its properties
or update an existing one with the setters.
}
\examples{
new("DataSetSelector", size = 10)

dataSetSelector(position = "left")
setPosition(.Object = dataSetSelector(), position = "left")

}
