% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GaugeArrow.R
\name{gaugeArrow}
\alias{gaugeArrow}
\title{Constructor.}
\usage{
gaugeArrow(fillAlpha, alpha = 1, axis, ...)
}
\arguments{
\item{\code{...}:}{{Properties of GaugeArrow.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/GaugeArrow}}}}
}
\value{
An \code{\linkS4class{GaugeArrow}} object
}
\description{
Constructor.
}
\examples{
gaugeArrow()
}

