% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeBand.R
\docType{class}
\name{GaugeBand-class}
\alias{GaugeBand-class}
\title{GaugeBand class}
\description{
GaugeBand class
}
\section{Slots}{

\describe{
\item{\code{id}}{Object of class \code{character}}

\item{\code{listeners}}{Object of class \code{"list"} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{Object of class \code{"list"},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{Object of class \code{numeric}.

Unique id of a band..}
}}
\author{
DataKnowledge
}

