% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmObject.R
\docType{methods}
\name{getOtherProperties,ANY-method}
\alias{getOtherProperties,ANY-method}
\title{Getter}
\usage{
\S4method{getOtherProperties}{ANY}(.Object)
}
\description{
Getter
}
\examples{
# Getter for otherProperties
ls <- list(categoryAxis = list(gridPosition = "start"), test = 1)
amPieChart( otherProperties = ls) \%>>\% getOtherProperties
}
\seealso{
\code{\linkS4class{AmObject}}

Other AmObjects methods: \code{\link{setProperties,AmObject-method}}
}

