% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Guide.R
\docType{class}
\name{Guide-class}
\alias{Guide-class}
\title{Guide class}
\description{
Guide class
}
\section{Slots}{

\describe{
\item{\code{\code{fillAlpha}:}}{{ Object of clas \code{numeric}.
Specifies if a grid line is placed on the center of a cell or on the beginning of a cell.
Possible values are: "start" and "middle" This setting doesn't work if parseDates is set to true.}}

\item{\code{\code{valueAxis}:}}{{ Object of class \code{list},
containing properties of a \linkS4class{ValueAxis} class.
As you can add guides directly to the chart, you might need to specify which which value axis should be used.
}}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

