% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Guide.R
\docType{methods}
\name{guide}
\alias{guide}
\alias{setFillAlpha,Guide,numeric-method}
\title{Constructor.}
\usage{
guide(fillAlpha, valueAxis, value, ...)

\S4method{setFillAlpha}{Guide,numeric}(.Object, fillAlpha)
}
\arguments{
\item{\code{...}:}{{Properties of Guide.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/Guide}}}}
}
\value{
An \code{\linkS4class{Guide}} object
}
\description{
Constructor.

SETTER
}
\examples{
guide(fillAlpha = .4, value = 1)
guide(fillAlpha = .4, adjustBorderColor = TRUE, gridThickness = 1)
guide() \%>>\% setFillAlpha(1)
}

