% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmBalloon.R
\docType{methods}
\name{initialize,AmBalloon-method}
\alias{initialize,AmBalloon-method}
\title{Initialize an AmBalloon}
\usage{
\S4method{initialize}{AmBalloon}(.Object, adjustBorderColor, color,
  cornerRadius, fillColor, ...)
}
\arguments{
\item{\code{adjustBorderColor}:}{Object of class \code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}

\item{\code{color}:}{Object of class \code{character}.
Color of text in the balloon.}

\item{\code{cornerRadius}:}{Object of class \code{numeric}.
Balloon corner radius.}

\item{\code{cornerRadius}:}{Object of class \code{character}.
Balloon background color. Usually balloon background color is set by the chart.
Only if "adjustBorderColor" is "true" this color will be used.}

\item{\code{...}:}{Properties of AmBalloon.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/AmBalloon}}.}
}
\value{
An object of class \code{\linkS4class{AmBalloon}}.
}
\description{
Initialize an AmBalloon
}
\examples{
new("AmBalloon", cornerRadius = 10)
}
\seealso{
\code{\linkS4class{AmBalloon}} S4 class

Other AmBalloon methods: \code{\link{amBalloon}};
  \code{\link{setAdjustBorderColor}};
  \code{\link{setColor}}; \code{\link{setFillColor}}
}

