% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmGraph.R
\docType{methods}
\name{initialize,AmGraph-method}
\alias{initialize,AmGraph-method}
\title{Initialize an AmGraph}
\usage{
\S4method{initialize}{AmGraph}(.Object, animationPlayed = FALSE, balloonText,
  title, type, valueField, ...)
}
\description{
Initialize an AmGraph
}
\examples{
new("AmGraph", valueField = "value")
}

