% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmLegend.R
\docType{methods}
\name{initialize,AmLegend-method}
\alias{initialize,AmLegend-method}
\title{Initialize}
\usage{
\S4method{initialize}{AmLegend}(.Object, useGraphSettings, ...)
}
\arguments{
\item{.Object}{}

\item{useGraphSettings}{Object of class \code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}
}
\description{
Initialize
}
\examples{
new("AmLegend", useGraphSettings = TRUE)
}

