% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmLegend.R
\name{initialize,AmLegend-method}
\alias{initialize,AmLegend-method}
\alias{amLegend}
\alias{legend}
\alias{setUseGraphSettings}
\alias{setUseGraphSettings,AmLegend,logical-method}
\title{Initializes legend of the chart}
\usage{
\S4method{initialize}{AmLegend}(.Object, useGraphSettings, ...)

amLegend(useGraphSettings, ...)

legend(useGraphSettings, ...)

setUseGraphSettings(.Object, useGraphSettings)

\S4method{setUseGraphSettings}{AmLegend,logical}(.Object, useGraphSettings)
}
\arguments{
\item{.Object}{\linkS4class{AmLegend}.}

\item{useGraphSettings}{\code{logical}, 
if TRUE, border color will be changed when user rolls-over the slice, graph, 
etc, instead of background color.}

\item{...}{Other properties of \linkS4class{AmLegend}.
See \url{https://docs.amcharts.com/3/javascriptstockchart/AmLegend}.}
}
\description{
Constructor for an AmLegend.
}
\examples{
new("AmLegend", useGraphSettings = TRUE)
amLegend(useGraphSettings = FALSE)
rAmCharts:::legend(useGraphSettings = FALSE)
setUseGraphSettings(.Object = amLegend(), useGraphSettings = TRUE)
}
