% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{initialize,AmStockChart-method}
\alias{initialize,AmStockChart-method}
\title{Initialize an AmStockChart}
\usage{
\S4method{initialize}{AmStockChart}(.Object, categoryAxesSettings,
  chartCursorSettings, chartScrollbarSettings, comparedDataSets, dataSets,
  dataSetSelector, legendSettings, mainDataSet, panels, panelsSettings,
  periodSelector, stockEventsSettings, theme, valueAxesSettings,
  pathToImages = "http://www.amcharts.com/lib/3/images/", ...)
}
\description{
Method for initializing any S4 class provided by the package.
}
\examples{
new("AmStockChart", theme = "dark")
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

Other AmChart methods: \code{\link{addArrow}};
  \code{\link{addAxe}}; \code{\link{addGraph}};
  \code{\link{addGuide}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}}; \code{\link{addListener}};
  \code{\link{addSegment}}; \code{\link{addSubData}};
  \code{\link{addTitle}}; \code{\link{addTrendLine}};
  \code{\link{addValueAxes}};
  \code{\link{initialize,AmChart-method}};
  \code{\link{setAllLabels}}; \code{\link{setArrows}};
  \code{\link{setAxes}}; \code{\link{setBalloon}};
  \code{\link{setCategoryAxis}};
  \code{\link{setCategoryField}};
  \code{\link{setChartCursor}};
  \code{\link{setChartScrollbar}};
  \code{\link{setCreditsPosition}};
  \code{\link{setDataProvider}},
  \code{\link{setDataProvider}}; \code{\link{setExport}};
  \code{\link{setGraphs}}; \code{\link{setGraph}},
  \code{\link{setGraph}}; \code{\link{setGuides}};
  \code{\link{setLegend}};
  \code{\link{setSubChartProperties}};
  \code{\link{setTitles}}; \code{\link{setTrendLines}};
  \code{\link{setValueAxes}}; \code{\link{setValueAxis}}
}

