% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AxisBase.R
\docType{methods}
\name{addGuide,AxisBase-method}
\alias{addGuide,AxisBase-method}
\title{Add a Guide}
\usage{
\S4method{addGuide}{AxisBase}(.Object, guide = NULL, ...)
}
\arguments{
\item{.Object}{\linkS4class{AxisBase}}

\item{guide}{\linkS4class{Guide}}

\item{...}{Properties of \code{\linkS4class{Guide}}.}
}
\description{
Add a Guide
}
\examples{
library(pipeR)
valueAxis(axisTitleOffset = 12, tickLength = 10) \%>>\%
addGuide(fillAlpha = .4, adjustBorderColor = TRUE, gridThickness = 1)
}

