% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CategoryAxis.R
\docType{methods}
\name{initialize,CategoryAxis-method}
\alias{initialize,CategoryAxis-method}
\title{Initialize a CategoryAxis}
\usage{
\S4method{initialize}{CategoryAxis}(.Object, gridPosition, guides, ...)
}
\arguments{
\item{gridPosition}{Object of class \code{character}.}

\item{guides}{Object of class \code{list}. List of guides.}
}
\description{
Initialize a CategoryAxis
}
\examples{
\donttest{
new("CategoryAxis")
new("CategoryAxis", gridPosition = "start", 1) # 1 is not take into account
}
guides <- list(guide(fillAlpha = .4, adjustBorderColor = TRUE), guide(fillAlpha = .4, adjustBorderColor = TRUE))
new("CategoryAxis", gridPosition = "start",  gridThickness = 1, guides = guides)
}

