% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartScrollbar.R
\docType{methods}
\name{initialize,ChartScrollbar-method}
\alias{initialize,ChartScrollbar-method}
\title{Initialize a ChartScrollbar}
\usage{
\S4method{initialize}{ChartScrollbar}(.Object, graph,
  pathToImages = "http://www.amcharts.com/lib/3/images/", updateOnReleaseOnly,
  ...)
}
\description{
Initialize a ChartScrollbar
}
\examples{
new("ChartScrollbar", graph = "g1")
new("ChartScrollbar", graph = amGraph(test = 1))
}

