% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{initialize,DataSet-method}
\alias{initialize,DataSet-method}
\title{Initialize a DataSet}
\usage{
\S4method{initialize}{DataSet}(.Object, compared = FALSE, dataProvider,
  fieldMappings, stockEvents, ...)
}
\description{
Initialize a DataSet
}
\examples{
new("DataSet")
}

