% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetSelector.R
\docType{methods}
\name{initialize,DataSetSelector-method}
\alias{dataSetSelector}
\alias{initialize,DataSetSelector-method}
\alias{setPosition}
\alias{setPosition,DataSetSelector,character-method}
\title{Initialize  DataSetSelector}
\usage{
\S4method{initialize}{DataSetSelector}(.Object, position, ...)

dataSetSelector(position, ...)

setPosition(.Object, position)

\S4method{setPosition}{DataSetSelector,character}(.Object, position)
}
\arguments{
\item{.Object}{\linkS4class{DataSetSelector}}

\item{position}{\code{character}.
Possible values: "right", "left", "top", "bottom".
"top" and "bottom" positions has a limitation -
only one data set can be selected for comparing.}

\item{...}{Other properties.}
}
\value{
(updated) \linkS4class{DataSetSelector}
}
\examples{
new("DataSetSelector", size = 10)
dataSetSelector(position = "left")
setPosition(.Object = dataSetSelector(), position = "left")
setPosition(.Object = dataSetSelector(), position = "left")
}

