% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeArrow.R
\docType{methods}
\name{initialize,GaugeArrow-method}
\alias{initialize,GaugeArrow-method}
\title{Initialize a GaugeArrow}
\usage{
\S4method{initialize}{GaugeArrow}(.Object, alpha = 1, axis, ...)
}
\description{
Initialize a GaugeArrow
}
\examples{
new("GaugeArrow")
}

