% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeAxis.R
\docType{methods}
\name{initialize,GaugeAxis-method}
\alias{initialize,GaugeAxis-method}
\title{Initialize a GaugeAxis}
\usage{
\S4method{initialize}{GaugeAxis}(.Object, axisAlpha = 1, bands, ...)
}
\description{
Initialize a GaugeAxis
}
\examples{
new("GaugeAxis", alpha = 1)
}

