% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeAxis.R
\docType{methods}
\name{initialize,GaugeAxis-method}
\alias{addBand}
\alias{addBand,GaugeAxis-method}
\alias{gaugeAxis}
\alias{initialize,GaugeAxis-method}
\alias{setBands}
\alias{setBands,GaugeAxis,list-method}
\title{Initialize a GaugeAxis}
\usage{
\S4method{initialize}{GaugeAxis}(.Object, axisAlpha = 1, bands, ...)

gaugeAxis(axisAlpha = 1, bands, ...)

setBands(.Object, bands)

\S4method{setBands}{GaugeAxis,list}(.Object, bands)

addBand(.Object, band = NULL, ...)

\S4method{addBand}{GaugeAxis}(.Object, band = NULL, ...)
}
\arguments{
\item{.Object}{\linkS4class{GaugeAxis}.}

\item{axisAlpha}{\code{numeric}.}

\item{bands}{\code{list} of \linkS4class{GaugeBand}.
Bands are used to draw color fills between specified values.}

\item{...}{other properties.}

\item{band}{\linkS4class{GaugeBand}.}
}
\description{
Initialize a GaugeAxis
}
\section{Functions}{
\itemize{
\item \code{gaugeAxis}: 
}}
\examples{
new("GaugeAxis", alpha = 1)
gaugeAxis()
bands <- list(gaugeBand(), gaugeBand())
gaugeAxis(bands = bands)
\dontrun{
# error
bands <- list(gaugeBand(), test = 1)
gaugeAxis(bands = bands)
}
addBand(.Object = gaugeAxis(), band = gaugeBand(test = "foo"))
}

