% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GaugeBand.R
\docType{methods}
\name{initialize,GaugeBand-method}
\alias{initialize,GaugeBand-method}
\title{Initialize a GaugeBand}
\usage{
\S4method{initialize}{GaugeBand}(.Object, alpha = 1, id, ...)
}
\description{
Initialize a GaugeBand
}
\examples{
new("GaugeBand")
}

