% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaugeBand.R
\docType{methods}
\name{initialize,GaugeBand-method}
\alias{gaugeBand}
\alias{initialize,GaugeBand-method}
\alias{setID}
\alias{setID,GaugeBand-method}
\title{Initialize a GaugeBand}
\usage{
\S4method{initialize}{GaugeBand}(.Object, alpha = 1, id, ...)

gaugeBand(alpha = 1, id, ...)

setID(.Object, id)

\S4method{setID}{GaugeBand}(.Object, id)
}
\arguments{
\item{.Object}{\linkS4class{GaugeBand} (or "GaugeBand" for initialize).}

\item{alpha}{\code{numeric}.}

\item{id}{\code{character}.}

\item{...}{other properties.}
}
\value{
(updated) .Object of class \linkS4class{GaugeBand}.
}
\examples{
new("GaugeBand")
gaugeBand()
setID(.Object = gaugeBand(), id = "1")
}

