% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Guide.R
\docType{methods}
\name{initialize,Guide-method}
\alias{initialize,Guide-method}
\title{Initialize a Guide}
\usage{
\S4method{initialize}{Guide}(.Object, fillAlpha, valueAxis, value, ...)
}
\description{
Initialize a Guide
}
\examples{
new("Guide", fillAlpha = 0.1,  gridThickness = 1, value = 1)
}

