% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{initialize,Label-method}
\alias{initialize,Label-method}
\title{Initialize}
\usage{
\S4method{initialize}{Label}(.Object, text, bold, x, y, ...)
}
\description{
Initialize
}
\examples{
new("Label", x = 10)
}

