% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{initialize,Label-method}
\alias{initialize,Label-method}
\alias{label}
\alias{setBold}
\alias{setBold,Label,logical-method}
\alias{setText,Label,character-method}
\alias{setX}
\alias{setX,Label,numeric-method}
\alias{setY}
\alias{setY,Label,numeric-method}
\title{Initialize}
\usage{
\S4method{initialize}{Label}(.Object, text, bold, x, y, ...)

label(text, bold, x, y, ...)

setBold(.Object, bold)

\S4method{setBold}{Label,logical}(.Object, bold)

\S4method{setText}{Label,character}(.Object, text)

setX(.Object, x)

\S4method{setX}{Label,numeric}(.Object, x)

setY(.Object, y)

\S4method{setY}{Label,numeric}(.Object, y)
}
\arguments{
\item{.Object}{\linkS4class{Label}.}

\item{text}{\code{character}.
Text of a title.}

\item{bold}{\code{character}.
Specifies if label is bold or not.}

\item{x}{\code{numeric}.
X position of a label.}

\item{y}{\code{numeric}.
Y position of a label.}

\item{...}{Other properties.}
}
\value{
(updated) .Object of class \linkS4class{Label}.
}
\description{
Initialize
}
\section{Functions}{
\itemize{
\item \code{label}: 
}}
\examples{
new("Label", x = 10)
label(text = "bonjour")
setBold(.Object = label(), bold = TRUE)
setText(.Object = label(), text = "Bonjour")
setX(.Object = label(), x = 16)
setY(.Object = label(), y = 16)
}

