% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PeriodSelector.R
\docType{methods}
\name{initialize,PeriodSelector-method}
\alias{initialize,PeriodSelector-method}
\title{Initialize a PeriodSelector}
\usage{
\S4method{initialize}{PeriodSelector}(.Object, periods, ...)
}
\description{
Initialize a PeriodSelector
}
\examples{
new( "PeriodSelector" )
}

