% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockEvent.R
\docType{methods}
\name{initialize,StockEvent-method}
\alias{initialize,StockEvent-method}
\title{Initialize a StockEvent}
\usage{
\S4method{initialize}{StockEvent}(.Object, backgroundAlpha = 1, stockGraph,
  ...)
}
\description{
Initialize a StockEvent
}
\examples{
new("StockEvent")
}

