% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockPanel.R
\docType{methods}
\name{initialize,StockPanel-method}
\alias{initialize,StockPanel-method}
\title{Initialize a StockPanel}
\usage{
\S4method{initialize}{StockPanel}(.Object, drawOnAxis, stockGraphs, stockLegend,
  ...)
}
\description{
Initialize a StockPanel
}
\examples{
new("StockPanel")
}

