% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Title.R
\docType{methods}
\name{initialize,Title-method}
\alias{initialize,Title-method}
\title{Initialize A Title}
\usage{
\S4method{initialize}{Title}(.Object, text, size, ...)
}
\description{
Initialize A Title
}
\examples{
new("Title", size = 10)
}

