% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Title.R
\docType{methods}
\name{initialize,Title-method}
\alias{initialize,Title-method}
\alias{setSize}
\alias{setSize,Title,numeric-method}
\alias{setText,Title,character-method}
\alias{title}
\title{Initialize A Title}
\usage{
\S4method{initialize}{Title}(.Object, text, size, ...)

title(text, size, ...)

\S4method{setText}{Title,character}(.Object, text)

setSize(.Object, size)

\S4method{setSize}{Title,numeric}(.Object, size)
}
\arguments{
\item{.Object}{\code{\linkS4class{Title}}}

\item{text}{\code{character}.
Text of a title.}

\item{size}{\code{numeric}.
Text size of a title.}

\item{...}{Other properties.}
}
\value{
(updated) \code{\linkS4class{Title}}
}
\examples{
new("Title", size = 10)
title(text = "bonjour")
setText(.Object = title(), text = "Bonjour")
library(pipeR)
title() \%>>\% setSize(16)
}

