% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ValueAxis.R
\docType{methods}
\name{initialize,ValueAxis-method}
\alias{initialize,ValueAxis-method}
\title{Initialize}
\usage{
\S4method{initialize}{ValueAxis}(.Object, title, guides, ...)
}
\arguments{
\item{\code{title}:}{{Object of class \code{character}}}

\item{\code{guides}:}{{Object of class \code{list}. List of guides.}}
}
\description{
Initialize
}
\examples{
\dontrun{
new("ValueAxis", title = "Hello !", 1) # 1 is not take into account

# If one element of guides is not a Guide object, it shows an error
guides <- list(guide(fillAlpha = .4), b = 1)
new("ValueAxis", title = "Hello !",  gridThickness = 1, guides = guides)
}

guides <- list(guide(fillAlpha = .4), guide(fillAlpha = .5))
new("ValueAxis", title = "Hello !",  gridThickness = 1, guides = guides)
}

