% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Label.R
\docType{class}
\name{Label-class}
\alias{Label-class}
\title{Label class}
\description{
Label class
}
\section{Slots}{

\describe{
\item{\code{\code{text}:}}{{Object of class \code{character}.
Text of a title.}}

\item{\code{\code{x}:}}{{Object of class \code{numeric}.
Text x of a title.}}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

