% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmBalloon.R
\docType{methods}
\name{listProperties,AmBalloon-method}
\alias{listProperties,AmBalloon-method}
\title{List attributes of an AmGraph object}
\usage{
\S4method{listProperties}{AmBalloon}(.Object)
}
\arguments{
\item{.Object:}{\code{\linkS4class{AmBalloon}}.}
}
\description{
List attributes of an AmGraph object
}
\examples{
amBalloon(adjustBorderColor = TRUE)
}

