% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmGraph.R
\docType{methods}
\name{listProperties,AmGraph-method}
\alias{listProperties,AmGraph-method}
\title{List attributes of an AmGraph object}
\usage{
\S4method{listProperties}{AmGraph}(.Object)
}
\description{
This function is used to list attributes before addind to graphs (attribute of AmChart)
}
\examples{
library(pipeR)
amGraph(balloonText = "toto", type = "type", valueField = "valueField") \%>>\%
listProperties

amGraph(balloonText = "toto", type = "type") \%>>\%
listProperties
}

