% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmLegend.R
\docType{methods}
\name{listProperties,AmLegend-method}
\alias{listProperties,AmLegend-method}
\title{List properties}
\usage{
\S4method{listProperties}{AmLegend}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
is(listProperties(amLegend(useGraphSettings = TRUE)), "list")
}

