% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmObject.R
\docType{methods}
\name{listProperties,AmObject-method}
\alias{listProperties,AmObject-method}
\title{List attributes of an AmObject}
\usage{
\S4method{listProperties}{AmObject}(.Object)
}
\description{
List attributes of an AmObject
}
\details{
Factor the code for the commone attributes
}

