% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmObject.R
\name{listProperties}
\alias{listProperties}
\alias{listProperties,AmObject-method}
\title{List properties of an S4 object}
\usage{
listProperties(.Object)

\S4method{listProperties}{AmObject}(.Object)
}
\arguments{
\item{.Object}{any class object of the package}
}
\value{
A list containing all the chart's properties.
}
\description{
Each S4 class implements the method to list its properties
(usefull to update complex properties).
}
\examples{
amChart(type = "serial")
}
