% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{listProperties,AmStockChart-method}
\alias{listProperties,AmStockChart-method}
\title{List attributes of an AmStockChart}
\usage{
\S4method{listProperties}{AmStockChart}(.Object)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}
}
\description{
This method lists attributes of an AmChart to fit the API
}
\details{
For certain attributes we do not verify if they are NULL, see constructor.
}
\examples{
library(pipeR)
amChart() \%>>\% setProperties(test = 1) \%>>\% listProperties
}

