% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AxisBase.R
\docType{methods}
\name{listProperties,AxisBase-method}
\alias{listProperties,AxisBase-method}
\title{List properties}
\usage{
\S4method{listProperties}{AxisBase}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}

