% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CategoryAxis.R
\docType{methods}
\name{listProperties,CategoryAxis-method}
\alias{listProperties,CategoryAxis-method}
\title{List properties}
\usage{
\S4method{listProperties}{CategoryAxis}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
\dontshow{
library(pipeR)
categoryAxis(ignoreAxisWidth = TRUE) \%>>\% listProperties() \%>>\% class
categoryAxis(ignoreAxisWidth = TRUE) \%>>\% listProperties() \%>>\% length
categoryAxis(ignoreAxisWidth = TRUE) \%>>\% setGridPosition("start") \%>>\% listProperties()
}
}

