% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartCursor.R
\docType{methods}
\name{listProperties,ChartCursor-method}
\alias{listProperties,ChartCursor-method}
\title{List properties}
\usage{
\S4method{listProperties}{ChartCursor}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
new("ChartCursor", oneBalloonOnly = TRUE)
}

