% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartScrollbar.R
\docType{methods}
\name{listProperties,ChartScrollbar-method}
\alias{listProperties,ChartScrollbar-method}
\title{List properties}
\usage{
\S4method{listProperties}{ChartScrollbar}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
new("ChartScrollbar", updateOnReleaseOnly = TRUE)
}

