% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeArrow.R
\docType{methods}
\name{listProperties,GaugeArrow-method}
\alias{listProperties,GaugeArrow-method}
\title{List properties}
\usage{
\S4method{listProperties}{GaugeArrow}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
lapply(list(gaugeArrow(alpha = .4, value = 1), gaugeArrow(alpha = .5)), listProperties)
}

