% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeAxis.R
\docType{methods}
\name{listProperties,GaugeAxis-method}
\alias{listProperties,GaugeAxis-method}
\title{List properties}
\usage{
\S4method{listProperties}{GaugeAxis}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}

