% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeBand.R
\docType{methods}
\name{listProperties,GaugeBand-method}
\alias{listProperties,GaugeBand-method}
\title{List properties}
\usage{
\S4method{listProperties}{GaugeBand}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
lapply(list(gaugeBand(fillAlpha = .4, value = 1), gaugeBand(fillAlpha = .5)), listProperties)
}

