% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Guide.R
\docType{methods}
\name{listProperties,Guide-method}
\alias{listProperties,Guide-method}
\title{List properties}
\usage{
\S4method{listProperties}{Guide}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
lapply(list(guide(fillAlpha = .4, value = 1), guide(fillAlpha = .5)), listProperties)
}

