% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{listProperties,Label-method}
\alias{listProperties,Label-method}
\title{List properties}
\usage{
\S4method{listProperties}{Label}(.Object)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{Label}}.}

\item{y}{Object of class \code{numeric}.
Y position of a label.}
}
\value{
The updated .Object.

Properties of the object in a list
}
\description{
List properties
}
\examples{
label(text = "balloonText")
}

