% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockEvent.R
\docType{methods}
\name{listProperties,StockEvent-method}
\alias{listProperties,StockEvent-method}
\title{List properties}
\usage{
\S4method{listProperties}{StockEvent}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}

