% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StockPanel.R
\docType{methods}
\name{listProperties,StockPanel-method}
\alias{listProperties,StockPanel-method}
\title{List properties}
\usage{
\S4method{listProperties}{StockPanel}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}

