% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Title.R
\docType{methods}
\name{listProperties,Title-method}
\alias{listProperties,Title-method}
\title{List properties}
\usage{
\S4method{listProperties}{Title}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
new("Title", text = "foo")
}

