% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrendLine.R
\docType{methods}
\name{listProperties,TrendLine-method}
\alias{listProperties,TrendLine-method}
\title{List properties}
\usage{
\S4method{listProperties}{TrendLine}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
trendLine( initialValue = 1, valueAxis = valueAxis(axisTitleOffset = 12, tickLength = 10) )
}

