% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ValueAxis.R
\docType{methods}
\name{listProperties,ValueAxis-method}
\alias{listProperties,ValueAxis-method}
\title{List properties}
\usage{
\S4method{listProperties}{ValueAxis}(.Object)
}
\arguments{
\item{.Object}{}
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
library(pipeR)
\dontshow{
valueAxis( axisTitleOffset = 12, tickLength = 10 ) \%>>\% listProperties \%>>\% class
}
valueAxis( axisTitleOffset = 12, tickLength = 10 ) \%>>\%
addGuide( fillAlpha = .4, adjustBorderColor = TRUE, gridThickness = 1 ) \%>>\%
setProperties( axisTitleOffset = 12 ) \%>>\% listProperties
}

