% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PeriodSelector.R
\docType{class}
\name{PeriodSelector-class}
\alias{PeriodSelector-class}
\title{PeriodSelector class}
\description{
PeriodSelector class
}
\section{Slots}{

\describe{
\item{\code{periods}}{Object of clas \code{numeric}.
Array of predefined period objects.
Period object has 4 properties - period, count, label and selected.
Possible period values are:
"ss" - seconds, "mm" - minutes, "hh" - hours, "DD" - days, "MM" - months and "YYYY" - years.
property "count" specifies how many periods this button will select.
"label" will be displayed on a button and "selected" is a boolean
which specifies if this button is selected when chart is initialized or not.
Example: {period:"DD", count:10, label:"10 days", selected:false}.}

\item{\code{listeners}}{Object of class \code{"list"} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{Object of class \code{"list"},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{Object of class \code{numeric}.}
}}
\author{
DataKnowledge
}

