% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PeriodSelector.R
\name{periodSelector}
\alias{periodSelector}
\title{Constructor.}
\usage{
periodSelector(periods, ...)
}
\arguments{
\item{\code{...}:}{{Properties of PeriodSelector.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/PeriodSelector}}}}
}
\value{
An \code{\linkS4class{PeriodSelector}} object
}
\description{
Constructor.
}
\examples{
periodSelector(fillAlpha = .4, value = 1)
periodSelector(fillAlpha = .4, adjustBorderColor = TRUE, gridThickness = 1)
}

