% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_plot.R
\docType{methods}
\name{plot,AmChart-method}
\alias{plot,AmChart-method}
\title{PLOTTING METHOD}
\usage{
\S4method{plot}{AmChart}(x, y, width = "100\%", height = "500px",
  background = "#ffffff", ...)
}
\description{
Basic method to plot an AmChart
}
\details{
Plots an object of class \code{\linkS4class{AmChart}}
}

