% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmObject.R
\docType{methods}
\name{print,AmObject-method}
\alias{print,AmObject-method}
\title{Visualize with print}
\usage{
\S4method{print}{AmObject}(x, ...)
}
\arguments{
\item{x}{\linkS4class{AmChart}}

\item{...}{Other properties.}
}
\description{
Visualize with print
}
\examples{
print(new("AmChart", categoryField = "variables"))
}
\seealso{
Other Visualizations: \code{\link{show,AmObject-method}}
}

