% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_shinyUtils.R
\docType{methods}
\name{renderAmChart}
\alias{renderAmChart}
\title{SHINY}
\usage{
\S4method{renderAmChart}{ANY}(expr, env, quoted)
}
\arguments{
\item{amChart:}{an \code{\linkS4class{AmChart}} object}
}
\description{
Widget output function for use in Shiny
}

