% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_shinyUtils.R
\docType{methods}
\name{renderAmCharts}
\alias{renderAmCharts}
\alias{renderAmCharts,ANY-method}
\alias{renderAmCharts,AmChart-method}
\title{SHINY}
\usage{
renderAmCharts(expr, env, ...)

renderAmCharts(expr, env, ...)

\S4method{renderAmCharts}{ANY}(expr, env, quoted)

\S4method{renderAmCharts}{AmChart}(expr, env)
}
\arguments{
\item{expr}{\code{expression}.}

\item{env}{see htmlwidgets doc}

\item{...}{other arguments.}

\item{quoted}{\code{logical}.}

\item{expr}{\code{expression}.}

\item{env}{see htmlwidgets doc.}

\item{quoted}{\code{logical}.}
}
\description{
Widget output function for use in Shiny

Widget output function for use in Shiny
}

