% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyExamples.R
\name{runShinyExamples}
\alias{runShinyExamples}
\title{Few shiny examples}
\usage{
runShinyExamples()
}
\description{
Launch a shiny app with amChart examples
}
\examples{
\dontrun{
runShinyExamples()
}
}

