% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{seeOfficialAPI}
\alias{seeOfficialAPI}
\title{See AmCharts API}
\usage{
seeOfficialAPI(class = NULL)
}
\arguments{
\item{class}{Object of class \code{character}.
Name of the class to see documentation.
Please respect lower and upper case.}
}
\description{
Open a window in your browser at the referenced documentation
under \url{http://docs.amcharts.com/3/javascriptstockchart/}.
}
\examples{
seeOfficialAPI()
seeOfficialAPI("AmChart")
}

