% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmBalloon.R
\docType{methods}
\name{setAdjustBorderColor}
\alias{setAdjustBorderColor}
\title{Setter for adjustBorderColor}
\usage{
\S4method{setAdjustBorderColor}{AmBalloon,logical}(.Object, adjustBorderColor)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{AmBalloon}}.}

\item{\code{adjustBorderColor}:}{Object of class \code{logical}.}
}
\value{
The updated object of class \code{\linkS4class{AmBalloon}}.
}
\description{
Setter for adjustBorderColor
}
\examples{
amBalloon() \%>>\% setAdjustBorderColor(TRUE)
}
\seealso{
\code{\linkS4class{AmBalloon}} S4 class

Other AmBalloon methods: \code{\link{amBalloon}};
  \code{\link{initialize,AmBalloon-method}};
  \code{\link{setColor}}; \code{\link{setFillColor}}

Other AmBalloon setters: \code{\link{setCornerRadius}};
  \code{\link{setFillColor}}
}

