% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setAllLabels}
\alias{setAllLabels}
\alias{setAllLabels,AmChart,list-method}
\title{Setter for allLabels}
\usage{
setAllLabels(.Object, allLabels, ...)

\S4method{setAllLabels}{AmChart,list}(.Object, allLabels)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{allLabels}{object of class \code{list}.}

\item{...}{Other properties.}
}
\value{
The updater object of class \code{\linkS4class{AmChart}}
}
\description{
Setter for allLabels
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,allLabels = list}: 
}}
\examples{
allLabels <- list(label(text = "balloonText"), label(text = "column"))
setAllLabels(.Object = amChart(), allLabels = allLabels)
amChart(allLabels = allLabels)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class
}

