% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setArrows}
\alias{setArrows}
\alias{setArrows,AmChart-method}
\title{Setter for arrows}
\usage{
setArrows(.Object, arrows = NULL)

\S4method{setArrows}{AmChart}(.Object, arrows = NULL)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{arrows}{(optionnal) list of \code{\linkS4class{GaugeArrow}}.}
}
\description{
Setter for arrows
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
library(pipeR)
# Setter for arrows
amChart() \%>>\% setArrows()
arrows <- list( gaugeArrow(value = 130), gaugeArrow(value = 150) )
amChart() \%>>\% setArrows(arrows)
amChart(arrows = arrows)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class
}

