% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeArrow.R
\docType{methods}
\name{setAxis,GaugeArrow-method}
\alias{setAxis,GaugeArrow-method}
\title{SETTER}
\usage{
\S4method{setAxis}{GaugeArrow}(.Object, axis = NULL, ...)
}
\description{
SETTER
}
\examples{
library(pipeR)
gaugeArrow() \%>>\% setAxis()
}

