% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setBalloon}
\alias{setBalloon}
\alias{setBalloon,AmChart-method}
\title{Setter}
\usage{
setBalloon(.Object, amBalloon = NULL, ...)

\S4method{setBalloon}{AmChart}(.Object, amBalloon = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{amBalloon}{object of class \code{\linkS4class{AmBalloon}}.}

\item{...}{properties of \code{\linkS4class{AmBalloon}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.

The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
setBalloon(.Object = amChart(), adjustBorderColor = TRUE)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{AmBalloon}} S4 class
}

