% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmGraph.R
\docType{methods}
\name{setBalloonText}
\alias{setBalloonText}
\title{Setter}
\usage{
\S4method{setBalloonText}{AmGraph,character}(.Object, balloonText)
}
\description{
Setter
}
\examples{
library(pipeR)
amGraph() \%>>\% setBalloonText("balloonText")
}

