% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeAxis.R
\docType{methods}
\name{setBands,GaugeAxis,list-method}
\alias{setBands,GaugeAxis,list-method}
\title{SETTER}
\usage{
\S4method{setBands}{GaugeAxis,list}(.Object, bands)
}
\description{
SETTER
}
\examples{
bands <- list(gaugeBand(), gaugeBand())
gaugeAxis(bands = bands)
\dontrun{
# error
bands <- list(gaugeBand(), test = 1)
gaugeAxis(bands = bands)
}
}

