% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setCategoryAxis}
\alias{setCategoryAxis}
\alias{setCategoryAxis,AmChart-method}
\title{Setter}
\usage{
setCategoryAxis(.Object, categoryAxis = NULL, ...)

\S4method{setCategoryAxis}{AmChart}(.Object, categoryAxis = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{categoryAxis}{object of class \code{\linkS4class{CategoryAxis}}.}

\item{...}{properties of \code{\linkS4class{CategoryAxis}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
library(pipeR)
# Setter for categoryAxis
amChart() \%>>\% setCategoryAxis(categoryAxis(gridPosition = "start"))
amChart() \%>>\% setCategoryAxis(gridPosition = "start")
\dontrun{
# The argument categoryAxis must be an object of class CategoryAxis
amChart() \%>>\% setCategoryAxis(categoryAxis = "error")
}
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{CategoryAxis}} S4 class
}

