% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setCategoryField}
\alias{setCategoryField}
\alias{setCategoryField,AmChart,character-method}
\title{Setter}
\usage{
setCategoryField(.Object, categoryField)

\S4method{setCategoryField}{AmChart,character}(.Object, categoryField)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}}

\item{categoryField}{object of class \code{category}}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,categoryField = character}: 
}}
\examples{
setCategoryField(.Object = amChart(), categoryField = "category")
}
\seealso{
\code{\linkS4class{AmChart}} S4 class
}

