% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setChartCursor}
\alias{setChartCursor}
\alias{setChartCursor,AmChart-method}
\title{Setter}
\usage{
setChartCursor(.Object, chartCursor = NULL, ...)

\S4method{setChartCursor}{AmChart}(.Object, chartCursor = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{chartCursor}{Object of class \code{\linkS4class{ChartCursor}}.}

\item{...}{properties of \code{\linkS4class{ChartCursor}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
library(pipeR)
amChart() \%>>\% setChartCursor(chartCursor(oneBallOnly = TRUE))

# same result
amChart() \%>>\% setChartCursor(oneBallOnly = TRUE)
object <- amChart() \%>>\% setChartCursor()
chartCursor() \%>>\% class
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{ChartCursor}} S4 class
}

