% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setChartScrollbar}
\alias{setChartScrollbar}
\alias{setChartScrollbar,AmChart-method}
\title{Setter}
\usage{
setChartScrollbar(.Object, chartScrollbar = NULL, ...)

\S4method{setChartScrollbar}{AmChart}(.Object, chartScrollbar = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}}

\item{chartScrollbar}{object of class \code{\linkS4class{ChartScrollbar}}.}

\item{...}{properties of \code{\linkS4class{ChartScrollbar}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
library(pipeR)
amChart() \%>>\% setChartScrollbar(chartScrollbar(oneBallOnly = TRUE))
amChart() \%>>\% setChartScrollbar()
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{ChartScrollbar}} S4 class
}

