% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setDataProvider,AmChart,data.frame-method}
\alias{setDataProvider,AmChart,data.frame-method}
\title{Setter for dataProvider}
\usage{
\S4method{setDataProvider}{AmChart,data.frame}(.Object, dataProvider,
  keepNA = TRUE)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmChart}}.}

\item{dataProvider}{object of class \code{data.frame}.}

\item{keepNA}{object of class \code{logical}, default \code{TRUE}.
Indicates if \code{NULL} values have to be kept or ignored.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter for dataProvider
}
\examples{
library(pipeR)
amChart() \%>>\% setDataProvider(data.frame(key = c("FR", "US"), value = c(20,10)))
}
\seealso{
\code{\linkS4class{AmChart}} S4 class
}

