% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{setDataSets}
\alias{setDataSets}
\title{Setter}
\usage{
\S4method{setDataSets}{AmStockChart}(.Object, dataSets)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{\code{dataSets}:}{Object of class \code{\linkS4class{list}}.
Each element of this list must be a dataSet}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter
}
\examples{
library(pipeR)
amStockChart() \%>>\% setDataSets(list(dataSet(compared = FALSE), dataSet(compared = FALSE)))
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

\code{\linkS4class{DataSet}} S4 class

Other AmStockChart methods: \code{\link{addComparedDataSet}};
  \code{\link{addDataSet}}; \code{\link{addPanel}};
  \code{\link{setCategoryAxesSettings}};
  \code{\link{setChartCursorSettings}};
  \code{\link{setChartScrollbarSettings}};
  \code{\link{setComparedDataSets}};
  \code{\link{setDataSetSelector}};
  \code{\link{setLegendSettings}};
  \code{\link{setMainDataSet}};
  \code{\link{setPanelsSettings}}; \code{\link{setPanels}};
  \code{\link{setPeriodSelector}};
  \code{\link{setStockEventsSettings}};
  \code{\link{setValueAxesSettings}}

Other AmStockChart setters: \code{\link{addComparedDataSet}};
  \code{\link{addDataSet}}; \code{\link{addPanel}};
  \code{\link{setCategoryAxesSettings}};
  \code{\link{setChartCursorSettings}};
  \code{\link{setChartScrollbarSettings}};
  \code{\link{setComparedDataSets}};
  \code{\link{setDataSetSelector}};
  \code{\link{setLegendSettings}};
  \code{\link{setMainDataSet}};
  \code{\link{setPanelsSettings}}; \code{\link{setPanels}};
  \code{\link{setPeriodSelector}};
  \code{\link{setStockEventsSettings}};
  \code{\link{setValueAxesSettings}}
}

