% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{setDataSetSelector,AmStockChart-method}
\alias{setDataSetSelector,AmStockChart-method}
\title{Setter}
\usage{
\S4method{setDataSetSelector}{AmStockChart}(.Object, ...)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{...}{Property of datSetSelector}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter
}
\examples{
library(pipeR)
amStockChart() \%>>\% setDataSetSelector(width = 180)
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

Other AmStockChart methods: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setMainDataSet,AmStockChart-method}};
  \code{\link{setPanels,AmStockChart,list-method}};
  \code{\link{setPanelsSettings,AmStockChart-method}};
  \code{\link{setPeriodSelector,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}};
  \code{\link{setValueAxesSettings,AmStockChart-method}}

Other AmStockChart setters: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setMainDataSet,AmStockChart-method}};
  \code{\link{setPanels,AmStockChart,list-method}};
  \code{\link{setPanelsSettings,AmStockChart-method}};
  \code{\link{setPeriodSelector,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}};
  \code{\link{setValueAxesSettings,AmStockChart-method}}
}

