% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setExport}
\alias{setExport}
\alias{setExport,AmChart-method}
\title{Setter for export}
\usage{
setExport(.Object, enabled = TRUE, ...)

\S4method{setExport}{AmChart}(.Object, enabled = TRUE, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{enabled}{object of class \code{logical}, default \code{TRUE}.
Should the export button be shown ?}

\item{...}{properties for export}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter for export
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
library(pipeR)
amChart() \%>>\% setExport()
}
\seealso{
\code{\linkS4class{AmChart}} S4 class
}

