% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{setFieldMappings,DataSet,list-method}
\alias{setFieldMappings,DataSet,list-method}
\title{Add fieldMapping}
\usage{
\S4method{setFieldMappings}{DataSet,list}(.Object, fieldMappings)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{DataSet}}.}

\item{fieldMappings}{Object of class \code{list}}
}
\value{
The updated \code{.Object}
}
\description{
Add fieldMapping
}
\examples{
library(pipeR)
dataSet() \%>>\% addFieldMapping(fromField ="val1", toField ="value")
}

