% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmBalloon.R
\docType{methods}
\name{setFillColor}
\alias{setFillColor}
\title{Setter for fillColor}
\usage{
\S4method{setFillColor}{AmBalloon,character}(.Object, fillColor)
}
\arguments{
\item{.Object:}{Object of class \code{\linkS4class{AmBalloon}}.}
}
\value{
The updated object of class \code{\linkS4class{AmBalloon}}.
}
\description{
Setter for fillColor
}
\examples{
library(pipeR)
amBalloon() \%>>\% setFillColor("#FFFFFF")
}
\seealso{
\code{\linkS4class{AmBalloon}} S4 class

Other AmBalloon methods: \code{\link{amBalloon}};
  \code{\link{initialize,AmBalloon-method}};
  \code{\link{setAdjustBorderColor}};
  \code{\link{setColor}}

Other AmBalloon setters: \code{\link{setAdjustBorderColor}};
  \code{\link{setCornerRadius}}
}

