% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrendLine.R
\docType{methods}
\name{setFinalValue,TrendLine,numeric-method}
\alias{setFinalValue,TrendLine,numeric-method}
\title{SETTER}
\usage{
\S4method{setFinalValue}{TrendLine,numeric}(.Object, finalValue)
}
\description{
SETTER
}
\examples{
library(pipeR)
trendLine() \%>>\% setFinalValue(16)
}

