% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrendLine.R
\docType{methods}
\name{setFinalXValue,TrendLine,numeric-method}
\alias{setFinalXValue,TrendLine,numeric-method}
\title{SETTER}
\usage{
\S4method{setFinalXValue}{TrendLine,numeric}(.Object, finalXValue)
}
\description{
SETTER
}
\examples{
trendLine() \%>>\% setFinalXValue(16)
}

