% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setGraph,AmChart-method}
\alias{setGraph,AmChart-method}
\title{Setter for graph}
\usage{
\S4method{setGraph}{AmChart}(.Object, graph = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}} and type \code{gantt}.}

\item{graph}{Object of class \code{\linkS4class{AmGraph}}.}

\item{...}{properties of \code{\linkS4class{AmGraph}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter for graph
}
\details{
Method to use in case of AmChart of type \code{gantt}.
For other type see methods \code{setGraphs} or \code{addGraph}.
}
\examples{
setGraph(.Object = amGanttChart())
}
\seealso{
\code{\linkS4class{AmChart}} S4 class
}

