% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setGraphs}
\alias{setGraphs}
\alias{setGraphs,AmChart,list-method}
\title{Setter}
\usage{
setGraphs(.Object, graphs)

\S4method{setGraphs}{AmChart,list}(.Object, graphs)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}}

\item{graphs}{Object of class \code{list}.
Each element must be an object of class \code{\linkS4class{AmGraph}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,graphs = list}: 
}}
\examples{
graphs <- list(amGraph(balloonText = "balloonText"), amGraph(type = "column"))
setGraphs(.Object = amChart(), graphs = graphs)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{AmGraph}} S4 class
}

