% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CategoryAxis.R
\docType{methods}
\name{setGridPosition,CategoryAxis,character-method}
\alias{setGridPosition,CategoryAxis,character-method}
\title{SETTER}
\usage{
\S4method{setGridPosition}{CategoryAxis,character}(.Object, gridPosition)
}
\description{
SETTER
}
\examples{
library(pipeR)
categoryAxis() \%>>\% setGridPosition("start")
}

