% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeBand.R
\docType{methods}
\name{setID,GaugeBand-method}
\alias{setID,GaugeBand-method}
\title{SETTER}
\usage{
\S4method{setID}{GaugeBand}(.Object, id)
}
\description{
SETTER
}
\examples{
library(pipeR)
gaugeBand() \%>>\% setID("1")
}

