% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrendLine.R
\docType{methods}
\name{setInitialValue,TrendLine,numeric-method}
\alias{setInitialValue,TrendLine,numeric-method}
\title{SETTER}
\usage{
\S4method{setInitialValue}{TrendLine,numeric}(.Object, initialValue)
}
\description{
SETTER
}
\examples{
trendLine() \%>>\% setInitialValue(16)
}

