% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrendLine.R
\docType{methods}
\name{setInitialXValue,TrendLine,numeric-method}
\alias{setInitialXValue,TrendLine,numeric-method}
\title{SETTER}
\usage{
\S4method{setInitialXValue}{TrendLine,numeric}(.Object, initialXValue)
}
\description{
SETTER
}
\examples{
library(pipeR)
trendLine() \%>>\% setInitialXValue(16)
}

