% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setLegend}
\alias{setLegend}
\alias{setLegend,AmChart-method}
\title{Setter for Legend}
\usage{
setLegend(.Object, amLegend = NULL, ...)

\S4method{setLegend}{AmChart}(.Object, amLegend = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{amLegend}{object of class \code{\linkS4class{AmLegend}}, default \code{NULL}.}

\item{...}{properties of \code{\linkS4class{AmLegend}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter for Legend
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
library(pipeR)
# Without chaining
setLegend(.Object = amChart(), amLegend = amLegend(useGraphSettings = TRUE))
setLegend(.Object = amChart(), useGraphSettings = TRUE)

# With chaining
amChart() \%>>\% setLegend(useGraphSettings = TRUE)
amChart() \%>>\% setLegend( amLegend(useGraphSettings = TRUE) )
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{AmLegend}} S4 class
}

