% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartCursor.R
\docType{methods}
\name{setOneBalloonOnly,ChartCursor,logical-method}
\alias{setOneBalloonOnly,ChartCursor,logical-method}
\title{SETTER}
\usage{
\S4method{setOneBalloonOnly}{ChartCursor,logical}(.Object, oneBalloonOnly)
}
\description{
SETTER
}
\examples{
chartCursor() \%>>\% setOneBalloonOnly(TRUE)
}

