% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{setPanels,AmStockChart,list-method}
\alias{setPanels,AmStockChart,list-method}
\title{Setter}
\usage{
\S4method{setPanels}{AmStockChart,list}(.Object, panels)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{panels}{Object of class \code{\linkS4class{list}}.
Each element of this list must be a stockPanel}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter
}
\examples{
library(pipeR)
amStockChart() \%>>\% setPanels(list(stockPanel(compared = TRUE), stockPanel(compared = TRUE)))
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

\code{\linkS4class{StockPanel}} S4 class

Other AmStockChart methods: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setDataSetSelector,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setMainDataSet,AmStockChart-method}};
  \code{\link{setPanelsSettings,AmStockChart-method}};
  \code{\link{setPeriodSelector,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}};
  \code{\link{setValueAxesSettings,AmStockChart-method}}

Other AmStockChart setters: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setDataSetSelector,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setMainDataSet,AmStockChart-method}};
  \code{\link{setPanelsSettings,AmStockChart-method}};
  \code{\link{setPeriodSelector,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}};
  \code{\link{setValueAxesSettings,AmStockChart-method}}
}

