% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{setPanelsSettings,AmStockChart-method}
\alias{setPanelsSettings,AmStockChart-method}
\title{Setter for PanelsSettings}
\usage{
\S4method{setPanelsSettings}{AmStockChart}(.Object, ...)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{...}{Properties of \code{PanelsSettings}.}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter for PanelsSettings
}
\details{
Use this methode in case of an AmStockChart.
}
\examples{
library(pipeR)
amStockChart() \%>>\% setPanelsSettings(backgroundAlpha = 0)
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

Other AmStockChart methods: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setDataSetSelector,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setMainDataSet,AmStockChart-method}};
  \code{\link{setPanels,AmStockChart,list-method}};
  \code{\link{setPeriodSelector,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}};
  \code{\link{setValueAxesSettings,AmStockChart-method}}

Other AmStockChart setters: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setDataSetSelector,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setMainDataSet,AmStockChart-method}};
  \code{\link{setPanels,AmStockChart,list-method}};
  \code{\link{setPeriodSelector,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}};
  \code{\link{setValueAxesSettings,AmStockChart-method}}
}

