% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{setPeriodSelector}
\alias{setPeriodSelector}
\title{Setter for PeriodSelector}
\usage{
\S4method{setPeriodSelector}{AmStockChart}(.Object, periodSelector = NULL,
  ...)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{\code{periodSelector}:}{Object of class \code{\linkS4class{PeriodSelector}}.}

\item{\code{...}:}{Properties of the class \code{\linkS4class{PeriodSelector}}.}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter for PeriodSelector
}
\details{
Use this methode in case of an AmStockChart.
}
\examples{
library(pipeR)
amStockChart() \%>>\% setPeriodSelector(dateFormat = "DD-MM-YYYY")
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

Other AmStockChart methods: \code{\link{addComparedDataSet}};
  \code{\link{addDataSet}}; \code{\link{addPanel}};
  \code{\link{setCategoryAxesSettings}};
  \code{\link{setChartCursorSettings}};
  \code{\link{setChartScrollbarSettings}};
  \code{\link{setComparedDataSets}};
  \code{\link{setDataSetSelector}};
  \code{\link{setDataSets}};
  \code{\link{setLegendSettings}};
  \code{\link{setMainDataSet}};
  \code{\link{setPanelsSettings}}; \code{\link{setPanels}};
  \code{\link{setStockEventsSettings}};
  \code{\link{setValueAxesSettings}}

Other AmStockChart setters: \code{\link{addComparedDataSet}};
  \code{\link{addDataSet}}; \code{\link{addPanel}};
  \code{\link{setCategoryAxesSettings}};
  \code{\link{setChartCursorSettings}};
  \code{\link{setChartScrollbarSettings}};
  \code{\link{setComparedDataSets}};
  \code{\link{setDataSetSelector}};
  \code{\link{setDataSets}};
  \code{\link{setLegendSettings}};
  \code{\link{setMainDataSet}};
  \code{\link{setPanelsSettings}}; \code{\link{setPanels}};
  \code{\link{setStockEventsSettings}};
  \code{\link{setValueAxesSettings}}
}

