% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Title.R
\docType{methods}
\name{setSize,Title,numeric-method}
\alias{setSize,Title,numeric-method}
\title{SETTER}
\usage{
\S4method{setSize}{Title,numeric}(.Object, size)
}
\description{
SETTER
}
\examples{
title() \%>>\% setSize(16)
}

