% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{setStockEvents,DataSet,list-method}
\alias{setStockEvents,DataSet,list-method}
\title{SETTER}
\usage{
\S4method{setStockEvents}{DataSet,list}(.Object, stockEvents)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{DataSet}}.}

\item{stockEvents}{Object of class \code{list} containing \code{\linkS4class{StockEvent}}}
}
\value{
The updated .Object
}
\description{
SETTER
}

