% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockEvent.R
\docType{methods}
\name{setStockGraph,StockEvent-method}
\alias{setStockGraph,StockEvent-method}
\title{SETTER}
\usage{
\S4method{setStockGraph}{StockEvent}(.Object, stockGraph = NULL, ...)
}
\description{
SETTER
}
\examples{
stockEvent() \%>>\% setStockGraph()
}

