% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockPanel.R
\docType{methods}
\name{setStockGraphs,StockPanel-method}
\alias{setStockGraphs,StockPanel-method}
\title{SETTER}
\usage{
\S4method{setStockGraphs}{StockPanel}(.Object, stockGraphs)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{stockGraphs}{}
}
\value{
The updated \code{.Object}
}
\description{
SETTER
}
\examples{
library(pipeR)
stockPanel() \%>>\% setStockGraphs( list( stockGraph(comparable = TRUE), stockGraph(comparable = FALSE) ) )
}

