% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockPanel.R
\docType{methods}
\name{setStockLegend,StockPanel-method}
\alias{setStockLegend,StockPanel-method}
\title{SETTER}
\usage{
\S4method{setStockLegend}{StockPanel}(.Object, stockLegend = NULL, ...)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{stockLegend}{}
}
\value{
The updated \code{.Object}
}
\description{
SETTER
}
\examples{
library(pipeR)
stockPanel() \%>>\% setStockLegend( stockLegend() )
}

