% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setSubChartProperties}
\alias{setSubChartProperties}
\alias{setSubChartProperties,AmChart-method}
\title{Add subData for drilldrown}
\usage{
setSubChartProperties(.Object, .subObject = NULL, ...)

\S4method{setSubChartProperties}{AmChart}(.Object, .subObject = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{.subObject}{object of class \code{\linkS4class{AmChart}}.}

\item{...}{Properties of \code{\linkS4class{AmChart}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Add subData for drilldrown
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
library(pipeR)
amSerialChart() \%>>\% setSubChartProperties(type = "pie")
}
\seealso{
\code{\linkS4class{AmChart}} S4 class
}

