% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{setText,Label,character-method}
\alias{setText,Label,character-method}
\title{SETTER}
\usage{
\S4method{setText}{Label,character}(.Object, text)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{Label}}}

\item{text}{Object of class \code{character}.
Text of a title.}
}
\value{
The updated .Object
}
\description{
SETTER
}
\examples{
library(pipeR)
label() \%>>\% setText("Bonjour")
}

