% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Title.R
\docType{methods}
\name{setText,Title,character-method}
\alias{setText,Title,character-method}
\title{SETTER}
\usage{
\S4method{setText}{Title,character}(.Object, text)
}
\description{
SETTER
}
\examples{
library(pipeR)
title() \%>>\% setText("Bonjour")
}

