% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ValueAxis.R
\docType{methods}
\name{setTitle,ValueAxis,character-method}
\alias{setTitle,ValueAxis,character-method}
\title{SETTER}
\usage{
\S4method{setTitle}{ValueAxis,character}(.Object, title)
}
\description{
SETTER
}
\examples{
library(pipeR)
valueAxis() \%>>\% setTitle("Hello !")
}

