% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setTitles}
\alias{setTitles}
\alias{setTitles,AmChart,list-method}
\title{Set a list of Title's}
\usage{
setTitles(.Object, titles)

\S4method{setTitles}{AmChart,list}(.Object, titles)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{titles}{object of class \code{list}.
Each element must be an object of class \code{\linkS4class{Title}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Set a list of Title's
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,titles = list}: 
}}
\examples{
library(pipeR)
# Setter for titles
titles <- list(title(text = "balloonText"), title(text = "column"))
amChart() \%>>\% setTitles(titles)
amChart(titles = titles)
\dontrun{
titles <- list(title(text = "balloonText"), text = "column")
amChart(titles = titles)
}
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{Title}} S4 class
}

