% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setTrendLines}
\alias{setTrendLines}
\alias{setTrendLines,AmChart,list-method}
\title{Setter for trendLines}
\usage{
setTrendLines(.Object, trendLines)

\S4method{setTrendLines}{AmChart,list}(.Object, trendLines)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{trendLines}{object of class \code{list}.
Each element must be of class \code{\linkS4class{TrendLine}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter for trendLines
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,trendLines = list}: 
}}
\examples{
trendLines <- list(trendLine(initialValue = 1, finalValue = 5),
                   trendLine(initialValue = 7, finalValue = 19))
setTrendLines(.Object = amChart(), trendLines = trendLines)
amChart(trendLines = trendLines) # Equivalent
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{TrendLine}} S4 class
}

