% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R, R/AmGraph.R
\docType{methods}
\name{setType}
\alias{setType}
\title{Setter for Type}
\usage{
\S4method{setType}{AmChart,character}(.Object, type)

\S4method{setType}{AmGraph,character}(.Object, type)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{AmChart}}.}

\item{\code{type}:}{Object of class \code{\linkS4class{character}}.}
}
\description{
Setter for Type

Setter
}
\examples{
library(pipeR)
amChart() \%>>\% setType("pie")
library(pipeR)
amGraph() \%>>\% setType("type")
}

