% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartScrollbar.R
\docType{methods}
\name{setUpdateOnReleaseOnly,ChartScrollbar,logical-method}
\alias{setUpdateOnReleaseOnly,ChartScrollbar,logical-method}
\title{SETTER}
\usage{
\S4method{setUpdateOnReleaseOnly}{ChartScrollbar,logical}(.Object,
  updateOnReleaseOnly)
}
\description{
SETTER
}
\examples{
library(pipeR)
chartScrollbar() \%>>\% setUpdateOnReleaseOnly(TRUE)
}

