% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmLegend.R
\docType{methods}
\name{setUseGraphSettings,AmLegend,logical-method}
\alias{setUseGraphSettings,AmLegend,logical-method}
\title{SETTER}
\usage{
\S4method{setUseGraphSettings}{AmLegend,logical}(.Object, useGraphSettings)
}
\description{
SETTER
}
\examples{
amLegend() \%>>\% setUseGraphSettings(TRUE)
}

