% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setValueAxes}
\alias{setValueAxes}
\alias{setValueAxes,AmChart,list-method}
\title{Setter for ValueAxes}
\usage{
setValueAxes(.Object, valueAxes = NULL, ...)

\S4method{setValueAxes}{AmChart,list}(.Object, valueAxes)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{valueAxes}{object of class \code{list}.
Each element must be of class \linkS4class{ValueAxis}.}

\item{...}{Properties of \linkS4class{ValueAxis}.}
}
\value{
The updated object of class \linkS4class{AmChart}.
}
\description{
Add a list of ValueAxes objects to an Amchart.
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,valueAxes = list}: 
}}
\examples{
valueAxes <- list(valueAxis(axisTitleOffset = 12, tickLength = 10),
                  valueAxis(axisTitleOffset = 10, tickLength = 10))
setValueAxes(.Object = amChart(), valueAxes = valueAxes)
\dontrun{
lapply(valueAxes, listProperties)
amChart(valueAxes = valueAxes)
}
}

