% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{setValueAxesSettings,AmStockChart-method}
\alias{setValueAxesSettings,AmStockChart-method}
\title{Setter for valueAxesSettings}
\usage{
\S4method{setValueAxesSettings}{AmStockChart}(.Object,
  valueAxesSettings = NULL, ...)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{...}{Named properties of \code{ValueAxesSettings}.}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter for valueAxesSettings
}
\examples{
library(pipeR)
amStockChart() \%>>\% setValueAxesSettings(autoGridCount = "TRUE")
}
\seealso{
Other AmStockChart methods: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setDataSetSelector,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setMainDataSet,AmStockChart-method}};
  \code{\link{setPanels,AmStockChart,list-method}};
  \code{\link{setPanelsSettings,AmStockChart-method}};
  \code{\link{setPeriodSelector,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}}

Other AmStockChart setters: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setDataSetSelector,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setMainDataSet,AmStockChart-method}};
  \code{\link{setPanels,AmStockChart,list-method}};
  \code{\link{setPanelsSettings,AmStockChart-method}};
  \code{\link{setPeriodSelector,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}}
}

