% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Guide.R
\docType{methods}
\name{setValueAxis,Guide-method}
\alias{setValueAxis,Guide-method}
\title{SETTER}
\usage{
\S4method{setValueAxis}{Guide}(.Object, valueAxis = NULL, ...)
}
\description{
SETTER
}
\examples{
library(pipeR)
guide() \%>>\% setValueAxis(list(a = 1))
}

