% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setValueAxis,AmChart-method}
\alias{setValueAxis,AmChart-method}
\title{Setter for valueAxis}
\usage{
\S4method{setValueAxis}{AmChart}(.Object, valueAxis = NULL, ...)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmChart}}
with \code{type = "gantt"}.}

\item{valueAxis}{Object of class \code{\linkS4class{ValueAxis}}, default \code{NULL}.}

\item{...}{properties of \code{\linkS4class{ValueAxis}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter for valueAxis
}
\examples{
setValueAxis(.Object = amGanttChart())
}

