% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrendLine.R
\docType{methods}
\name{setValueAxisX,TrendLine-method}
\alias{setValueAxisX,TrendLine-method}
\title{SETTER}
\usage{
\S4method{setValueAxisX}{TrendLine}(.Object, valueAxisX = NULL, ...)
}
\description{
SETTER
}
\examples{
trendLine() \%>>\% setValueAxisX(title = "Hello !", axisTitleOffset = 12)
valueAxis <- valueAxis(title = "Hello !", axisTitleOffset = 12)
trendLine(valueAxis = valueAxis)
}

