% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmGraph.R
\docType{methods}
\name{setValueField}
\alias{setValueField}
\title{Setter}
\usage{
\S4method{setValueField}{AmGraph,character}(.Object, valueField)
}
\description{
Setter
}
\examples{
library(pipeR)
amGraph() \%>>\% setValueField("valueField")
}

