% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartCursor.R
\docType{methods}
\name{setValueLineAxis,ChartCursor-method}
\alias{setValueLineAxis,ChartCursor-method}
\title{SETTER}
\usage{
\S4method{setValueLineAxis}{ChartCursor}(.Object, valueLineAxis = NULL, ...)
}
\description{
SETTER
}
\examples{
chartCursor() \%>>\% setValueLineAxis( title = "Hello !", axisTitleOffset = 12 )
}

