% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{setX,Label,numeric-method}
\alias{setX,Label,numeric-method}
\title{SETTER}
\usage{
\S4method{setX}{Label,numeric}(.Object, x)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{Label}}}

\item{x}{Object of class \code{numeric}.
X position of a label.}
}
\value{
The updated .Object.
}
\description{
SETTER
}
\examples{
library(pipeR)
label() \%>>\% setX(16)
}

