% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{setY,Label,numeric-method}
\alias{setY,Label,numeric-method}
\title{SETTER}
\usage{
\S4method{setY}{Label,numeric}(.Object, y)
}
\description{
SETTER
}
\examples{
library(pipeR)
label() \%>>\% setY(16)
}

