% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sharedGenerics.R
\name{setBalloon}
\alias{addGuide}
\alias{setBalloon}
\alias{setDataProvider}
\alias{setExport}
\alias{setGraph}
\alias{setText}
\alias{setTitle}
\alias{setType}
\alias{setValueAxis}
\title{Shared generics methods}
\usage{
setBalloon(.Object, amBalloon = NULL, ...)

setDataProvider(.Object, dataProvider, keepNA = TRUE)

setExport(.Object, enabled = TRUE, ...)

setTitle(.Object, title)

setType(.Object, type)

setGraph(.Object, graph = NULL, ...)

addGuide(.Object, guide = NULL, ...)

setText(.Object, text)

setValueAxis(.Object, valueAxis = NULL, ...)
}
\arguments{
\item{.Object}{\linkS4class{AmObject}.}

\item{amBalloon}{\linkS4class{AmBalloon}.}

\item{...}{Other properties.}

\item{dataProvider}{\code{data.frame}.}

\item{keepNA}{\code{logical}, default \code{TRUE}.
Indicates if \code{NULL} values have to be kept or ignored.}

\item{enabled}{\code{logical}, default \code{TRUE}.
Should the export button be shown ?}

\item{title}{\code{character}.}

\item{type}{\code{character}.}

\item{graph}{\linkS4class{AmGraph}.}

\item{guide}{\linkS4class{Guide}.}

\item{text}{\code{character}.}

\item{valueAxis}{\linkS4class{ValueAxis}.}
}
\description{
These methods are shared by at least two different classes
}
\details{
\code{setBalloon} is shared by AmChart and AmStockChart

setDataProvider(..) is shared by AmGraph and DataSet

setExport(...) is Shared by AmChart and AmStockChart

setTitle(...) is Shared by AmGraph and ValueAxis

setType(...) is shared by AmGraph and AmChart

setGraph(...) is shared by AmChart and ChartScrollbar

addGuide(...) is shared by AxisBase and AmChart

setText(...) is shared by Title and Label

setValueAxis(...) is shared by AmChart(type = "gantt"), TrendLine and Guide
}

