% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinyExamples}
\alias{shinyExamples}
\title{Few shiny examples}
\usage{
shinyExamples()
}
\description{
Launch a shiny app with amChart examples
}
\examples{
\donttest{
library(pipeR)
library(shiny)
shinyExamples()
}
}

