% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmObject.R
\docType{methods}
\name{show,AmObject-method}
\alias{show,AmObject-method}
\title{Visualize with show}
\usage{
\S4method{show}{AmObject}(object)
}
\arguments{
\item{object:}{\code{\linkS4class{AmObject}} object}
}
\description{
Visualize with show
}
\examples{
new("AmChart")
}
\seealso{
\code{\linkS4class{AmObject}}

Other AmObject methods: \code{\link{print,AmObject-method}}

Other Visualizations: \code{\link{print,AmObject-method}}
}

