% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{show,AmStockChart-method}
\alias{show,AmStockChart-method}
\title{Visualize AmStockChart with show}
\usage{
\S4method{show}{AmStockChart}(object)
}
\arguments{
\item{object}{\linkS4class{AmStockChart}}
}
\seealso{
Other Visualizations: \code{\link{print,AmObject-method}},
  \code{\link{show,AmChart-method}},
  \code{\link{show,AmObject-method}}
}

