% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockEvent.R
\docType{class}
\name{StockEvent-class}
\alias{StockEvent-class}
\title{StockEvent class}
\description{
StockEvent class
}
\section{Slots}{

\describe{
\item{\code{stockGraph}}{\code{list} containing properties of stockGraph.
This is the graph on which event will be displayed.
You can use a reference to the stock graph object or id of the graph.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

