% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmGraph.R
\name{stockGraph}
\alias{stockGraph}
\title{Constructor for a stockGraph (class AmGraph)}
\usage{
stockGraph(animationPlayed = FALSE, balloonText, title, type, valueField, ...)
}
\arguments{
\item{animationPlayed}{\code{logical}.}

\item{balloonText}{\code{character}.
Balloon text. You can use tags like [[value]], [[description]], [[percents]], [[open]], [[category]]
or any other field name from your data provider. HTML tags can also be used.}

\item{title}{\code{character}. Graph title.}

\item{type}{\code{character}.
Type of the graph. Possible values are: "line", "column", "step", "smoothedLine", "candlestick", "ohlc".
XY and Radar charts can only display "line" otherArguments graphs.}

\item{valueField}{\code{character}.
Name of the value field in your dataProvider.}

\item{...}{Other properties}
}
\value{
An object of class \linkS4class{AmGraph}.
}
\description{
Constructor used for \linkS4class{AmStockChart}
}
\examples{
# --- constructor
stockGraph(balloonText = "balloonText", "type" = "column",
           valueField = "value", animationPlayed = TRUE)
           
}

