% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StockPanel.R
\docType{class}
\name{StockPanel-class}
\alias{StockPanel-class}
\title{StockPanel class}
\description{
StockPanel class
}
\section{Slots}{

\describe{
\item{\code{\code{drawOnAxis}:}}{Object of class \code{list}. Containing properties of ValueAxis.
Specifies on which value axis user can draw trend lines.
Set drawingIconsEnabled to true if you want drawing icons to be visible.
First value axis will be used if not set here.
You can use a reference to the value axis object or id of value axis.}

\item{\code{\code{stockGraphs}:}}{Object of class \code{list}.
Each element must be have been created with stockGraph(*)}

\item{\code{\code{stockLegend}:}}{Object of class \code{list}.
Each element must be have been created with stockLegend(*)}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

