% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockPanel.R
\name{stockPanel}
\alias{stockPanel}
\title{#’ Constructor.}
\usage{
stockPanel(drawOnAxis, stockGraphs, stockLegend, ...)
}
\arguments{
\item{\code{...}:}{{Properties of StockPanel.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/StockPanel}}}}
}
\value{
An \code{\linkS4class{StockPanel}} object
}
\description{
#’ Constructor.
}
\examples{
stockPanel()
}

