% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Title.R
\docType{class}
\name{Title-class}
\alias{Title-class}
\title{Title class}
\description{
Title class
}
\section{Slots}{

\describe{
\item{\code{\code{text}:}}{{Object of class \code{character}.
Text of a title.}}

\item{\code{\code{size}:}}{{Object of class \code{numeric}.
Text size of a title.}}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

