% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TrendLine.R
\docType{class}
\name{TrendLine-class}
\alias{TrendLine-class}
\title{TrendLine class}
\description{
TrendLine class
}
\section{Slots}{

\describe{
\item{\code{\code{finalValue}:}}{Object of class \code{numeric}.
Value at which trend line should end.}

\item{\code{\code{finalXValue}:}}{Object of class \code{numeric}.
Used by XY chart only. X value at which trend line should end.}

\item{\code{\code{initialValue}:}}{{Object of class \code{numeric}.
Value from which trend line should start.}}

\item{\code{\code{initialValue}:}}{{Object of class \code{numeric}.
Used by XY chart only. X value from which trend line should start.}}

\item{\code{\code{valueAxis}:}}{{Object of class \code{\linkS4class{ValueAxis}}.
Value axis of the trend line. Will use first value axis of the chart if not set any.
You can use a reference to the value axis object or id of value axis.}}

\item{\code{\code{valueAxisX}:}}{{Object of class \code{\linkS4class{ValueAxis}}.
Used by XY chart only. X axis of trend line.
Will use first X axis of the chart if not set any.
You can use a reference to the value axis object or id of value axis.}}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

