% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ValueAxis.R
\name{valueAxis}
\alias{valueAxis}
\title{Constructor for an AmGraph}
\usage{
valueAxis(title, ...)
}
\arguments{
\item{title}{Object of class \code{character}. Title of the axis.}

\item{...}{Properties of ValueAxis.
See \url{http://docs.amcharts.com/3/javascriptcharts/ValueAxis}}
}
\value{
An \code{\linkS4class{ValueAxis}} object
}
\description{
Constructor for an AmGraph
}
\examples{
valueAxis(title = "Hello !", axisTitleOffset = 12)
}

