% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_AmCharts.R
\name{add_export_dependency}
\alias{add_export_dependency}
\title{Add the export dependency to an htmlwidget}
\usage{
add_export_dependency(widget)
}
\arguments{
\item{widget}{An htmlwidget.}
}
\value{
Return the updated widget with the 'export' dependency.
}
\description{
Add the 'export' dependency to an htmlwidget.
You can only manipulate the htmlwidget if you call the method 'plot' with an rAmChart.
}
