% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addAxe}
\alias{addAxe}
\alias{addAxe,AmChart-method}
\title{Setter}
\usage{
addAxe(.Object, axe = NULL, ...)

\S4method{addAxe}{AmChart}(.Object, axe = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}}

\item{axe}{object of class \code{\linkS4class{GaugeAxis}}.}

\item{...}{properties of \code{\linkS4class{GaugeAxis}}}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
addAxe(.Object = amChart(), bandAlpha = 1)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{GaugeAxis}} S4 class

Other AmChart methods: \code{\link{addArrow}},
  \code{\link{addArrow,AmChart-method}};
  \code{\link{addGraph}},
  \code{\link{addGraph,AmChart-method}};
  \code{\link{addGuide,AmChart-method}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}},
  \code{\link{addLabel,AmChart-method}};
  \code{\link{addListener,AmObject,character,character-method}};
  \code{\link{addSegment}},
  \code{\link{addSegment,AmChart,numeric-method}};
  \code{\link{addSubData}},
  \code{\link{addSubData,AmChart,numeric,data.frame-method}};
  \code{\link{addTitle}},
  \code{\link{addTitle,AmChart-method}};
  \code{\link{addTrendLine}},
  \code{\link{addTrendLine,AmChart-method}};
  \code{\link{addValueAxes}},
  \code{\link{addValueAxes,AmChart-method}};
  \code{\link{setAllLabels}},
  \code{\link{setAllLabels,AmChart,list-method}};
  \code{\link{setArrows}},
  \code{\link{setArrows,AmChart-method}};
  \code{\link{setAxes}},
  \code{\link{setAxes,AmChart,list-method}};
  \code{\link{setBalloon}},
  \code{\link{setBalloon,AmChart-method}};
  \code{\link{setCategoryAxis}},
  \code{\link{setCategoryAxis,AmChart-method}};
  \code{\link{setCategoryField}},
  \code{\link{setCategoryField,AmChart,character-method}};
  \code{\link{setChartCursor}},
  \code{\link{setChartCursor,AmChart-method}};
  \code{\link{setChartScrollbar}},
  \code{\link{setChartScrollbar,AmChart-method}};
  \code{\link{setCreditsPosition}},
  \code{\link{setCreditsPosition,AmChart,character-method}};
  \code{\link{setDataProvider,AmChart,data.frame-method}},
  \code{\link{setDataProvider,DataSet,data.frame-method}};
  \code{\link{setExport}},
  \code{\link{setExport,AmChart-method}};
  \code{\link{setGraph,AmChart-method}},
  \code{\link{setGraph,ChartScrollbar-method}};
  \code{\link{setGraphs}},
  \code{\link{setGraphs,AmChart,list-method}};
  \code{\link{setGuides}},
  \code{\link{setGuides,AmChart,list-method}};
  \code{\link{setLegend}},
  \code{\link{setLegend,AmChart-method}};
  \code{\link{setSubChartProperties}},
  \code{\link{setSubChartProperties,AmChart-method}};
  \code{\link{setTitles}},
  \code{\link{setTitles,AmChart,list-method}};
  \code{\link{setTrendLines}},
  \code{\link{setTrendLines,AmChart,list-method}};
  \code{\link{setValueAxes}},
  \code{\link{setValueAxes,AmChart,list-method}};
  \code{\link{setValueAxis,AmChart-method}}

Other AmChart setters: \code{\link{addArrow}},
  \code{\link{addArrow,AmChart-method}};
  \code{\link{addGraph}},
  \code{\link{addGraph,AmChart-method}};
  \code{\link{addGuide,AmChart-method}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}},
  \code{\link{addLabel,AmChart-method}};
  \code{\link{addListener,AmObject,character,character-method}};
  \code{\link{addTitle}},
  \code{\link{addTitle,AmChart-method}};
  \code{\link{addTrendLine}},
  \code{\link{addTrendLine,AmChart-method}};
  \code{\link{addValueAxes}},
  \code{\link{addValueAxes,AmChart-method}};
  \code{\link{setAllLabels}},
  \code{\link{setAllLabels,AmChart,list-method}};
  \code{\link{setArrows}},
  \code{\link{setArrows,AmChart-method}};
  \code{\link{setAxes}},
  \code{\link{setAxes,AmChart,list-method}};
  \code{\link{setBalloon}},
  \code{\link{setBalloon,AmChart-method}};
  \code{\link{setCategoryAxis}},
  \code{\link{setCategoryAxis,AmChart-method}};
  \code{\link{setCategoryField}},
  \code{\link{setCategoryField,AmChart,character-method}};
  \code{\link{setChartCursor}},
  \code{\link{setChartCursor,AmChart-method}};
  \code{\link{setChartScrollbar}},
  \code{\link{setChartScrollbar,AmChart-method}};
  \code{\link{setCreditsPosition}},
  \code{\link{setCreditsPosition,AmChart,character-method}};
  \code{\link{setDataProvider,AmChart,data.frame-method}},
  \code{\link{setDataProvider,DataSet,data.frame-method}};
  \code{\link{setExport}},
  \code{\link{setExport,AmChart-method}};
  \code{\link{setGraph,AmChart-method}},
  \code{\link{setGraph,ChartScrollbar-method}};
  \code{\link{setGraphs}},
  \code{\link{setGraphs,AmChart,list-method}};
  \code{\link{setGuides}},
  \code{\link{setGuides,AmChart,list-method}};
  \code{\link{setLegend}},
  \code{\link{setLegend,AmChart-method}};
  \code{\link{setSubChartProperties}},
  \code{\link{setSubChartProperties,AmChart-method}};
  \code{\link{setTitles}},
  \code{\link{setTitles,AmChart,list-method}};
  \code{\link{setTrendLines}},
  \code{\link{setTrendLines,AmChart,list-method}};
  \code{\link{setValueAxes}},
  \code{\link{setValueAxes,AmChart,list-method}};
  \code{\link{setValueAxis,AmChart-method}}
}

