% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{addFieldMapping,DataSet-method}
\alias{addFieldMapping,DataSet-method}
\title{Add fieldMapping}
\usage{
\S4method{addFieldMapping}{DataSet}(.Object, ...)
}
\description{
Add fieldMapping
}
\examples{
library(pipeR)
dataSet() \%>>\% addFieldMapping(fromField ="val1", toField ="value")
}

