% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addSegment}
\alias{addSegment}
\title{Add a segment to a category of AmChart}
\usage{
\S4method{addSegment}{AmChart,numeric}(.Object, categoryIDs, sgts)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{AmChart}}.}

\item{\code{categoryIDs}:}{Object of class \code{numeric}.}

\item{\code{sgts}:}{Object of class \code{data.frame}
( or \code{list} of \code{data.frame} for multiple add ).}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Add a segment to a category of AmChart
}
\details{
Use this methode in case of an AmChart.
}
\examples{
library(pipeR)
amGanttChart( segmentsField = "segments", dataProvider = data.frame( category = c( "John", "Julia") ) ) \%>>\%
addSegment( 1, data.frame(start = 7, duration = 2:3, task = c("Task #1", "Task #2") ) ) \%>>\%
addSegment( 2, data.frame(start = 10, duration = 2:3, task = c("Task #1", "Task #2") ) )

ls <- list( data.frame(start = 7, duration = 2:3, task = c("Task #1", "Task #2") ),
data.frame(start = 10, duration = 2:3, task = c("Task #1", "Task #2") ) )
amGanttChart( segmentsField = "segments", dataProvider = data.frame( category = c( "John", "Julia") ) ) \%>>\%
addSegment( 1:2,  ls)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

Other AmChart methods: \code{\link{addArrow}};
  \code{\link{addAxe}}; \code{\link{addGraph}};
  \code{\link{addGuide}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}}; \code{\link{addListener}};
  \code{\link{addSubData}}; \code{\link{addTitle}};
  \code{\link{addTrendLine}}; \code{\link{addValueAxes}};
  \code{\link{initialize,AmChart-method}};
  \code{\link{initialize,AmStockChart-method}};
  \code{\link{setAllLabels}}; \code{\link{setArrows}};
  \code{\link{setAxes}}; \code{\link{setBalloon}};
  \code{\link{setCategoryAxis}};
  \code{\link{setCategoryField}};
  \code{\link{setChartCursor}};
  \code{\link{setChartScrollbar}};
  \code{\link{setCreditsPosition}};
  \code{\link{setDataProvider}},
  \code{\link{setDataProvider}}; \code{\link{setExport}};
  \code{\link{setGraphs}}; \code{\link{setGraph}},
  \code{\link{setGraph}}; \code{\link{setGuides}};
  \code{\link{setLegend}};
  \code{\link{setSubChartProperties}};
  \code{\link{setTitles}}; \code{\link{setTrendLines}};
  \code{\link{setValueAxes}}; \code{\link{setValueAxis}}
}

