% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{addStockEvent,DataSet-method}
\alias{addStockEvent,DataSet-method}
\title{Add stockEvent}
\usage{
\S4method{addStockEvent}{DataSet}(.Object, stockEvent = NULL, ...)
}
\description{
Add stockEvent
}
\examples{
dataSet()
}

