% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockPanel.R
\docType{methods}
\name{addStockGraph,StockPanel-method}
\alias{addStockGraph,StockPanel-method}
\title{Setter}
\usage{
\S4method{addStockGraph}{StockPanel}(.Object, stockGraph = NULL, ...)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{stockGraph}{(optionnal) Object of class \code{\linkS4class{AmGraph}}.
Use stockGraph(*) constructor}

\item{...}{Properties of the \code{\linkS4class{AmGraph}} to add.}
}
\value{
The updated \code{.Object}

The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter
}
\examples{
library(pipeR)
stockPanel() \%>>\% addStockGraph(comparable = FALSE)
stockPanel() \%>>\% addStockGraph( stockGraph(comparable = FALSE) )
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

\code{\linkS4class{AmGraph}} S4 class
}

