% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{addTitle}
\alias{addTitle}
\alias{addTitle,AmChart-method}
\title{Add a Title}
\usage{
addTitle(.Object, title = NULL, ...)

\S4method{addTitle}{AmChart}(.Object, title = NULL, ...)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{title}{object of class \code{\linkS4class{Title}}, default \code{NULL}.}

\item{...}{properties of class \code{\linkS4class{Title}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Add a Title
}
\section{Methods (by class)}{
\itemize{
\item \code{AmChart}: 
}}
\examples{
addTitle(.Object = amChart(), text = "balloonText")
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{Title}} S4 class
}

