% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Guide.R
\docType{methods}
\name{addValueAxis}
\alias{addValueAxis}
\title{Add a ValueAxis}
\usage{
\S4method{addValueAxis}{Guide}(.Object, valueAxis = NULL, ...)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{Guide}}.}

\item{\code{valueAxis}:}{Object of class \code{\linkS4class{ValuesAxes}}.}
}
\value{
The updated object of class \code{\linkS4class{Guide}}.
}
\description{
Add a ValueAxis
}
\examples{
library(pipeR)
guide() \%>>\% addValueAxis( axisTitleOffset = 12, tickLength = 10 )
}
\seealso{
\code{\linkS4class{Guide}} S4 class

\code{\linkS4class{ValueAxis}} S4 class
}

