% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmBalloon.R
\docType{class}
\name{AmBalloon-class}
\alias{AmBalloon-class}
\title{AmBalloon class}
\description{
AmBalloon class
}
\section{Slots}{

\describe{
\item{\code{\code{adjustBorderColor}:}}{Object of class \code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}

\item{\code{\code{color}:}}{Object of class \code{character}.
Color of text in the balloon.}

\item{\code{\code{cornerRadius}:}}{Object of class \code{numeric}.
Balloon corner radius.}

\item{\code{\code{cornerRadius}:}}{Object of class \code{character}.
Balloon background color. Usually balloon background color is set by the chart.
Only if "adjustBorderColor" is "true" this color will be used.}
}}
\section{Slots}{

}
\author{
DataKnowledge
}
\seealso{
Other rAmChart classes: \code{\link{AmChart-class}};
  \code{\link{AmGraph-class}}; \code{\link{AmObject-class}}
}

