% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmBalloon.R
\docType{class}
\name{AmBalloon-class}
\alias{AmBalloon-class}
\title{AmBalloon class}
\description{
Creates the balloons (tooltips) of the chart.
It follows the mouse cursor when you roll-over the data items.
The framework generates the instances automatically, you just have to adjust
the appearance to your needs.
}
\details{
Run \code{api("AmBalloon")} for more information and all avalaible properties.
}
\section{Slots}{

\describe{
\item{\code{adjustBorderColor}}{\code{logical}.
If TRUE, border color will be changed when user rolls-over the slice, graph, 
etc, instead of background color.}

\item{\code{color}}{\code{character}.
Balloon text color.}

\item{\code{cornerRadius}}{\code{numeric}.
Balloon corner radius.}

\item{\code{fillColor}}{\code{character}.
Balloon background color. It is usually defined by the chart itself. If
"adjustBorderColor" is set to TRUE, the balloon background color will be
equal to "fillColor".}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must be a character string.
See examples for details.
Inherited from \linkS4class{AmObject}.}

\item{\code{otherProperties}}{\code{list}
containing other avalaible properties not yet implemented in the package.
Inherited from \linkS4class{AmObject}.}

\item{\code{value}}{\code{numeric}.
Inherited from \linkS4class{AmObject}.}
}}

