% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmBalloon.R
\name{amBalloon}
\alias{amBalloon}
\title{Constructor for an AmBalloon}
\usage{
amBalloon(adjustBorderColor, color, cornerRadius, fillColor, ...)
}
\arguments{
\item{adjustBorderColor:}{Object of class \code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}

\item{color:}{Object of class \code{character}.
Color of text in the balloon.}

\item{cornerRadius:}{Object of class \code{numeric}.
Balloon corner radius.}

\item{fillColor:}{Object of class \code{character}.
Balloon background color. Usually balloon background color is set by the chart.
Only if "adjustBorderColor" is "true" this color will be used.}

\item{...:}{Properties of AmBalloon.
See \url{http://docs.amcharts.com/3/javascriptcharts/AmBalloon}.}
}
\value{
The updated \code{\linkS4class{AmBalloon}} object.
}
\description{
Constructor for an AmBalloon
}
\examples{
object <- amBalloon(adjustBorderColor = TRUE)
}
\seealso{
Other AmBalloon methods: \code{\link{initialize,AmBalloon-method}};
  \code{\link{setAdjustBorderColor}};
  \code{\link{setColor}}; \code{\link{setFillColor}}

Other rAmChart class constructors: \code{\link{amAngularGaugeChart}},
  \code{\link{amChart}}, \code{\link{amFunnelChart}},
  \code{\link{amGanttChart}}, \code{\link{amPieChart}},
  \code{\link{amRadarChart}}, \code{\link{amSerialChart}},
  \code{\link{amXYChart}}
}

