% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amBarplot.R
\name{amBarplot}
\alias{amBarplot}
\title{Plotting bar chart using rAmCharts}
\usage{
amBarplot(x, y, data, xlab = "", ylab = "", groups_color = NULL,
  horiz = FALSE, labelRotation = 0, stack_type = "none",
  layered = FALSE, show_values = FALSE, depth = 0,
  dataDateFormat = NULL, minPeriod = ifelse(!is.null(dataDateFormat), "DD",
  ""), ...)
}
\arguments{
\item{x}{\code{character}, column name for x-axis or \code{numeric} value of the corresponding column.
It is optional if argument \code{data} has row names.}

\item{y}{\code{character}, column name for y-axis or \code{numeric}  vector
of the corresponding column. If you want to display a grouped barchart
or a stacked barchart, y is a vector of characters or numerics.}

\item{data}{\code{data.frame} dataframe with values to display.
You can add a column "color" (character, colors in hexadecimal). You can
also add a column "description" (character) containing the text you want to
display when mouse is on the graphic ('<br>' for a new line).}

\item{xlab}{\code{character} label for x-axis.}

\item{ylab}{\code{character} label for y-axis.}

\item{groups_color}{\code{character} vector of colors in hexadecimal, 
same length as y.}

\item{horiz}{\code{boolean} TRUE for an horizontal chart, FALSE for a vertical one}

\item{labelRotation}{\code{numeric} Rotation angle of a label. Only horizontal axis' values can be rotated.
If you set this for vertical axis, the setting will be ignored. Possible values from -90 to 90.}

\item{stack_type}{\code{character}, "regular" if you wish stacked bars, "100" if
you want 100 percent stacked bars. Default is set to "none".}

\item{layered}{\code{boolean} TRUE for layered. If TRUE, stack_type must be set
to "none".}

\item{show_values}{\code{boolean} TRUE to display values.}

\item{depth}{\code{numeric} if > 0, chart is displayed in 3D. Value between 0 and 100.}

\item{dataDateFormat}{\code{character}, default 'NULL'. Even if your chart parses dates,
you can pass them as strings in your data - 
all you need to do is to set data date format and the chart will parse dates to date objects.
Check this page for available formats.
Please note that two-digit years (YY) as well as literal month names (MMM)  are NOT supported in this setting.}

\item{minPeriod}{Specifies the shortest period of your data.
This should be set only if dataDateFormat is not 'NULL'.
Possible period values:
fff - milliseconds, ss - seconds, mm - minutes, hh - hours, DD - days, MM - months, YYYY - years.
It's also possible to supply a number for increments, i.e. '15mm'
which will instruct the chart that your data is supplied in 15 minute increments.}

\item{...}{see \code{\link{amOptions}} for more options}
}
\value{
An object of class \linkS4class{AmChart}.
}
\description{
amBarplot computes a bar chart of the given values.
}
\details{
\strong{Notice about labels:}
if the chart has many columns, several labels might be hidden.
It depends on the width of the conatainer where the chart is displayed.
Zoom on the chart to see if the chart can contain all labels. If not, use the parameter labelRotation.
You can also add a cursor to your chart...
}
\examples{
stopifnot(require(pipeR))

#Basic Example : column chart
data_bar <- data.frame(country = c("USA", "China", "Japan", "Germany", 
                                   "UK", "France", "India", "Spain",
                                   "Netherlands", "Russia", "South Korea",
                                   "Canada"),
                       visits = c(3025, 1882, 1809, 1322, 1122, 1114, 
                                  984, 711, 665, 580, 443, 441),
                       color = c("#FF0F00", "#FF6600", "#FF9E01", "#FCD202",
                                 "#F8FF01", "#B0DE09", "#04D215", "#0D8ECF",
                                 "#0D52D1", "#2A0CD0", "#8A0CCF", "#CD0D74"),
                       stringsAsFactors = FALSE)

# test with label rotation
amBarplot(x = "country", y = "visits", data = data_bar, labelRotation = -90)

#horizontal bar
amBarplot(x = "country", y = "visits", data = data_bar, horiz = TRUE, labelRotation = -90)

#3D bar
amBarplot(x = "country", y = "visits", data = data_bar, depth = 15, labelRotation = -90)

#display values
amBarplot(x = "country", y = "visits", data = data_bar, show_values = TRUE, labelRotation = -90)

#grouped columns
data_gbar <- data.frame(year = c("2005", "2006", "2007", "2008", "2009"),
                        day = c("01/06/2005", "02/06/2005", "03/06/2005",
                                "04/06/2005", "05/06/2005"),
                        month = c("06/2005", "07/2005", "08/2005",
                                  "09/2005", "10/2005"),
                        income = c(23.5, 26.2, 30.1, 29.5, 24.6),
                        expenses = c(18.1, 22.8, 23.9, 25.1, 25),
                        stringsAsFactors = FALSE)

amBarplot(x = "year", y = c("income", "expenses"), data = data_gbar)

# Parse dates

# default label: firt day of each year
pipeR::pipeline(
  amBarplot(x = "year", y = c("income", "expenses"), data = data_gbar,
            dataDateFormat = "YYYY", minPeriod = "YYYY"),
  setChartCursor()
)

# default label: first day of each month
pipeR::pipeline(
  amBarplot(x = "month", y = c("income", "expenses"), data = data_gbar,
            dataDateFormat = "MM/YYYY", minPeriod = "MM"),
  setChartCursor()
)

pipeR::pipeline(
  amBarplot(x = "day", y = c("income", "expenses"), data = data_gbar,
            dataDateFormat = "DD/MM/YYYY"),
  setChartCursor()
)

#change groups colors
amBarplot(x = "year", y = c("income", "expenses"), data = data_gbar, 
      groups_color = c("#87cefa", "#c7158"))

#stacked bars
amBarplot(x = "year", y = c("income", "expenses"), data = data_gbar, stack_type = "regular")

#100\% stacked bars
amBarplot(x = "year", y = c("income", "expenses"), data = data_gbar, stack_type = "100")

#layered bars
amBarplot(x = "year", y = c("income", "expenses"), data = data_gbar, layered = TRUE)

#data with row names
dataset <- data.frame(get(x = "USArrests", pos = "package:datasets"))
amBarplot(y = c("Murder", "Assault", "UrbanPop", "Rape"), data = dataset, stack_type = "regular")
}

