% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_amBoxplot.R
\name{amBoxplot}
\alias{amBoxplot}
\alias{amBoxplot.data.frame}
\alias{amBoxplot.default}
\alias{amBoxplot.formula}
\alias{amBoxplot.matrix}
\title{Plotting boxplot using rAmCharts}
\usage{
amBoxplot(object, ...)

\method{amBoxplot}{default}(object, xlab = NULL, ylab = NULL, ylim = NULL,
  names = NULL, col = "#1e90ff", horizontal = FALSE, ...)

\method{amBoxplot}{data.frame}(object, id = NULL, xlab = NULL,
  ylab = NULL, ylim = NULL, col = NULL, horizontal = FALSE, ...)

\method{amBoxplot}{matrix}(object, use.cols = TRUE, xlab = NULL,
  ylab = NULL, ylim = NULL, col = NULL, horizontal = FALSE, ...)

\method{amBoxplot}{formula}(object, data = NULL, id = NULL, xlab = NULL,
  ylab = NULL, ylim = NULL, col = NULL, horizontal = FALSE, ...)
}
\arguments{
\item{object}{a vector, data.frame, a matrix, or a formula !}

\item{...}{Other parameters for \link{amOptions}.}

\item{xlab, ylab}{\code{character}, labels of the axis.}

\item{ylim}{the range of y values with sensible defaults.}

\item{names}{in case of vector, name on x-axis}

\item{col}{color(s) to be used to fill the boxplot}

\item{horizontal}{Boolean. Rotate boxplot ?}

\item{id}{in case of using a data.frame, column name of id for identify outliers}

\item{use.cols}{for matrix only. Boxplot on columns or rows ?}

\item{data}{a data.frame from which the variables in formula should be taken}
}
\value{
An object of class \linkS4class{AmChart}.
}
\description{
amBoxplot computes a boxplot of the given data values.
Can be a vector, a data.frame, or a matrix
}
\examples{
# ---
# formula
# ---

(obj <- amBoxplot(count ~ spray, data = InsectSprays))
print(obj)

# adding parameters
amBoxplot(count ~ spray, data = InsectSprays, ylim = c(0,50),
          xlab = "spray", col = c("darkblue", "gray"))

# transpose
amBoxplot(count ~ spray, data = InsectSprays, ylim = c(0,50), xlab = "spray", horizontal = TRUE)

# using a custom colum to identify outliers
InsectSprays$id <- paste0("ID : ", 1:nrow(InsectSprays))
amBoxplot(count ~ spray, data = InsectSprays, id = "id")

# parameter for amOptions
amBoxplot(count ~ spray, data = InsectSprays, main = "amcharts")

# ---
# data.frame
# ---

don <- data.frame(a = 1:10, b = 1:5)
amBoxplot(don, ylim = c(0,15))

# parameter for amOptions
amBoxplot(count ~ spray, data = InsectSprays, creditsPosition = "top-right")

# ---
# matrix
# ---

x <- matrix(nrow = 10, ncol = 5, rnorm(50))

amBoxplot(x) # on columns
colnames(x) <- LETTERS[1:5]
amBoxplot(x) # with names
amBoxplot(x, use.cols = FALSE, col = c("blue", "red"))

# parameter for amOptions
amBoxplot(x, export = TRUE, exportFormat = "SVG")

# ---
# vector
# ---

amBoxplot(rnorm(100))

# parameter for amOptions
amBoxplot(x, zoom = TRUE)
}

