% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amCandlestick.R
\name{amCandlestick}
\alias{amCandlestick}
\title{Plotting candlestick chart using rAmCharts}
\usage{
amCandlestick(data, period_type = "day", main = "", mainSize = 15,
  xlab = "", ylab = "", horiz = FALSE, scrollbar = FALSE,
  positive_color = "#7f8da9", negative_color = "#db4c3c")
}
\arguments{
\item{data}{\code{data.frame} dataframe with at least 5 columns : 
date (character yyyy-mm-dd), open (numeric), close (numeric), low (numeric),
high (numeric).}

\item{period_type}{\code{character} either 'day', 'month' or 'year'}

\item{main}{\code{character}, title of the graph}

\item{mainSize}{\code{numeric}, size of the title of the graph.}

\item{xlab}{\code{character} label for x-axis.}

\item{ylab}{\code{character} label for y-axis.}

\item{horiz}{\code{boolean} TRUE for an horizontal chart, FALSE for a vertical one}

\item{scrollbar}{\code{boolean} TRUE if you want an horizontal scrollbar to zoom}

\item{positive_color}{\code{character} color for positive values 
(in hexadecimal)}

\item{negative_color}{\code{character} color for negative values
(in hexadecimal)}
}
\description{
amCandlestick computes a candlestick chart of the given value.
}
\examples{

# Basic example
data_candle <- data.frame(date = c("2015-01-01", "2015-01-02", "2015-01-03",
                                   "2015-01-04", "2015-01-05", "2015-01-06",
                                   "2015-01-07", "2015-01-08", "2015-01-09",
                                   "2015-01-10", "2015-01-11", "2015-01-20"),
                          open = c(136.65, 135.26, 132.90, 134.94, 136.76, 131.11,
                                   123.12, 128.32, 128.29, 122.74, 117.01, 122.01),
                          close = c(136.4, 131.8, 135.2, 135.0, 134.0, 126.3, 
                                    125.0, 127.7, 124.0, 119.9, 117.0, 122.0),
                          low = c(134.15, 131.50, 128.30, 132.63, 132.00, 125.09,
                                  120.30, 126.50, 123.71, 119.65, 111.62, 119.82),
                          high = c(136.96, 135.95, 135.27, 137.24, 136.86, 133.00,
                                   127.75, 129.35, 128.30, 124.86, 118.50, 123.50), 
                          stringsAsFactors = FALSE)
                                   
amCandlestick(data = data_candle)

#Horizontal chart :
amCandlestick(data = data_candle, horiz = TRUE)

#With a scrollbar :
amCandlestick(data = data_candle, scrollbar = TRUE)

# datas over months
data_candle$date <- c("2015-01-01", "2015-02-01", "2015-03-01",
                      "2015-04-01", "2015-05-01", "2015-06-01",
                      "2015-07-01", "2015-08-01", "2015-09-01",
                      "2015-10-01", "2015-11-01", "2015-12-01")
                      
amCandlestick(data = data_candle, period_type = "month")

}

