% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_amCandlestick.R
\name{amCandlestick}
\alias{amCandlestick}
\title{Plotting candlestick chart using rAmCharts}
\usage{
amCandlestick(data, xlab = "", ylab = "", horiz = FALSE,
  positiveColor = "#7f8da9", negativeColor = "#db4c3c", names = c("low",
  "open", "close", "high"), dataDateFormat = NULL,
  minPeriod = ifelse(!is.null(dataDateFormat), "DD", ""), ...)
}
\arguments{
\item{data}{\code{data.frame} dataframe with at least 5 columns: 
category, open (numeric), close (numeric), low (numeric),
high (numeric). See \link{data_candleStick1} and \link{data_candleStick2}.}

\item{xlab}{\code{character} label for x-axis.}

\item{ylab}{\code{character} label for y-axis.}

\item{horiz}{\code{logical} TRUE for an horizontal chart, FALSE for a vertical one}

\item{positiveColor}{\code{character} color for positive values (in hexadecimal).}

\item{negativeColor}{\code{character} color for negative values (in hexadecimal).}

\item{names}{\code{character}, names for the tooltip. Default set to c("low", "open", "close", "high").}

\item{dataDateFormat}{\code{character}, default set to NULL. Even if your chart parses dates,
you can pass them as strings in your dataframe - 
all you need to do is to set data date format and the chart will parse dates to date objects.
Check this page for available formats.
Please note that two-digit years (YY) as well as literal month names (MMM)  are NOT supported in this setting.}

\item{minPeriod}{\code{character} minPeriod Specifies the shortest period of your data.
This should be set only if dataDateFormat is not NULL.
Possible period values:
fff - milliseconds, ss - seconds, mm - minutes, hh - hours, DD - days, MM - months, YYYY - years.
It's also possible to supply a number for increments, i.e. '15mm'
which will instruct the chart that your data is supplied in 15 minute increments.}

\item{...}{see \link{amOptions} for more options}
}
\description{
amCandlestick computes a candlestick chart of the given value.
}
\examples{
# load data
data("data_candleStick2")

# Basic example
amCandlestick(data = data_candleStick2)

# change colors
amCandlestick(data = data_candleStick2, positiveColor = "black", negativeColor = "green")

# naming the axes
amCandlestick(data = data_candleStick2, xlab = "categories", ylab = "values")

# Rotate the labels for x axis
amCandlestick(data = data_candleStick2, labelRotation = 90)

# change names
amCandlestick(data = data_candleStick2, names = c("min", "begin", "end", "max"))

#Horizontal chart :
amCandlestick(data = data_candleStick2, horiz = TRUE)

#  Parse date              
amCandlestick(data = data_candleStick2, dataDateFormat = "YYYY-MM-DD")

# datas over months
data_candleStick2$category <- c("2015-01-01", "2015-02-01", "2015-03-01",
                                "2015-04-01", "2015-05-01", "2015-06-01",
                                "2015-07-01", "2015-08-01", "2015-09-01",
                                "2015-10-01", "2015-11-01", "2015-12-01")

amCandlestick(data = data_candleStick2, dataDateFormat = "YYYY-MM-DD", minPeriod = "MM")

# decimal precision
require(pipeR)
amCandlestick(data = data_candleStick2, horiz = TRUE) \%>>\%
  setProperties(precision = 2) 
}
\seealso{
\itemize{
\item{\url{https://dataknowledge.github.io/introduction_ramcharts/}}
}
}

