% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_shinyUtils.R
\docType{methods}
\name{amChartsOutput}
\alias{amChartsOutput}
\alias{amChartsOutput,character,characterOrMissing-method}
\title{SHINY}
\usage{
amChartsOutput(outputId, type = NULL, width, height)

\S4method{amChartsOutput}{character,characterOrMissing}(outputId, type = NULL,
  width, height)
}
\arguments{
\item{outputId}{\code{character}.}

\item{type}{\code{character}
indicating the chart type.}

\item{width}{\code{character}.}

\item{height}{\code{character}.}
}
\description{
Widget output function for use in Shiny
}

