% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmChart_shinyUtils.R
\name{amChartsOutput}
\alias{amChartsOutput}
\title{SHINY}
\usage{
amChartsOutput(outputId, type = NULL, width = "100\%",
  height = "400px")
}
\arguments{
\item{outputId}{\code{character}, output variable to read the chart from.}

\item{type}{\code{character}, indicating the chart type.}

\item{width}{\code{character}, the width of the chart container.}

\item{height}{\code{character}, the height of the chart container.}
}
\description{
Widget output function for use in Shiny
}
