% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmChart_shinyUtils.R
\name{amChartsOutput}
\alias{amChartsOutput}
\title{SHINY}
\usage{
amChartsOutput(outputId, type = NULL, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{\code{character}.}

\item{type}{\code{character}
indicating the chart type.}

\item{width}{\code{character}.}

\item{height}{\code{character}.}
}
\description{
Widget output function for use in Shiny
}

