% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amHist.R
\name{amHist}
\alias{amHist}
\alias{amHist.numeric}
\title{Plot an histogram}
\usage{
amHist(x, ...)

\method{amHist}{numeric}(x, col = "gray", border = "gray", freq = TRUE,
  plot = TRUE, labels = FALSE, xlab, ylab, ylim, control_hist, ...)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{...}{Other parameters for \link{amOptions}.}

\item{col}{a colour to be used to fill the bars.}

\item{border}{a colour for the borders.}

\item{freq}{\code{logical}; if \code{TRUE}, the histogram graphic is a representation of frequencies,
the counts component of the result; if FALSE, probability densities,
component density, are plotted (so that the histogram has a total area of one).
Defaults to TRUE if and only if breaks are equidistant (and probability is not specified).}

\item{plot}{logical. If \code{TRUE} (default),
an histogram is plotted, otherwise a list of breaks and counts is returned.
In the latter case, a warning is used if (typically graphical)
arguments are specified that only apply to the \code{plot = TRUE} case.}

\item{labels}{\code{logical} default \code{FALSE}.
Additionally draw labels on top of bars.
if \code{TRUE}, draw the counts or rounded densities;
if labels is a \code{character}, draw itself.}

\item{xlab, ylab}{\code{character}, labels of the axis.}

\item{ylim}{the range of y values with sensible defaults.}

\item{control_hist}{(optional) named \code{list()} containing parameters for computing the histogram.}
}
\value{
An object of class \linkS4class{AmChart}.
}
\description{
amHist computes a histogram of the given data values.
If \code{plot = TRUE}, the resulting object of class \code{"histogram"} before it is returned.
}
\examples{
x <- replicate(1000, {
  if (round(runif(1))) {
    rnorm(1)
  } else {
    rnorm(1, mean = 5)
  }
})

# Default method
(object <- amHist(x = x))
print(object)

# Without plot
amHist(x = x, plot = FALSE)

# With options
amHist(x = x, border = "blue")
amHist(x = x, col = "lightblue", breaks = 100)
amHist(x = x, col = "grey", breaks = 100)
amHist(x = x, col = "gray")
amHist(x = x, main = "Histogram", ylab = "y-axis", xlab = "x-axis", col = "red")
amHist(x = x, main = "Histogram", ylab = "y-axis", xlab = "x-axis", ylim = c(10, 15))
amHist(x = x, main = "Histogram", ylab = "y-axis", xlab = "x-axis")

# options for computing the histogram
amHist(x = x, control_hist = list(breaks = "Scott"))


}

