% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmLegend.R
\docType{class}
\name{AmLegend-class}
\alias{AmLegend-class}
\title{AmLegend class}
\description{
AmLegend class
}
\section{Slots}{

\describe{
\item{\code{\code{useGraphSettings}:}}{{Object of class \code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

