% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmLegend.R
\docType{class}
\name{AmLegend-class}
\alias{AmLegend-class}
\title{AmLegend class}
\description{
AmLegend class
}
\section{Slots}{

\describe{
\item{\code{useGraphSettings}}{Object of class \code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}

\item{\code{listeners}}{Object of class \code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{Object of class \code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{object of class \code{numeric}.}
}}
\author{
DataKnowledge
}

