% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_amPlot.R
\name{amLines}
\alias{amLines}
\title{amLines adds a serie to a graph.}
\usage{
amLines(chart, x = NULL, y = NULL, type = c("points", "line",
  "smoothedLine"), col = "#0066cc", title)
}
\arguments{
\item{chart}{\linkS4class{AmChart}. Chart you wish to add the new serie.}

\item{x}{\code{numeric}, equivalent to y, deprecated.}

\item{y}{\code{numeric}.}

\item{type}{(optionnal) \code{character}. Possible values are : "l" for line, 
"p" for points, "sl" for smoothed line.}

\item{col}{\code{character}, color of the new serie.}

\item{title}{\code{character}, name of the new serie, used when legend is enabled.}
}
\description{
amLines adds a new serie to an existing serial chart.
}
\note{
It is supposed here that x or y corresponds to the y-axis, and the x-axis
is automatically linked to the x values of the chart "chart". That is why it makes
sense to give the y argument.
}
\examples{
require(pipeR)
amPlot(x = rnorm(100), type = 'sl') \%>>\%
  amLines(x = rnorm(100), type = "p")
  
\donttest{
amPlot(x = rnorm(100), type = 'sl') \%>>\%
  amLines(x = rnorm(100), col = "blue") \%>>\%
  amLines(x = rnorm(100), type = "sl") \%>>\%
  amLines(x = rnorm(100), type = "p")

# For an XY chart
x <- sort(rnorm(100))
y1 <- rnorm(100, sd = 10)
y2 <- rnorm(100, sd = 10)
y3 <- rnorm(100, sd = 10)
amPlot(x = x, y = y1) \%>>\%
  amLines(x = y2, col = "blue") \%>>\%
  amLines(x = y3, type = "p")
}

}

