% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_amPlot.R
\name{amLines}
\alias{amLines}
\title{amLines adds a serie to a graph.}
\usage{
amLines(chart, x = NULL, y = NULL, type, col, title)
}
\arguments{
\item{chart}{\linkS4class{AmChart}. Chart you wish to add the new serie.}

\item{x}{\code{numeric}, equivalent to y, deprecated.}

\item{y}{\code{numeric}.}

\item{type}{(optionnal) \code{character}. Possible values are : "l" for line, 
"p" for points, "sl" for smoothed line.}

\item{col}{\code{character}, color of the new serie.}

\item{title}{\code{character}, name of the new serie, used when legend is enabled.}
}
\description{
amLines adds a new serie to an existing serial chart.
}
\note{
It is supposed here that x or y corresponds to the y-axis, and the x-axis
is automatically linked to the x values of the chart "chart". That is why it makes
sense to give the y argument.
}
\examples{
library(pipeR)
library(data.table)

# ----------------
# EXAMPLES NUMERIC
# ----------------

# ---
# x, y plot if x is a factor or a character
# ---

xc <- paste("cat.", 1:100)
xf <- factor(xc)
y <- rnorm(length(xc))

# Change type
amPlot(x = xc, y = y, type = 'l')
amPlot(x = xf, y = y, type = 'l') # x can be either a character or a factor
amPlot(x = xc, y = y, type = "sl")
amPlot(x = xc, y = y, type = "st")
amPlot(x = xc, y = y, type = "p")
amPlot(x = xc, y = y,type = "b")

# ---
# x, y plot if x is a date (character)
# ---
start <- as.POSIXct('01-01-2015', format = '\%d-\%m-\%Y')
end <- as.POSIXct('31-12-2015', format = '\%d-\%m-\%Y')
date <- seq.POSIXt(from = start, to = end, by = 'day')
date <- format(date, '\%m-\%d-\%Y')

y <- rnorm(length(date))
amPlot(x = date, y = y, type = 'l') # default parseDates = FALSE

amPlot(x = date, y = y, type = 'l', parseDates = TRUE, dataDateFormat = "MM-DD-YYYY")


# ---
# x, y plot if both x, y are numeric
# ---

x <- sort(rnorm(100))
y <- rnorm(100, sd = 10)

# Change type
amPlot(x = x, y = y) # default type = "p"
amPlot(x = x, y = y, type = "l")

# Add a legend
amPlot(x = x, y = y, type = "l") \%>>\%
  setLegend(useGraphSettings = TRUE)

# Add a scrollbar
amPlot(x = x, y = y, scrollbar = TRUE)
amPlot(x = x, y = y, scrollbar = TRUE)

# Change lty
amPlot(x = x, y = y, type = "l", lty = 2)

# Add weight
weights <- rnorm(100, sd = 5)
amPlot(x = x, y = y, type = "p", weights = weights)

# Change cex
amPlot(x = x, y = y, type = "p", cex = 1)
amPlot(x = x, y = y, type = "p", cex = 10)


# Change color
amPlot(x = x, y = y, col = "blue")

col <- factor(c(rep(1,25), rep(2,50), rep(3,25)))
amPlot(x = x, y = y, col = col)

# Add title from amOptions
amPlot(x = x, y = y, main = "title")

# ---
# x plot
# ---
x <- rnorm(100)

# Simple scatter plot
amPlot(x = x)

# Add title
amPlot(x = x, main = "Title")

# Set a title
amPlot(x = x, main = "Title")

# Change type (set to "p" by default)
amPlot(x = x, type = "l")
amPlot(x = x, type = "sl")
amPlot(x = x, type = "st")
amPlot(x = x, type = "p")
amPlot(x = x, type = "b")

# Change color
amPlot(x = rnorm(100), col = "lightblue")

# Change bullets
x <- rnorm(100)
amPlot(x = x, bullet = "diamond")
amPlot(x = x, bullet = "square")
amPlot(x = x, bullet = "triangleUp")
amPlot(x = x, bullet = "triangleDown")
amPlot(x = x, bullet = "triangleLeft")
amPlot(x = x, bullet = "triangleRight")
amPlot(x = x, bullet = "bubble")
amPlot(x = x, bullet = "yError", error = sample(100))
amPlot(x = x, bullet = "xError", error = sample(100))

# Change lty
x <- rnorm(100)
amPlot(x = x, type = "b", lty = 1)
amPlot(x = x, type = "b", lty = 6)

# Change cex
x <- rnorm(100)
amPlot(x = x, cex = 1)
amPlot(x = x, cex = 10)

# Change lwd
x <- rnorm(100)
amPlot(x = x, type = "st", lwd = 1)
amPlot(x = x, type = "st", lwd = 2)

# Add a title from amOptions
amPlot(x = x, type = "st", lwd = 2, main = "My title")

# -------------------
# EXAMPLES DATA.FRAME
# -------------------

iris <- get("iris", "package:datasets")

# Select the column to plot with a character or a numeric
amPlot(iris, columns = "Sepal.Length")
amPlot(iris, columns = 1)

# Select several columns
amPlot(iris, col = colnames(iris)[1:2], type = c("l", "st"))

# Select all numeric columns
amPlot(iris)

# Add a legend 
amPlot(iris, col = colnames(iris)[1:2], type = c("l", "st")) \%>>\%
  setLegend(useGraphSettings = TRUE)

# Add parameter from amOptions
co2 <- data.table(get("CO2", "package:datasets"))
amPlot(co2, zoom = TRUE)

# -------------------
# ADD OTHER SERIE(S)
# -------------------

# For a simple chart
amPlot(x = rnorm(100), type = 'sl') \%>>\%
  amLines(x = rnorm(100), type = "p")

amPlot(x = rnorm(100), type = 'sl') \%>>\%
  amLines(x = rnorm(100), col = "blue") \%>>\%
  amLines(x = rnorm(100), type = "sl") \%>>\%
  amLines(x = rnorm(100), type = "p")

# For an XY chart
x <- sort(rnorm(100))
y1 <- rnorm(100, sd = 10)
y2 <- rnorm(100, sd = 10)
y3 <- rnorm(100, sd = 10)
amPlot(x = x, y = y1) \%>>\%
  amLines(x = y2, col = "blue") \%>>\%
  amLines(x = y3, type = "p")


# -------------------
# FORMULA
# -------------------
iris <- data.table(get("iris", "package:datasets"))

amPlot(x = iris$Sepal.Length, y = iris$Sepal.Width, xlab = "Sepal.Length")
amPlot(Sepal.Length~Sepal.Width, data = iris)
amPlot(Petal.Length + Sepal.Length ~ Sepal.Width, data = iris) \%>>\%
  setLegend(useGraphSettings = TRUE)

# Add parameter from amOptions
amPlot(Sepal.Length~Sepal.Width, data = iris, zoom = TRUE)
}

