% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmObject.R
\docType{class}
\name{AmObject-class}
\alias{AmObject-class}
\title{AmObject class}
\description{
This is a virtual class for representing any Am** class
}
\section{Slots}{

\describe{
\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must be a character string.}

\item{\code{otherProperties}}{\code{list}
containing other avalaible properties not yet implemented in the package.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

