% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_amOptions.R
\name{amOptions}
\alias{amOptions}
\title{amOptions}
\usage{
amOptions(chart, theme = "none", legend = FALSE, legendPosition = "right",
  legendAlign = "left", export = FALSE, exportFormat = character(),
  creditsPosition = "top-left", main = character(), mainColor = "#000000",
  mainSize = 15, zoom = FALSE, scrollbar = FALSE, scrollbarHeight = 20,
  labelRotation = 0)
}
\arguments{
\item{chart}{\linkS4class{AmChart}.}

\item{theme}{\code{character}, chn. Can be "none","light","dark","patterns","chalk",
default "none".}

\item{legend}{\code{logical}, default \code{FALSE}. Add a legend to the chart ?.}

\item{legendPosition}{\code{character}. Can be "left", "right", "top" or "bottom", default "right".}

\item{legendAlign}{\code{character} controls the legend alignement. Can be "left","right" or "center",
default "left". Only use if \code{legend = TRUE}.}

\item{export}{\code{logical}, default FALSE. Display export feature ?.}

\item{exportFormat}{\code{character} export format to keep, must be in JPG, PNG ,SVG,
CSV, JSON, PDF, XLSX, PRINT.}

\item{creditsPosition}{\code{character},  control credits position,
can be "top-left", "top-right", "bottom-left" or "bottom-right", default top-left.}

\item{main}{\code{character}, title of graphic, default empty \code{character}.}

\item{mainColor}{\code{character}, color of title (html-color), default "#000000".}

\item{mainSize}{\code{numeric}, color of title (html-color), default 15.}

\item{zoom}{\code{logical}, add a chart cursor, default FALSE.}

\item{scrollbar}{\code{logical}, default \code{FALSE}, display Scrollbar if \code{TRUE}.}

\item{scrollbarHeight}{\code{numeric}, Height in px, must be > 0.}

\item{labelRotation}{\code{numeric} Rotation angle of a label. Only horizontal axis' values can be rotated.
Must be between -90 and 90.}
}
\description{
Set the most common options for chart customization.
You can set other properties with the method \link{setProperties}.
See details for exception.
}
\details{
\strong{Exception:} 
\itemize{
 \item{It's impossible to export a gauge chart data as CSV.}
 }
}
\examples{
data(data_pie) 

library(pipeR)
#Export                 
amPie(data = data_pie)\%>>\%
  amOptions(export = TRUE)

#Legend
amPie(data = data_pie)\%>>\%
  amOptions(legend = TRUE)

#Legend position
amPie(data = data_pie)\%>>\%
  amOptions(legend = TRUE, legendPosition = "bottom")

#credits Position
amPie(data = data_pie)\%>>\%
  amOptions(creditsPosition = "bottom-right")

#credits Position
amPie(data = data_pie)\%>>\%
  amOptions(creditsPosition = "bottom-right")

#Theme
amPie(data = data_pie)\%>>\%
  amOptions(theme = "chalk")

#Title
amPie(data = data_pie)\%>>\%
  amOptions(main = "Social network", mainColor = "#FFFFFF", mainSize = 40, theme = "chalk")


#custom exemple
amPie(data = data_pie)\%>>\%
  amOptions(main = "Social network", mainColor = "#FFFFFF", mainSize = 40,
            theme = "dark", legend = TRUE, legendPosition = "bottom",
             creditsPosition = "bottom-right" )
}

