% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amOptions.R
\name{amOptions}
\alias{amOptions}
\title{amOptions}
\usage{
amOptions(chart, legend = FALSE, legendPosition = "right",
  legendAlign = "left", export = FALSE, exportFormat = NULL,
  creditsPosition = "top-left", theme = "none", main = "",
  mainColor = "#000000", mainSize = 15, zoom = FALSE, scrollbar = FALSE,
  scrollbarHeight = 20, ...)
}
\arguments{
\item{chart}{\linkS4class{AmChart}.}

\item{legend}{\code{boolean}, TRUE or FALSE, default FALSE, display legend if TRUE.}

\item{legendPosition}{\code{character}, control legend position,
can be "left", "right", "top" or "bottom", default "rigth". Only use if legend = TRUE.}

\item{legendAlign}{\code{character} control legend align, must be "left","rigth" or "center",
default "left". Only use if legend = TRUE.}

\item{export}{\code{boolean}, TRUE or FALSE, default FALSE, display export if TRUE}

\item{exportFormat}{\code{character} export format to keep, must be in JPG, PNG ,SVG,
CSV ,JSON, PDF, XLSX, PRINT}

\item{creditsPosition}{\code{character},  control credits position,
can be "top-left", "top-right", "bottom-left" or "bottom-right", default top-left}

\item{theme}{\code{character}, control theme.Can be "none","light","dark","patterns","chalk",
default "none}

\item{main}{\code{character}, title of graphic, default ""}

\item{mainColor}{\code{character}, color of title (html-color), default "#000000"}

\item{mainSize}{\code{numeric}, color of title (html-color), default 15}

\item{zoom}{\code{boolean}, activated zoom, default FALSE.}

\item{scrollbar}{\code{boolean}, TRUE or FALSE, default FALSE, display Scrollbar if TRUE}

\item{scrollbarHeight}{\code{numeric}, Height in px, must be >0}

\item{...}{Don't use...}
}
\description{
Most used options fot rAmCharts graphs customization
}
\examples{
data(data_pie) 

library(pipeR)
#Export                 
amPie(data = data_pie)\%>>\%
  amOptions(export = TRUE)

#Legend
amPie(data = data_pie)\%>>\%
  amOptions(legend = TRUE)

#Legend position
amPie(data = data_pie)\%>>\%
  amOptions(legend = TRUE, legendPosision = "bottom")

#credits Position
amPie(data = data_pie)\%>>\%
  amOptions(creditsPosition = "bottom-right")

#credits Position
amPie(data = data_pie)\%>>\%
  amOptions(creditsPosition = "bottom-right")

#Theme
amPie(data = data_pie)\%>>\%
  amOptions(theme = "chalk")

#Title
amPie(data = data_pie)\%>>\%
  amOptions(main = "Social network", mainColor = "#FFFFFF", mainSize = 40, theme = "chalk")


#custom exemple
amPie(data = data_pie)\%>>\%
  amOptions(main = "Social network", mainColor = "#FFFFFF", mainSize = 40,
            theme = "dark", legend = TRUE, legendPosision = "bottom",
             creditsPosition = "bottom-right" )
}
\author{
DataKnowledge
}

