% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amPie.R
\name{amPie}
\alias{amPie}
\title{Plotting pie chart using rAmCharts}
\usage{
amPie(data, show_values = TRUE, third_dim = FALSE, inner_radius = 0, ...)
}
\arguments{
\item{data}{\code{data.frame} dataframe with at least 2 columns : 
label (character), value (numeric).
You can add a third column "color" (character, colors in hexadecimal).}

\item{show_values}{\code{boolean} TRUE to display values.}

\item{third_dim}{\code{boolean} if TRUE, chart is displayed in 3D}

\item{inner_radius}{\code{numeric} value between 0 and 100}

\item{...}{see \code{\link{amOptions}} for more options}
}
\description{
amPie computes a pie chart of the given value.
}
\examples{
#Basic example
data_pie <- data.frame(label = c("Facebook", "Twitter", "LinkedIn", "Google+", 
                                  "Pinterest"),
                       value = c(38, 25, 15, 14, 8), stringsAsFactors = FALSE)
                       
amPie(data = data_pie)


#don't display values
amPie(data = data_pie, show_values = FALSE)

#3D pie
amPie(data = data_pie, third_dim = TRUE)

#donut chart
amPie(data = data_pie, inner_radius = 50)

#all
amPie(data = data_pie, inner_radius = 50, third_dim = TRUE, show_values = FALSE)
}

