% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_amPie.R
\name{amPie}
\alias{amPie}
\title{Plotting pie chart using rAmCharts}
\usage{
amPie(data, show_values = TRUE, depth = 0, inner_radius = 0, ...)
}
\arguments{
\item{data}{\code{data.frame}, dataframe with at least 2 columns : 
label (character), value (numeric). See \code{\link{data_pie}}
You can add a third column "color" (character, colors in hexadecimal).}

\item{show_values}{\code{logical}, TRUE to display values.}

\item{depth}{\code{numeric}, if > 0, chart is displayed in 3D, value
between 0 and 100}

\item{inner_radius}{\code{numeric}, value between 0 and 100}

\item{...}{see \code{\link{amOptions}} for more options.}
}
\description{
amPie computes a pie chart of the given value.
}
\examples{
\donttest{
# Load data
data("data_pie")

# Reference example : pie chart                
amPie(data = data_pie)

# Don't display values
amPie(data = data_pie, show_values = FALSE)

# 3D pie
amPie(data = data_pie, depth = 10)

# Donut chart
amPie(data = data_pie, inner_radius = 50)

# All parameters
amPie(data = data_pie, inner_radius = 50, depth = 10, show_values = FALSE)
}
}
\seealso{
\itemize{
\item{\url{https://dataknowledge.github.io/introduction_ramcharts/}}
}
}

