% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amPlot.R
\name{amPlot}
\alias{amPlot}
\alias{amPlot.data.frame}
\alias{amPlot.default}
\alias{amPlot.numeric}
\title{Plot data using rAmCharts}
\usage{
amPlot(x, ...)

\method{amPlot}{default}(x, ...)

\method{amPlot}{numeric}(x, y, bullet = "round", type = "p", col = "gray",
  export = FALSE, weights = NULL, scrollbar = FALSE, precision = 2,
  cursor = TRUE, id, error, xlab, ylab, main, lty, cex, lwd, xlim, ylim,
  hideYScrollbar, hideXScrollbar, ...)

\method{amPlot}{data.frame}(x, columns, type = "l", cursor = TRUE,
  scrollbar = FALSE, export = FALSE, legend = TRUE, precision = 2, xlab,
  ylab, main, ...)
}
\arguments{
\item{x}{the coordinates of points in the plot numeric,
data.frame, formula.}

\item{...}{Don't use... For S3 Definition...}

\item{y}{the y coordinates of points in the plot,
optional if x is an appropriate structure.}

\item{bullet}{default "round"}

\item{type}{what type of plot should be drawn, default "p".}

\item{col}{default "gray"}

\item{export}{default FALSE}

\item{weights}{for x/y charts only, default rep(1, x)}

\item{scrollbar}{default TRUE}

\item{precision}{default 2}

\item{cursor}{default TRUE}

\item{id}{for x/y charts only, default 1:length(x),}

\item{error}{only when type is "xError" "yError" default NULL,}

\item{xlab}{a title for the x axis, default NULL}

\item{ylab}{a title for the y axis, default NULL}

\item{main}{an overall title for the plot, default ""}

\item{lty}{line type (dashes)}

\item{cex}{bullet size}

\item{lwd}{line width}

\item{xlim}{range of x}

\item{ylim}{range of y}

\item{hideYScrollbar}{(\code{TRUE} by default)
optionnal for xy charts when scrollbar is enabled.
Specifies if Scrollbar of Y axis (vertical) should be hidden.}

\item{hideXScrollbar}{(\code{FALSE} by default)
optionnal for xy charts when scrollbar is enabled.
Specifies if Scrollbar of X axis (horizontal) should be hidden.}
}
\description{
amPlot computes a plot of the given data values.
Can be a vector, a data.frame, or a formula
}
\examples{
library(pipeR)
library(rAmCharts)

# ----------------
# EXAMPLES NUMERIC
# ----------------

# ---
# x, y plot
# ---

x <- sort(rnorm(100))
y <- rnorm(100, sd = 10)

# change type
amPlot(x = x, y = y) # default type = "p"
amPlot(x = x, y = y, type = "sl")
amPlot(x = x, y = y, type = "l")

# add a scrollbar
amPlot(x = x, y = y, scrollbar = TRUE)
amPlot(x = x, y = y, scrollbar = TRUE, hideYScrollbar = FALSE)

# change lty
amPlot(x = x, y = y, type = "l", lty = 1)

# add weight
weights <- rnorm(100, sd = 5)
amPlot(x = x, y = y, type = "p", weights = weights)

# change cex
amPlot(x = x, y = y, type = "p", cex = 1)
amPlot(x = x, y = y, type = "p", cex = 50)

# add title
amPlot(x = rnorm(100), y = rnorm(100), main = "title")

# ---
# x plot
# ---

amPlot(x = rnorm(100))

# add title
amPlot(x = rnorm(100), main = "Title")

# set a title
amPlot(x = rnorm(100), main = "Title")

# provide only x
amPlot(x = rnorm(100))

# change type (set to "l" by default)
x <- rnorm(100)
amPlot(x = rnorm(100, mean = 10))
amPlot(x = x, type = "sl")
amPlot(x = x, type = "st")
amPlot(x = x, type = "p")
amPlot(x = x, type = "b")

# change color
amPlot(x = rnorm(100), col = "lightblue")

# disable the cursor
amPlot(x = rnorm(100), cursor = FALSE)

# allow scrollbar
amPlot(x = rnorm(100), scrollbar = TRUE)

# change bullets
x <- rnorm(100)
amPlot(x = x, bullet = "diamond")
amPlot(x = x, bullet = "square")
amPlot(x = x, bullet = "triangleUp")
amPlot(x = x, bullet = "triangleDown")
amPlot(x = x, bullet = "triangleLeft")
amPlot(x = x, bullet = "triangleRight")
amPlot(x = x, bullet = "bubble")
amPlot(x = x, bullet = "yError", error = sample(100))
amPlot(x = x, bullet = "xError", error = sample(100))

# change lty
x <- rnorm(100)
amPlot(x = x, type = "b", lty = 1)
amPlot(x = x, type = "b", lty = 6)

# change cex
x <- rnorm(100)
amPlot(x = x, cex = 1)
amPlot(x = x, cex = 10)

# change lwd
x <- rnorm(100)
amPlot(x = x, type = "st", lwd = 1)
amPlot(x = x, type = "st", lwd = 2)

# -------------------
# EXAMPLES DATA.FRAME
# -------------------

iris <- data.table(get("iris", "package:datasets"))
amPlot(iris, columns = "Sepal.Length")

amPlot(iris, col = colnames(iris)[1:2], type = c("l", "st"))
amPlot(iris)

co2 <- data.table(get("CO2", "package:datasets"))
amPlot(co2)
library(pipeR)
library(rAmCharts)

# ----------------
# EXAMPLES NUMERIC
# ----------------

# ---
# x, y plot
# ---

x <- sort(rnorm(100))
y <- rnorm(100, sd = 10)

# change type
amPlot(x = x, y = y) # default type = "p"
amPlot(x = x, y = y, type = "sl")
amPlot(x = x, y = y, type = "l")

# add a scrollbar
amPlot(x = x, y = y, scrollbar = TRUE)
amPlot(x = x, y = y, scrollbar = TRUE, hideYScrollbar = FALSE)

# change lty
amPlot(x = x, y = y, type = "l", lty = 1)

# add weight
weights <- rnorm(100, sd = 5)
amPlot(x = x, y = y, type = "p", weights = weights)

# change cex
amPlot(x = x, y = y, type = "p", cex = 1)
amPlot(x = x, y = y, type = "p", cex = 50)

# add title
amPlot(x = rnorm(100), y = rnorm(100), main = "title")

# ---
# x plot
# ---

amPlot(x = rnorm(100))

# add title
amPlot(x = rnorm(100), main = "Title")

# set a title
amPlot(x = rnorm(100), main = "Title")

# provide only x
amPlot(x = rnorm(100))

# change type (set to "l" by default)
x <- rnorm(100)
amPlot(x = rnorm(100, mean = 10))
amPlot(x = x, type = "sl")
amPlot(x = x, type = "st")
amPlot(x = x, type = "p")
amPlot(x = x, type = "b")

# change color
amPlot(x = rnorm(100), col = "lightblue")

# disable the cursor
amPlot(x = rnorm(100), cursor = FALSE)

# allow scrollbar
amPlot(x = rnorm(100), scrollbar = TRUE)

# change bullets
x <- rnorm(100)
amPlot(x = x, bullet = "diamond")
amPlot(x = x, bullet = "square")
amPlot(x = x, bullet = "triangleUp")
amPlot(x = x, bullet = "triangleDown")
amPlot(x = x, bullet = "triangleLeft")
amPlot(x = x, bullet = "triangleRight")
amPlot(x = x, bullet = "bubble")
amPlot(x = x, bullet = "yError", error = sample(100))
amPlot(x = x, bullet = "xError", error = sample(100))

# change lty
x <- rnorm(100)
amPlot(x = x, type = "b", lty = 1)
amPlot(x = x, type = "b", lty = 6)

# change cex
x <- rnorm(100)
amPlot(x = x, cex = 1)
amPlot(x = x, cex = 10)

# change lwd
x <- rnorm(100)
amPlot(x = x, type = "st", lwd = 1)
amPlot(x = x, type = "st", lwd = 2)

# -------------------
# EXAMPLES DATA.FRAME
# -------------------

iris <- data.table(get("iris", "package:datasets"))
amPlot(iris, columns = "Sepal.Length")

amPlot(iris, col = colnames(iris)[1:2], type = c("l", "st"))
amPlot(iris)

co2 <- data.table(get("CO2", "package:datasets"))
amPlot(co2)
\dontrun{
amPlot(iris)
}
}

