% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amRadar.R
\name{amRadar}
\alias{amRadar}
\title{Plotting radar using rAmCharts}
\usage{
amRadar(data, col = NULL, backTransparency = 0.5, main = "",
  legend = TRUE, export = FALSE, type = "circles", mainSize = 15,
  pch = "round")
}
\arguments{
\item{data}{: first column is names "label", other columns are series of values}

\item{col}{: color of series}

\item{backTransparency}{: background transparency}

\item{main}{: title of graph}

\item{legend}{: add legend, TRUE or FALSE}

\item{export}{: add export, TRUE or FALSE}

\item{type}{: "polygons" or "circle", type of radar}

\item{mainSize}{: font size}

\item{pch}{: symbols}
}
\description{
radar computes a radarplot of the given data values.
}

