% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amRadar.R
\name{amRadar}
\alias{amRadar}
\title{Plotting radar using rAmCharts}
\usage{
amRadar(data, col = NULL, backTransparency = 0.5, main = "",
  legend = TRUE, export = FALSE, type = "polygons", mainSize = 15,
  pch = "round")
}
\arguments{
\item{data}{first column is names "label", other columns are series of values}

\item{col}{color of series}

\item{backTransparency}{background transparency}

\item{main}{title of graph}

\item{legend}{add legend, TRUE or FALSE}

\item{export}{add export, TRUE or FALSE}

\item{type}{"polygons" or "circle", type of radar}

\item{mainSize}{font size}

\item{pch}{symbols}
}
\description{
radar computes a radarplot of the given data values.
}
\examples{
data <- data.frame(label = c("A", "Z", "E", "R", "T"),
                   Product1 = c(1, 2, 3, 4, 2), 
                   Product2 = c(2, 8, 1, 1, 0),
                   Product3 = c(1,1,2,2,4))
amRadar(data)

amRadar(data, main = "My title", export = TRUE,
        col = c("#0000FF","#00FF00","#FF0000"), 
        backTransparency = c(0,0.4),
        type = c("polygons"),
        pch = "triangleRight")

}

