% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amRadar.R
\name{amRadar}
\alias{amRadar}
\title{Plotting radar using rAmCharts}
\usage{
amRadar(data, col = NULL, backTransparency = 0.5, type = "polygons",
  pch = "round", ...)
}
\arguments{
\item{data}{a data frame first column is named "label" (character), other columns are series of values,
see \code{\link{data_radar}}}

\item{col}{\code{character} color(s) of serie(s) hexadecimal like "#00FF00"}

\item{backTransparency}{\code{numeric} background transparency, between 0 and 1}

\item{type}{\code{character} "polygons" or "circle", type of radar}

\item{pch}{\code{character} symbols must be in "round", "square", "triangleUp", 
"triangleDown", "triangleLeft", "triangleRight", "bubble", 
"diamond", "xError", "yError"}

\item{...}{see \code{\link{amOptions}} for more options}
}
\description{
radar computes a radarplot of the given data values.
}
\examples{
require(pipeR)
#Data
data("data_radar")

#Basic
amRadar(data_radar)

#col
amRadar(data_radar, col = "#FF0000")
amRadar(data_radar, col = c("#0000FF", "#00FF00", "#FF0000"))


#backTransparency
amRadar(data_radar, backTransparency = 0.6)
amRadar(data_radar, backTransparency = c(0, 0.4, 0.6))


#type
amRadar(data_radar, type = "circles")

#pch
amRadar(data_radar,  pch = "triangleRight")
amRadar(data_radar,  pch = "triangleLeft")

}

