% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amRadar.R
\name{amRadar}
\alias{amRadar}
\title{Plotting radar using rAmCharts}
\usage{
amRadar(data, col = NULL, backTransparency = 0.5, type = "polygons",
  pch = "round", ...)
}
\arguments{
\item{data}{a data frame first column is named "label" (character), other columns are series of values}

\item{col}{\code{character} color(s) of serie(s) hexadecimal like "#00FF00"}

\item{backTransparency}{\code{numeric} background transparency, between 0 and 1}

\item{type}{\code{character} "polygons" or "circle", type of radar}

\item{pch}{\code{character} symbols must be in "round", "square", "triangleUp", 
"triangleDown", "triangleLeft", "triangleRight", "bubble", 
"diamond", "xError", "yError"}

\item{...}{see \code{\link{amOptions}} for more options}
}
\description{
radar computes a radarplot of the given data values.
}
\examples{
require(pipeR)

data <- data.frame(label = c("A", "Z", "E", "R", "T"),
                   Product1 = c(1, 2, 3, 4, 2), 
                   Product2 = c(2, 8, 1, 1, 0),
                   Product3 = c(1,1,2,2,4))

#Basic
amRadar(data)

#col
amRadar(data, col = "#FF0000")
amRadar(data, col = c("#0000FF", "#00FF00", "#FF0000"))


#backTransparency
amRadar(data, backTransparency = 0.6)
amRadar(data, backTransparency = c(0, 0.4, 0.6))


#type
amRadar(data, type = "circles")

#pch
amRadar(data,  pch = "triangleRight")
amRadar(data,  pch = "triangleLeft")


}

