% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amGauge.R
\name{amSolidGauge}
\alias{amSolidGauge}
\title{Plotting solid gauge using rAmCharts}
\usage{
amSolidGauge(x, min = 0, max = 100, type = "full", width = 20,
  color = "", text = "", textSize = 20)
}
\arguments{
\item{x}{\code{integer} an integer of the value for which the solid gauge is desired.}

\item{min}{\code{numeric} minimal value possible}

\item{max}{\code{numeric} maximal value possible}

\item{type}{\code{character} type of gauge : "full" or "semi"}

\item{width}{\code{numeric} width of the gauge}

\item{color}{\code{character} hexadecimal color value or a vector of colors}

\item{text}{\code{character} text}

\item{textSize}{\code{numeric} text size}
}
\description{
amSolidGauge computes a gauge of the given value.
}
\examples{
#Full gauge
amSolidGauge(x = 75, min = 0, max = 100, type = "full", width = 20, 
             color = "#1e90ff", text = "\%",
             textSize = 50)

#Demi gauge
amSolidGauge(x = 70, min = 0, max = 100, type = "semi", width = 20, 
             color = c("#00ff00", "#ffd700", "#ff0000"), text = "\%",
             textSize = 25)
}

