% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amGauge.R
\name{amSolidGauge}
\alias{amSolidGauge}
\title{Plotting solid gauge using rAmCharts}
\usage{
amSolidGauge(x, min = 0, max = 100, type = "full", width = 20,
  main = "", mainSize = 15, color = "", legend = "", legendSize = 20)
}
\arguments{
\item{x}{\code{integer} an integer of the value for which the solid gauge is desired.}

\item{min}{\code{numeric} minimal value possible}

\item{max}{\code{numeric} maximal value possible}

\item{type}{\code{character} type of gauge : "full" or "semi"}

\item{width}{\code{numeric} width of the gauge}

\item{main}{\code{character}, title of the graph.}

\item{mainSize}{\code{numeric}, size of the title of the graph.}

\item{color}{\code{character} hexadecimal color value or a vector of colors}

\item{legend}{\code{character} text legend}

\item{legendSize}{\code{numeric} text legend size}
}
\description{
amSolidGauge computes a gauge of the given value.
}
\examples{
amSolidGauge(x = 75, min = 0, max = 100, type = "full", width = 20,
             main = "Solid Gaude : full", color = "#1e90ff", legend = "\%",
             legendSize = 50)
amSolidGauge(x = 70, min = 0, max = 100, type = "semi", width = 20,
             main = "Solid Gauge : semi", color = c("#00ff00", "#ffd700", "#ff0000"), legend = "\%",
             legendSize = 25)
}

