% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{class}
\name{AmStockChart-class}
\alias{AmStockChart-class}
\title{AmStockChart}
\description{
AmStockChart
}
\section{Slots}{

\describe{
\item{\code{\code{categoryAxesSettings}:}}{Object of class \code{"list"}.
List of a \code{\linkS4class{CategoryAxes}} properties.
CategoryAxesSettings settings set's settings common for all CategoryAxes of StockPanels.
If you change a property after the chart is initialized,
you should call stockChart.validateNow() method in order for it to work.
If there is no default value specified, default value of CategoryAxis class will be used.
you should get this axis from the chart and set properties to this object.}

\item{\code{\code{ChartCursorSettings}:}}{Object of class \code{"list"}.
List of a \code{\linkS4class{ChartCursor}} class properties.
ChartCursorSettings settings set's settings for chart cursor.
If you change a property after the chart is initialized,
you should call stockChart.validateNow() method in order for it to work.
If there is no default value specified, default value of ChartCursor class will be used.}

\item{\code{\code{ChartScrollbarSettings}:}}{Object of class \code{"list"}.
List of a \code{\linkS4class{ChartScrollbar}} class properties.
ChartScrollbarSettings settings set's settings for chart scrollbar.
If you change a property after the chart is initialized,
you should call stockChart.validateNow() method in order for it to work.
If there is no default value specified, default value of ChartScrollbar class will be used.}

\item{\code{\code{comparedDataSets}:}}{Object of class \code{"list"}. Properties of data sets selected for comparing.}

\item{\code{\code{dataSets}:}}{Object of class \code{"list"}.
Each element must a list of DataSet properties.}

\item{\code{\code{dataSetSelector}:}}{Object of class \code{"list"}. DataSetSelector properties.
You can add it if you have more than one data set and want users to be able to select/compare them.}

\item{\code{\code{legendSettings}:}}{Object of class \code{"list"}. Legend settings.}

\item{\code{\code{mainDataSet}:}}{Object of class \code{"list"}. Data set selected as main.}

\item{\code{\code{panels}:}}{Object of class \code{"list"}.
Each element must be a list stockPanel properties.}

\item{\code{\code{panelsSettings}:}}{Object of class \code{"list"}. Settings for stock panels.}

\item{\code{\code{periodSelector}:}}{Object of class \code{"list"}.
Period selector object.
You can add it if you want user's to be able to enter
date ranges or zoom chart with predefined period buttons.}

\item{\code{\code{stockEventsSettings}:}}{Object of class \code{"list"}. Settings for stock events.}

\item{\code{\code{valueAxesSettings}:}}{Object of class \code{"list"}. Settings for value axes.}
}}
\author{
Dataknowledge
}
\seealso{
\code{\url{http://docs.amcharts.com/3/javascriptstockchart/AmStockChart}}
}

