% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmStockChart.R
\docType{class}
\name{AmStockChart-class}
\alias{AmStockChart-class}
\title{AmStockChart}
\description{
Class to draw stock charts
}
\section{Slots}{

\describe{
\item{\code{balloon}}{\linkS4class{AmBalloon}.}

\item{\code{comparedDataSets}}{\code{list} of \linkS4class{DataSet}.
Properties of data sets selected for comparison.}

\item{\code{dataSets}}{\code{list} of \linkS4class{DataSet}.
Each element must be a list of DataSet properties.}

\item{\code{dataSetSelector}}{\code{list} of \linkS4class{DataSetSelector}.
You can add it if you have more than one data set and want users
to be able to select/compare them.}

\item{\code{mainDataSet}}{\linkS4class{DataSet}.
Data set selected as main.}

\item{\code{panels}}{\code{list} of \linkS4class{StockPanel}.}

\item{\code{periodSelector}}{\linkS4class{PeriodSelector}.
You can add it if you want users to be able to enter
date ranges or zoom chart with predefined period buttons.}

\item{\code{theme}}{\code{character}}

\item{\code{type}}{equals "stock"}

\item{\code{listeners}}{\code{list} containining the listeners to add to the chart.
The list must be named as in the official API. Each element must be a character string.}

\item{\code{otherProperties}}{\code{list}
containing other avalaible properties not yet implemented in the package.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
datastorm-open
}
\seealso{
\url{http://docs.amcharts.com/3/javascriptstockchart/AmStockChart}
}

