% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amWaterfall.R
\name{amWaterfall}
\alias{amWaterfall}
\title{Plotting waterfall chart using rAmCharts}
\usage{
amWaterfall(data, start = 0, main = "", mainSize = 15, horiz = FALSE)
}
\arguments{
\item{data}{\code{data.frame} dataframe with at least 3 columns :
label (character), value (numeric), operation (character : "plus", "minus", "total").
Your can add a third column "color" (character, colors in hexadecimal).}

\item{start}{\code{numeric} value from which to start}

\item{main}{\code{character}, title of the graph}

\item{mainSize}{\code{numeric}, size of the title of the graph.}

\item{horiz}{\code{boolean} TRUE for an horizontal bullet chart, FALSE for a vertical one}
}
\description{
amWaterfall computes a waterfall chart of the given value.
}
\examples{
amWaterfall(data = data.frame(label = c("Income 1", "Income 2", "Income 3", "Total 1",
                                       "Expenses 1", "Expenses 2", "Total 2", "Income 4",
                                       "Income 5", "Income 6", "Expenses 3","Total 3",
                                       "Expenses 4", "Expenses 5", "Total 4"),
                              value = c(5, 10, 15, 30, 10, 5, 15, 50, 10, 35, 10, 100,
                                        15, 60, 25),
                              operation = c(rep("plus", 3), "total", rep("minus", 2),
                                            "total", "plus", "minus", rep("plus", 2),
                                            "total", rep("minus", 2), "total"),
                                            stringsAsFactors = FALSE),
            start = 0, main = "Waterfall Example", mainSize = 15, horiz = FALSE)
}

