% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_amWaterfall.R
\name{amWaterfall}
\alias{amWaterfall}
\title{Plotting waterfall chart using rAmCharts}
\usage{
amWaterfall(data, start = 0, horiz = FALSE, show_values = FALSE, ...)
}
\arguments{
\item{data}{\code{data.frame}, dataframe with at least 3 columns : 
label (character), value (numeric), operation (character : "plus", "minus", "total").
You can add a third column "color" (character, colors in hexadecimal). You can
also add a column "description" (character) containing the text you want to
display when mouse is on the graphic ('<br>' for a new line).
See \code{\link{data_waterfall}}.}

\item{start}{\code{numeric}, value from which to start.}

\item{horiz}{\code{logical}, TRUE for an horizontal chart, FALSE for a vertical one.}

\item{show_values}{\code{logical}, TRUE to display values on the chart.}

\item{...}{see \code{\link{amOptions}} for more options.}
}
\description{
amWaterfall computes a waterfall chart of the given value.
}
\examples{
\donttest{
# Load data
data("data_waterfall")

# Reference example : waterfall chart 
amWaterfall(data = data_waterfall, show_values = TRUE)


# Change the orientation :
amWaterfall(data = data_waterfall, horiz = TRUE)         
}
}
\seealso{
\itemize{
\item{\url{https://dataknowledge.github.io/introduction_ramcharts/}}
}
}

