% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amWind.R
\name{amWind}
\alias{amWind}
\title{Plotting wind using rAmCharts}
\usage{
amWind(data, col = NULL, backTransparency = 0.5)
}
\arguments{
\item{data}{: columns are series of values, from week wind (first column) to strong wind (last column)}

\item{col}{: color of series}

\item{backTransparency}{: background transparency}
}
\description{
wind computes a windplot of the given data values.
}
\examples{
##data
wind <- data.frame(Week = c(1, 2, 3, 4,1, 2,1, 2),
                   Middle = c(2, 8, 1, 1,2, 8,1, 2),
                   Strong = c(1, 1, 2, 2,1, 1 ,1, 2))
#Basic
amWind(wind)


#Color
amWind(wind, col = "#0404B4")
amWind(wind, col = c("#0404B4","#01DF01","#FFBF00"))

#backTransparency
amWind(wind, col = c("#0404B4","#01DF01","#FFBF00"), backTransparency = 0.1)
amWind(wind, col = c("#0404B4","#01DF01","#FFBF00"), backTransparency = 1)
amWind(wind, col = c("#0404B4","#01DF01","#FFBF00"), backTransparency = c(0.1, 0.1, 1))


}
\seealso{
\code{\link{amRadar}}
}

