% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amWind.R
\name{amWind}
\alias{amWind}
\title{Plotting wind using rAmCharts}
\usage{
amWind(data, col = NULL, backTransparency = 0.5, main = "",
  legend = TRUE, export = FALSE, fontSize = 15, pch = "round")
}
\arguments{
\item{data}{: columns are series of values, from week wind (first column) to strong wind (last column)}

\item{col}{: color of series}

\item{backTransparency}{: background transparency}

\item{main}{: title of graph}

\item{legend}{: add legend, TRUE or FALSE}

\item{export}{: add export, TRUE or FALSE}

\item{fontSize}{: font size}

\item{pch}{: symbols}
}
\description{
wind computes a windplot of the given data values.
}
\seealso{
\code{\link{amRadar}}
}

