% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_amWind.R
\name{amWind}
\alias{amWind}
\title{Plotting wind using rAmCharts}
\usage{
amWind(data, col = NULL, backTransparency = 0.5, ...)
}
\arguments{
\item{data}{: columns are series of values, from week wind (first column)
to strong wind (last column). See \link{data_wind}.}

\item{col}{\code{character} color(s) of serie(s) hexadecimal like "#00FF00".}

\item{backTransparency}{\code{numeric} background transparency, between 0 and 1.}

\item{...}{see \link{amOptions} for more options.}
}
\description{
wind computes a windplot of the given data values.
}
\examples{
# load data
data("data_wind")

# basic
amWind(data_wind)

# color
amWind(data = data_wind, col = "#0404B4")
amWind(data = data_wind, col = c("#0404B4","#01DF01","#FFBF00"))

# backTransparency
amWind(data = data_wind, col = c("#0404B4","#01DF01","#FFBF00"), backTransparency = 0.1)
amWind(data = data_wind, col = c("#0404B4","#01DF01","#FFBF00"), backTransparency = 1)
amWind(data = data_wind, col = c("#0404B4","#01DF01","#FFBF00"), backTransparency = c(0.1, 0.1, 1))


}
\seealso{
#' @seealso 
\itemize{
\item{\url{https://dataknowledge.github.io/introduction_ramcharts/}}
\item{\link{amRadar}}
}
}

