% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AxisBase.R
\docType{class}
\name{AxisBase-class}
\alias{AxisBase-class}
\title{AxisBase class}
\description{
Base class for ValueAxis and CategoryAxis. It can not be explicitly instantiated.
}
\section{Slots}{

\describe{
\item{\code{guides}}{\code{list}.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must be a character string.}

\item{\code{otherProperties}}{\code{list}
containing other avalaible properties not yet implemented in the package.}

\item{\code{value}}{\code{numeric}.
Guides of this axis. Use addGuide method.}
}}
\author{
datastorm-open
}

