% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AxisBase.R
\docType{class}
\name{AxisBase-class}
\alias{AxisBase-class}
\title{AxisBase class}
\description{
AxisBase class
}
\section{Slots}{

\describe{
\item{\code{\code{guides}:}}{Object of class \code{list}.
Guides belonging to this axis. Use addGuide method}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

