% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CategoryAxis.R
\docType{class}
\name{CategoryAxis-class}
\alias{CategoryAxis-class}
\title{CategoryAxis class}
\description{
CategoryAxis class
}
\section{Slots}{

\describe{
\item{\code{\code{gridPosition}:}}{{ Object of class \code{character}.
Specifies if a grid line is placed on the center of a cell or on the beginning of a cell.
Possible values are: "start" and "middle" This setting doesn't work if parseDates is set to true.}}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

