% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CategoryAxis.R
\name{categoryAxis}
\alias{categoryAxis}
\title{#’ Constructor.}
\usage{
categoryAxis(gridPosition, ...)
}
\arguments{
\item{\code{...}:}{{Properties of CategoryAxis.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/CategoryAxis}}}}
}
\value{
An \code{\linkS4class{CategoryAxis}} object
}
\description{
#’ Constructor.
}
\examples{
new("CategoryAxis", gridPosition = "start")
categoryAxis(gridPosition = "start", adjustBorderColor = TRUE)
}

