% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ChartCursor.R
\docType{class}
\name{ChartCursor-class}
\alias{ChartCursor-class}
\title{ChartCursor class}
\description{
Creates a cursor for the chart which follows the mouse movements.
In case of AmSerialChart charts it shows the balloons of hovered data points.
}
\details{
Run \code{api("ChartCursor")} for more information and all avalaible properties.
}
\section{Slots}{

\describe{
\item{\code{oneBalloonOnly}}{\code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}

\item{\code{valueLineAxis}}{\code{list}.
Properties of Axis of value line. If you set valueLineBalloonEnabled to true,
but you have more than one axis, you can use this property
to indicate which axis should display balloon.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

