% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ChartCursor.R
\name{chartCursor}
\alias{chartCursor}
\title{#’ Constructor.}
\usage{
chartCursor(animationDuration = 0.3, oneBalloonOnly, valueLineAxis, ...)
}
\arguments{
\item{\code{...}:}{{Properties of ChartCursor.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/ChartCursor}}}}
}
\value{
An \code{\linkS4class{ChartCursor}} object
}
\description{
#’ Constructor.
}
\examples{
chartCursor()
chartCursor(oneBalloonOnly = TRUE)
}

