% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ChartScrollbar.R
\docType{class}
\name{ChartScrollbar-class}
\alias{ChartScrollbar-class}
\title{ChartScrollbar class}
\description{
ChartScrollbar class
}
\section{Slots}{

\describe{
\item{\code{\code{updateOnReleaseOnly}:}}{Object of class \code{logical}.
Specifies if the chart should be updated while dragging/resizing the scrollbar
or only at the moment when user releases mouse button.}

\item{\code{\code{graph}:}}{Object of class \code{AmGraph}.
Specifies which graph will be displayed in the scrollbar.}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

