% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartScrollbar.R
\name{chartScrollbar}
\alias{chartScrollbar}
\title{Constructor for an AmGraph}
\usage{
chartScrollbar(graph, pathToImages = "http://www.amcharts.com/lib/3/images/",
  updateOnReleaseOnly, ...)
}
\arguments{
\item{graph}{Object of class \code{AmGraph}.
Specifies which graph will be displayed in the scrollbar.}

\item{pathToImages}{}

\item{updateOnReleaseOnly}{Object of class \code{logical}.
Specifies if the chart should be updated while dragging/resizing the scrollbar
or only at the moment when user releases mouse button.}

\item{...}{Properties of ChartScrollbar
See \url{http://docs.amcharts.com/3/javascriptcharts/ChartScrollbar}}
}
\value{
An \code{\linkS4class{ChartScrollbar}} object
}
\description{
Constructor for an AmGraph
}
\examples{
chartScrollbar()
chartScrollbar(updateOnReleaseOnly = TRUE)
}

