% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmChart_shinyUtils.R
\name{controlShinyPlot}
\alias{controlShinyPlot}
\title{Test the class of an exepression}
\usage{
controlShinyPlot(x)
}
\arguments{
\item{x}{expression passed to 'renderAmCharts'.
Either an expression that generates an HTML widget.
Or an expression that generates an AmChart.}
}
\description{
Only used in the package in 'renderAmCharts'.
}
\details{
This function has only an internal purpose. Never use it.
}

