% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmChart_shinyUtils.R
\name{controlShinyPlot}
\alias{controlShinyPlot}
\title{Tests the class of an expression.}
\usage{
controlShinyPlot(x)
}
\arguments{
\item{x}{expression passed to 'renderAmCharts'.
Either an expression that generates an HTML widget, or an
expression that generates an AmChart.}
}
\description{
Only used in 'renderAmCharts'.
}
\details{
This function has only an internal purpose. Never use it.
}

