% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DataSet.R
\docType{class}
\name{DataSet-class}
\alias{DataSet-class}
\title{DataSet class}
\description{
DataSet is an object which holds all information about data for \linkS4class{AmStockChart}
}
\details{
Run \code{api("DataSet")} for more information.
}
\section{Slots}{

\describe{
\item{\code{dataProvider}}{\code{list}, the data set data.
Important: the data sets need to come pre-ordered in ascending order.
Data with incorrect order might result in visual and functional glitches on the chart.}

\item{\code{fieldMappings}}{\code{list}, field mappings.
Field mapping is an object with fromField and toField properties.
fromField is the name of your value field in dataProvider.
toField might be chosen freely,
it will be used to set value/open/close/high/low fields for the StockChart.
Example: {fromField:"val1", toField:"value"}.}

\item{\code{stockEvents}}{\code{list} of \linkS4class{StockEvent}.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must be a character string.}

\item{\code{otherProperties}}{\code{list}
containing other avalaible properties not yet implemented in the package.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

