% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{class}
\name{DataSet-class}
\alias{DataSet-class}
\title{DataSet class}
\description{
DataSet class
}
\section{Slots}{

\describe{
\item{\code{\code{dataProvider}:}}{Object of class \code{list}. The data set data.
Important: the data sets need to come pre-ordered in ascending order.
Data with incorrect order might result in visual and functional glitches on the chart.}

\item{\code{\code{fieldMappings}:}}{Object of class \code{list}.
Array of field mappings.
Field mapping is an object with fromField and toField properties.
fromField is a name of your value field in dataProvider.
toField might be chosen freely,
it will be used to set value/open/close/high/low fields for the StockGraph.
Example: {fromField:"val1", toField:"value"}.}

\item{\code{\code{stockEvents}:}}{Object of class \code{list}.
Containing properties of stockEvents.}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

