% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{class}
\name{DataSet-class}
\alias{DataSet-class}
\title{DataSet class}
\description{
DataSet is objects which holds all information about data.
}
\details{
Run \code{api("DataSet")} for mor information.
}
\section{Slots}{

\describe{
\item{\code{dataProvider}}{\code{list}. The data set data.
Important: the data sets need to come pre-ordered in ascending order.
Data with incorrect order might result in visual and functional glitches on the chart.}

\item{\code{fieldMappings}}{\code{list} field mappings.
Field mapping is an object with fromField and toField properties.
fromField is a name of your value field in dataProvider.
toField might be chosen freely,
it will be used to set value/open/close/high/low fields for the StockGraph.
Example: {fromField:"val1", toField:"value"}.}

\item{\code{stockEvents}}{\code{list} of \linkS4class{StockEvent}.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

