% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeArrow.R
\docType{class}
\name{GaugeArrow-class}
\alias{GaugeArrow-class}
\title{GaugeArrow class}
\description{
GaugeArrow class
}
\section{Slots}{

\describe{
\item{\code{\code{axis}:}}{Object of class \code{list}
containing properties of \code{\linkS4class{GaugeAxis}}.
Axis of the arrow. You can use reference to the axis or id of the axis.
If you don't set any axis, the first axis of a chart will be used.}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

