% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GaugeArrow.R
\docType{class}
\name{GaugeArrow-class}
\alias{GaugeArrow-class}
\title{GaugeArrow class}
\description{
Creates an arrow for AmAngularGaugeChart, multiple can be assigned.
}
\details{
Run \code{api("GaugeArrow")} for more information and all avalaible properties.
}
\section{Slots}{

\describe{
\item{\code{axis}}{\code{list} containing properties of \linkS4class{GaugeAxis}.
Axis of the arrow. You can use reference to the axis or id of the axis.
If you don't set any axis, the first axis of a chart will be used.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must be a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list}
containing other avalaible properties not yet coded in the package.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

