% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeArrow.R
\name{gaugeArrow}
\alias{gaugeArrow}
\title{Constructor for a GaugeArrow}
\usage{
gaugeArrow(fillAlpha, alpha = 1, axis, ...)
}
\arguments{
\item{...}{Properties of GaugeArrow.
See \url{http://docs.amcharts.com/3/javascriptcharts/GaugeArrow}}
}
\value{
An \code{\linkS4class{GaugeArrow}} object
}
\description{
Constructor for a GaugeArrow
}
\examples{
gaugeArrow()
}

