% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeAxis.R
\docType{class}
\name{GaugeAxis-class}
\alias{GaugeAxis-class}
\title{GaugeAxis class}
\description{
GaugeAxis class
}
\section{Slots}{

\describe{
\item{\code{\code{bands}:}}{Object of class \code{list}
containing properties of one or several \code{\linkS4class{GaugeBand}} objects.
Bands are used to draw color fills between specified values.}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

