% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GaugeAxis.R
\name{gaugeAxis}
\alias{gaugeAxis}
\title{Constructor.}
\usage{
gaugeAxis(fillAlpha, axisAlpha = 1, bands, ...)
}
\arguments{
\item{\code{...}:}{{Properties of GaugeAxis.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/GaugeAxis}}}}
}
\value{
An \code{\linkS4class{GaugeAxis}} object
}
\description{
Constructor.
}
\examples{
gaugeAxis()
}

