% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GaugeBand.R
\docType{class}
\name{GaugeBand-class}
\alias{GaugeBand-class}
\title{GaugeBand class}
\description{
GaugeBand class
}
\section{Slots}{

\describe{
\item{\code{\code{id}:}}{Object of class \code{character}
Unique id of a band..}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

