% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaugeBand.R
\name{gaugeBand}
\alias{gaugeBand}
\title{Constructor.}
\usage{
gaugeBand(alpha = 1, id, ...)
}
\arguments{
\item{...}{Properties of GaugeBand.
See \url{http://docs.amcharts.com/3/javascriptcharts/GaugeBand}}
}
\value{
An \code{\linkS4class{GaugeBand}} object
}
\description{
Constructor.
}
\examples{
gaugeBand()
}

