% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GaugeBand.R
\docType{methods}
\name{initialize,GaugeBand-method}
\alias{initialize,GaugeBand-method}
\alias{gaugeBand}
\alias{setID}
\alias{setID,GaugeBand-method}
\title{Initializes a GaugeBand}
\usage{
\S4method{initialize}{GaugeBand}(.Object, alpha = 1, id, ...)

gaugeBand(alpha = 1, id, ...)

setID(.Object, id)

\S4method{setID}{GaugeBand}(.Object, id)
}
\arguments{
\item{.Object}{\linkS4class{GaugeBand} (or "GaugeBand" for initialize).}

\item{alpha}{\code{numeric}.}

\item{id}{\code{character}.}

\item{...}{other properties of GaugeBand.}
}
\value{
(updated) .Object of class \linkS4class{GaugeBand}.
}
\description{
Uses the constructor to create the object
or update an existing one with the setters.
}
\examples{
# --- method 'initialize'
new("GaugeBand")

# --- constructor
gaugeBand(alpha = 2, id = "band2")

# --- set the 'id'
setID(.Object = gaugeBand(), id = "1")

}
