% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Guide.R
\docType{class}
\name{Guide-class}
\alias{Guide-class}
\title{Guide class}
\description{
Guide class
}
\section{Slots}{

\describe{
\item{\code{fillAlpha}}{\code{numeric}.
Specifies if a grid line is placed on the center of a cell or on the beginning of a cell.
Possible values are: "start" and "middle" This setting doesn't work if parseDates is set to true.}

\item{\code{valueAxis}}{\code{list},
containing properties of a \linkS4class{ValueAxis} class.
As you can add guides directly to the chart, you might need to specify
which value axis should be used.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

