% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Guide.R
\docType{class}
\name{Guide-class}
\alias{Guide-class}
\title{Guide class}
\description{
Creates a horizontal/vertical guideline-/area for
amSerialChart, amXYChart and amRadarChart charts,
automatically adapts it's settings from the axes if none has been specified.
}
\details{
Run \code{api("Guide")} for more information and all avalaible properties.
}
\section{Slots}{

\describe{
\item{\code{fillAlpha}}{\code{numeric}.
Specifies if a grid line is placed on the center of a cell or on the beginning of a cell.
Possible values are: "start" and "middle" This setting doesn't work if parseDates is set to TRUE.}

\item{\code{valueAxis}}{\linkS4class{ValueAxis}.
As you can add guides directly to the chart, you might need to specify 
which value axis should be used.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must be a character string.}

\item{\code{otherProperties}}{\code{list}
containing other avalaible properties not yet implemented in the package.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

