% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmLegend.R
\docType{methods}
\name{initialize,AmLegend-method}
\alias{initialize,AmLegend-method}
\title{Initialize}
\usage{
\S4method{initialize}{AmLegend}(.Object, useGraphSettings, ...)
}
\description{
Initialize
}
\examples{
new("AmLegend", useGraphSettings = TRUE)
}

