% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmLegend.R
\docType{methods}
\name{initialize,AmLegend-method}
\alias{amLegend}
\alias{initialize,AmLegend-method}
\alias{legend}
\alias{setUseGraphSettings}
\alias{setUseGraphSettings,AmLegend,logical-method}
\title{Initialize}
\usage{
\S4method{initialize}{AmLegend}(.Object, useGraphSettings, ...)

amLegend(useGraphSettings, ...)

legend(useGraphSettings, ...)

setUseGraphSettings(.Object, useGraphSettings)

\S4method{setUseGraphSettings}{AmLegend,logical}(.Object, useGraphSettings)
}
\arguments{
\item{.Object}{\linkS4class{AmLegend}.}

\item{useGraphSettings}{\code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}

\item{...}{Other properties of \linkS4class{AmLegend}.
See \url{http://docs.amcharts.com/3/javascriptstockchart/AmLegend}.}
}
\description{
Constructor for an AmLegend.

Constructor for an AmLegend.
}
\examples{
new("AmLegend", useGraphSettings = TRUE)
amLegend(useGraphSettings = FALSE)
legend(useGraphSettings = FALSE)
setUseGraphSettings(.Object = amLegend(), useGraphSettings = TRUE)
}

