% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AxisBase_setters.R
\docType{methods}
\name{addGuide,AxisBase,GuideOrMissing-method}
\alias{addGuide,AxisBase,GuideOrMissing-method}
\title{Add a Guide for AxisBase}
\usage{
\S4method{addGuide}{AxisBase,GuideOrMissing}(.Object, guide = NULL, ...)
}
\arguments{
\item{.Object}{children class of \linkS4class{AxisBase}.}

\item{guide}{(optional) \linkS4class{Guide}.}

\item{...}{properties of \linkS4class{Guide}
Argument for method \code{addGuide}.}
}
\value{
(possibly updated) .Object of class \linkS4class{AxisBase}.
}
\description{
Update an of object of inherited class \linkS4class{AxisBase}.
}
\examples{
addGuide(.Object = valueAxis(), fillAlpha = .4, adjustBorderColor = TRUE, gridThickness = 1)
# equivalent to:
guide_obj <- guide(fillAlpha = .4, adjustBorderColor = TRUE, gridThickness = 1)
addGuide(.Object = valueAxis(), guide = guide_obj)

}

