% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ChartCursor.R
\docType{methods}
\name{initialize,ChartCursor-method}
\alias{initialize,ChartCursor-method}
\title{Initialize a ChartCursor}
\usage{
\S4method{initialize}{ChartCursor}(.Object, animationDuration = 0.3,
  oneBalloonOnly, ...)
}
\description{
Initialize a ChartCursor
}
\examples{
new("ChartCursor", oneBalloonOnly = TRUE)
}

