% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartScrollbar.R
\docType{methods}
\name{initialize,ChartScrollbar-method}
\alias{initialize,ChartScrollbar-method}
\title{Initialize a ChartScrollbar}
\usage{
\S4method{initialize}{ChartScrollbar}(.Object, graph,
  pathToImages = "http://www.amcharts.com/lib/3/images/", updateOnReleaseOnly,
  ...)
}
\arguments{
\item{graph}{Object of class \code{AmGraph}.
Specifies which graph will be displayed in the scrollbar.}

\item{pathToImages}{URL to API images (for buttons)}

\item{updateOnReleaseOnly}{Object of class \code{logical}.
Specifies if the chart should be updated while dragging/resizing the scrollbar
or only at the moment when user releases mouse button.}

\item{...}{Properties of ChartScrollbar
See \url{http://docs.amcharts.com/3/javascriptcharts/ChartScrollbar}}
}
\description{
Initialize a ChartScrollbar
}
\examples{
new("ChartScrollbar", graph = "g1")
new("ChartScrollbar", graph = amGraph(test = 1))
}

