% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockEvent.R
\docType{methods}
\name{initialize,StockEvent-method}
\alias{initialize,StockEvent-method}
\alias{setStockGraph}
\alias{setStockGraph,StockEvent-method}
\alias{stockEvent}
\title{Initialize a StockEvent}
\usage{
\S4method{initialize}{StockEvent}(.Object, backgroundAlpha = 1, stockGraph,
  ...)

stockEvent(backgroundAlpha = 1, stockGraph, ...)

setStockGraph(.Object, stockGraph = NULL, ...)

\S4method{setStockGraph}{StockEvent}(.Object, stockGraph = NULL, ...)
}
\arguments{
\item{.Object}{\linkS4class{StockEvent}.}

\item{backgroundAlpha}{\code{numeric}.}

\item{stockGraph}{\linkS4class{AmGraph} created with stockGraph(*).
This is the graph on which event will be displayed.
You can use a reference to the stock graph object or id of the graph.}

\item{...}{Other properties.}
}
\value{
(updated) argument .Object of class \linkS4class{StockEvent}.
}
\description{
Initialize a StockEvent
}
\section{Functions}{
\itemize{
\item \code{stockEvent}: 
}}
\examples{
new("StockEvent")
stockEvent()
setStockGraph(.Object = stockEvent(), stockGraph = stockGraph(balloonText = "balloonText"))
setStockGraph(.Object = stockEvent(), stockGraph = "stockGraph1")
}

