% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{class}
\name{Label-class}
\alias{Label-class}
\title{Label class}
\description{
Label class
}
\section{Slots}{

\describe{
\item{\code{bold}}{Object of class \code{character}.
Specifies if label is bold or not.}

\item{\code{text}}{Object of class \code{character}.
Text of a title.}

\item{\code{x}}{Object of class \code{numeric}.
X position of a label.}

\item{\code{y}}{Object of class \code{numeric}.
Y position of a label.}

\item{\code{listeners}}{Object of class \code{"list"} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{Object of class \code{"list"},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{Object of class \code{numeric}.}
}}
\author{
DataKnowledge
}

