% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label.R
\docType{class}
\name{Label-class}
\alias{Label-class}
\title{Label class}
\description{
Creates a label on the chart which can be placed anywhere, multiple can be assigned.
}
\details{
Run \code{api("Label")} for more information and all avalaible properties.
}
\section{Slots}{

\describe{
\item{\code{bold}}{\code{character}. Specifies if label is bold or not.}

\item{\code{text}}{\code{character}. Text of a title.}

\item{\code{x}}{\code{numeric}. X position of a label.}

\item{\code{y}}{\code{numeric}. Y position of a label.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

