% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\name{label}
\alias{label}
\title{Constructor for an AmGraph}
\usage{
label(text, bold, x, y, ...)
}
\arguments{
\item{text}{Object of class \code{character}.
Text of a title.}

\item{bold}{Object of class \code{character}.
Specifies if label is bold or not.}

\item{x}{Object of class \code{numeric}.
X position of a label.}

\item{y}{Object of class \code{numeric}.
Y position of a label.}

\item{...}{Properties of Label.
See \url{http://docs.amcharts.com/3/javascriptcharts/Label}}
}
\value{
An \code{\linkS4class{Label}} object
}
\description{
Constructor for an AmGraph
}
\examples{
new("Label", text = "bonjour")
label(text = "bonjour")
}

