% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\name{label}
\alias{label}
\title{#’ Constructor.}
\usage{
label(text, bold, x, y, ...)
}
\arguments{
\item{\code{...}:}{{Properties of Label.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/Label}}}}
}
\value{
An \code{\linkS4class{Label}} object
}
\description{
#’ Constructor.
}
\examples{
new("Label", text = "bonjour")
label(text = "bonjour")
}

