% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmBalloon.R
\docType{methods}
\name{listProperties,AmBalloon-method}
\alias{listProperties,AmBalloon-method}
\title{List attributes of an AmGraph object}
\usage{
\S4method{listProperties}{AmBalloon}(.Object)
}
\description{
This function is used to list attributes before addind to graphs (attribute of AmChart)
}
\examples{
library(pipeR)
amBalloon(adjustBorderColor = TRUE) \%>>\% listProperties
}

