% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart.R
\docType{methods}
\name{listProperties,AmChart-method}
\alias{listProperties,AmChart-method}
\title{List attributes of an AmChart}
\usage{
\S4method{listProperties}{AmChart}(.Object)
}
\arguments{
\item{.Object}{\code{\linkS4class{AmChart}}}
}
\description{
This method lists attributes of an AmChart to fit the API
}
\details{
For certain attributes we do not verify if they are NULL, see constructor.
}
\examples{
amChart(test = 1)
}

