% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmGraph.R, R/AmLegend.R, R/AmObject.R
\docType{methods}
\name{listProperties,AmGraph-method}
\alias{listProperties}
\alias{listProperties,AmGraph-method}
\alias{listProperties,AmLegend-method}
\alias{listProperties,AmObject-method}
\title{List attributes of an S4 object}
\usage{
\S4method{listProperties}{AmGraph}(.Object)

\S4method{listProperties}{AmLegend}(.Object)

listProperties(.Object)

\S4method{listProperties}{AmObject}(.Object)
}
\arguments{
\item{.Object}{any class object of the package}
}
\description{
Attributes of an AmGraph object.

Attributes of an AmLegend object.

List attributes of an S4 object
}
\examples{
listProperties(amGraph(balloonText = "toto", type = "type", valueField = "valueField"))
listProperties(amGraph(balloonText = "toto", type = "type"))
listProperties(amLegend(useGraphSettings = TRUE))
library(pipeR)
object <- amChart() \%>>\% addListener("select", "function onSelect (properties) {
     alert('selected nodes: ' + properties.nodes);}")
}

