% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{listProperties,Label-method}
\alias{listProperties,Label-method}
\title{List properties}
\usage{
\S4method{listProperties}{Label}(.Object)
}
\value{
Properties of the object in a list
}
\description{
List properties
}
\examples{
label(text = "balloonText")
}

