% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_PeriodSelector.R
\docType{class}
\name{PeriodSelector-class}
\alias{PeriodSelector-class}
\title{PeriodSelector}
\description{
Define the PeriodSelector properties.
}
\section{Slots}{

\describe{
\item{\code{periods}}{\code{list}.
Period object has 4 properties - period, count, label and selected.
Possible period values are:
"ss" - seconds, "mm" - minutes, "hh" - hours, "DD" - days, "MM" - months and "YYYY" - years.
property "count" specifies how many periods this button will select.
"label" will be displayed on a button and "selected" is a boolean
which specifies if this button is selected when chart is initialized or not.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{Object of class \code{numeric}.}
}}
\author{
Dataknowledge
}

