% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmCharts_plot.R
\docType{methods}
\name{plot,AmCharts-method}
\alias{plot,AmCharts-method}
\title{PLOTTING METHOD}
\usage{
\S4method{plot}{AmCharts}(x, y, width = "100\%", height = "500px",
  background = "#ffffff", ...)
}
\arguments{
\item{x}{\linkS4class{AmChart}}

\item{y}{unused}

\item{width}{\code{charcter}.}

\item{height}{\code{charcter}.}

\item{background}{\code{charcter}.}

\item{...}{Other properties.}
}
\description{
Basic method to plot an AmChart
}
\details{
Plots an object of class \code{\linkS4class{AmChart}}
}

