% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_AmCharts.R
\name{plot,AmCharts-method}
\alias{plot,AmCharts-method}
\title{PLOTTING METHOD}
\usage{
\S4method{plot}{AmCharts}(x, y, width = "100\%", height = NULL, background = "#ffffff", ...)
}
\arguments{
\item{x}{\linkS4class{AmChart}}

\item{y}{unused.}

\item{width}{\code{character}.}

\item{height}{\code{character}.B}

\item{background}{\code{character}.}

\item{...}{Other properties.}
}
\description{
Basic method to plot an AmChart
}
\details{
Plots an object of class \code{\linkS4class{AmChart}}
}
