% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmObject.R
\docType{methods}
\name{print,AmObject-method}
\alias{print,AmObject-method}
\title{Visualize with print}
\usage{
\S4method{print}{AmObject}(x, withDetail = TRUE, ...)
}
\arguments{
\item{x}{\linkS4class{AmChart}.}

\item{withDetail}{\code{logical}. Should the detail be printed ?}

\item{...}{Other properties.}
}
\description{
Display the object in the console.
}
\details{
The method 'print' can also plot the object of class \linkS4class{AmChart}.
}
\examples{
print(new("AmChart", categoryField = "variables", type = "serial"))
print(new("AmChart", categoryField = "variables", type = "serial"), withDetail = FALSE)
}

