% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmObject.R
\docType{methods}
\name{print,AmObject-method}
\alias{print,AmObject-method}
\title{Visualize with print}
\usage{
\S4method{print}{AmObject}(x, ...)
}
\arguments{
\item{x:}{an \code{\linkS4class{AmChart}}}
}
\description{
Visualize with print
}
\examples{
print(new("AmChart", categoryField = "variables"))
}
\seealso{
\code{\linkS4class{AmObject}}

Other AmObject methods: \code{\link{show,AmObject-method}}

Other Visualizations: \code{\link{show,AmObject-method}}
}

