% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmObject.R
\docType{methods}
\name{print,AmObject-method}
\alias{print,AmObject-method}
\title{Visualize with print}
\usage{
\S4method{print}{AmObject}(x, withDetail = TRUE, ...)
}
\arguments{
\item{x}{\linkS4class{AmChart}.}

\item{withDetail}{\code{logical}, TRUE to display details.}

\item{...}{Other properties.}
}
\description{
Display the object in the console.
}
\details{
If the object possess a 'dataProvider' property, it will be hidden in the console.
To see if it's correctly registered use '@dataProvider'.
}
\examples{
print(new("AmChart", categoryField = "variables", type = "serial"))
print(new("AmChart", categoryField = "variables", type = "serial"), withDetail = FALSE)
}

