% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules_timeSeries.R
\name{rAmCharts-shinymodules-ts}
\alias{rAmCharts-shinymodules-ts}
\alias{rAmChartsTimeSeriesUI}
\alias{rAmChartsTimeSeriesServer}
\title{Shiny module to render large time-series data with live server-client aggregation}
\usage{
rAmChartsTimeSeriesUI(id, width = "100\%", height = "400px")

rAmChartsTimeSeriesServer(input, output, session, data, col_date, col_series,
  maxPoints = 600, tz = "UTC", ts = c("5 min", "10 min", "30 min", "hour",
  "3 hour", "12 hour", "day", "week", "month", "year"), fun_aggr = "mean",
  treat_missing = FALSE, maxgap = Inf, type_aggr = "first", ...)
}
\arguments{
\item{id}{character, used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}

\item{width}{\code{character}, the width of the chart container. For \code{amChartsOutput}.}

\item{height}{\code{character}, the height of the chart container. For \code{amChartsOutput}.}

\item{input}{standard, \code{shiny} input}

\item{output}{standard, \code{shiny} output}

\item{session}{standard, \code{shiny} session}

\item{data}{: data.frame to transform. Must be a reactive object.}

\item{col_date}{: Date column name, default to "date".
Must be "POSIXct" or "CET24" colum}

\item{col_series}{: Column name of quantitative variable(s) to be
transformed. Default to setdiff(colnames(data), "date")}

\item{maxPoints}{: Maximal number of rows in results}

\item{tz}{: Timezone of result. Defaut to "UTC".}

\item{ts}{: All enabled aggregation. Default to c("5 min",  "10 min", "30 min", "hour", "3 hour", "12 hour", "day", "week", "month", "year").
Can be a number, in seconds, or a character string
containing one of "min", "hour", "day"....
This can optionally be preceded by a positive integer
and a space}

\item{fun_aggr}{: Aggregation function to use ("min", "max", "sum", "mean").
Default to "mean".}

\item{treat_missing}{: Boolean. Default to FALSE
Whether or not to interpolate missing values ?
see \code{na.approx}}

\item{maxgap}{: When interpolate missing values with \code{na.approx}.
Maximum number of consecutive NAs to fill. Defaut to Inf.}

\item{type_aggr}{: Character. Type of aggregation
\itemize{
 \item{"first"}{ : Date/Time result is equal to minimum of sequence, and this minimum is included in aggregation}
 \item{"last"}{ : Date/Time result is equal to maximum of sequence, and this maximum is included in aggregation}
}}

\item{...}{: \link{amTimeSeries} arguments}
}
\value{
a reactive expression with aggregate data and ts
}
\description{
Shiny module to render large time-series data with live server-client aggregation
}
\examples{

\dontrun{

# ui
rAmChartTimeSeriesUI("ts_1")

# server
callModule(rAmChartTimeSeriesServer, "ts_1", data_stock_2, "date", "value", 
 periodFieldsSelection = TRUE)

}

}
