% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmChart_shinyUtils.R
\name{renderAmCharts}
\alias{renderAmCharts}
\title{SHINY}
\usage{
renderAmCharts(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates an HTML widget.}

\item{env}{the environment in which to evaluate expr.}

\item{quoted}{is expr a quoted expression (with quote())?
This is useful if you want to save an expression in a variable.}
}
\description{
Widget output function for use in Shiny
}

