% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_examples.R
\name{runExamples}
\alias{runExamples}
\title{Run example with shiny}
\usage{
runExamples()
}
\description{
See some examples in a shiny web application.
Both 'am' functions and basic functions are illustrated.
}
\examples{

\dontrun{
if (interactive()) runExamples()
}
}
