% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setAxes}
\alias{setAxes}
\alias{setAxes,AmChart,list-method}
\title{Setter for axes}
\usage{
setAxes(.Object, axes, ...)

\S4method{setAxes}{AmChart,list}(.Object, axes)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}}

\item{axes}{object of class \code{\linkS4class{GaugeAxis}}}

\item{...}{Other properties}
}
\description{
Setter for axes
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,axes = list}: 
}}
\examples{
axes <- list(gaugeAxis(value = 130), gaugeAxis(value = 150))
setAxes(.Object = amChart(), axes = axes)
amChart(axes = axes)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class
}

