% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{setBold,Label,logical-method}
\alias{setBold,Label,logical-method}
\title{SETTER}
\usage{
\S4method{setBold}{Label,logical}(.Object, bold)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{Label}}}

\item{bold}{Object of class \code{character}.
Specifies if label is bold or not.}
}
\value{
The updated .Object.
}
\description{
SETTER
}
\examples{
library(pipeR)
label() \%>>\% setBold(TRUE)
}

