% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{setBold,Label,logical-method}
\alias{setBold,Label,logical-method}
\title{SETTER}
\usage{
\S4method{setBold}{Label,logical}(.Object, bold)
}
\description{
SETTER
}
\examples{
label() \%>>\% setBold(TRUE)
}

