% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setCategoryAxis}
\alias{setCategoryAxis}
\title{Setter}
\usage{
\S4method{setCategoryAxis}{AmChart}(amChart, categoryAxis = NULL, ...)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{AmChart}}.}

\item{\code{categoryAxis}:}{Object of class \code{\linkS4class{CategoryAxis}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\examples{
library(pipeR)
# Setter for categoryAxis
amChart() \%>>\% setCategoryAxis(categoryAxis(gridPosition = "start"))
amChart() \%>>\% setCategoryAxis(gridPosition = "start")
\dontrun{
# The argument categoryAxis must be an object of class CategoryAxis
amChart() \%>>\% setCategoryAxis(categoryAxis = "error")
}
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{CategoryAxis}} S4 class

Other AmChart methods: \code{\link{addArrow}};
  \code{\link{addAxe}}; \code{\link{addGraph}};
  \code{\link{addGuide}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}}; \code{\link{addListener}};
  \code{\link{addSegment}}; \code{\link{addSubData}};
  \code{\link{addTitle}}; \code{\link{addTrendLine}};
  \code{\link{addValueAxes}};
  \code{\link{initialize,AmChart-method}};
  \code{\link{initialize,AmStockChart-method}};
  \code{\link{setAllLabels}}; \code{\link{setArrows}};
  \code{\link{setAxes}}; \code{\link{setBalloon}};
  \code{\link{setCategoryField}};
  \code{\link{setChartCursor}};
  \code{\link{setChartScrollbar}};
  \code{\link{setCreditsPosition}};
  \code{\link{setDataProvider}},
  \code{\link{setDataProvider}}; \code{\link{setExport}};
  \code{\link{setGraphs}}; \code{\link{setGraph}},
  \code{\link{setGraph}}; \code{\link{setGuides}};
  \code{\link{setLegend}};
  \code{\link{setSubChartProperties}};
  \code{\link{setTitles}}; \code{\link{setTrendLines}};
  \code{\link{setValueAxes}}; \code{\link{setValueAxis}}

Other AmChart setters: \code{\link{addArrow}};
  \code{\link{addAxe}}; \code{\link{addGraph}};
  \code{\link{addGuide}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}}; \code{\link{addListener}};
  \code{\link{addTitle}}; \code{\link{addTrendLine}};
  \code{\link{addValueAxes}}; \code{\link{setAllLabels}};
  \code{\link{setArrows}}; \code{\link{setAxes}};
  \code{\link{setBalloon}}; \code{\link{setCategoryField}};
  \code{\link{setChartCursor}};
  \code{\link{setChartScrollbar}};
  \code{\link{setCreditsPosition}};
  \code{\link{setDataProvider}},
  \code{\link{setDataProvider}}; \code{\link{setExport}};
  \code{\link{setGraphs}}; \code{\link{setGraph}},
  \code{\link{setGraph}}; \code{\link{setGuides}};
  \code{\link{setLegend}};
  \code{\link{setSubChartProperties}};
  \code{\link{setTitles}}; \code{\link{setTrendLines}};
  \code{\link{setValueAxes}}; \code{\link{setValueAxis}}
}

