% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmBalloon.R
\docType{methods}
\name{setColor}
\alias{setColor}
\title{Setter for color}
\usage{
\S4method{setColor}{AmBalloon,character}(.Object, color)
}
\arguments{
\item{.Object:}{Object of class \code{\linkS4class{AmBalloon}}.}
}
\value{
The updated object of class \code{\linkS4class{AmBalloon}}.
}
\description{
Setter for color
}
\examples{
library(pipeR)
amBalloon() \%>>\% setColor("#000000")
}
\seealso{
\code{\linkS4class{AmBalloon}} S4 class

Other AmBalloon methods: \code{\link{amBalloon}};
  \code{\link{initialize,AmBalloon-method}};
  \code{\link{setAdjustBorderColor}};
  \code{\link{setFillColor}}
}

