% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmBalloon.R
\docType{methods}
\name{setCornerRadius}
\alias{setCornerRadius}
\title{Setter for corner radius}
\usage{
\S4method{setCornerRadius}{AmBalloon,numeric}(.Object, cornerRadius)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{AmBalloon}}.}

\item{\code{cornerRadius}:}{Object of class \code{numeric}.}
}
\value{
The updated object of class \code{\linkS4class{AmBalloon}}.
}
\description{
Setter for corner radius
}
\examples{
library(pipeR)
amBalloon() \%>>\% setCornerRadius(5)
}
\seealso{
\code{\linkS4class{AmBalloon}}  S4 class

Other AmBalloon setters: \code{\link{setAdjustBorderColor}};
  \code{\link{setFillColor}}
}

