% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setCreditsPosition}
\alias{setCreditsPosition}
\alias{setCreditsPosition,AmChart,character-method}
\title{Setter}
\usage{
setCreditsPosition(.Object, creditsPosition)

\S4method{setCreditsPosition}{AmChart,character}(.Object, creditsPosition)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{creditsPosition}{object of class \code{character}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,creditsPosition = character}: 
}}
\examples{
setCreditsPosition(.Object = amChart(), creditsPosition = "top-right")
}
\seealso{
\code{\linkS4class{AmChart}} S4 class
}

