% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{setDataSetSelector}
\alias{setDataSetSelector}
\title{Setter}
\usage{
\S4method{setDataSetSelector}{AmStockChart}(.Object, ...)
}
\arguments{
\item{\code{.Object}:}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{\code{...}:}{Property of datSetSelector}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter
}
\examples{
library(pipeR)
amStockChart() \%>>\% setDataSetSelector(width = 180)
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

Other AmStockChart methods: \code{\link{addComparedDataSet}};
  \code{\link{addDataSet}}; \code{\link{addPanel}};
  \code{\link{setCategoryAxesSettings}};
  \code{\link{setChartCursorSettings}};
  \code{\link{setChartScrollbarSettings}};
  \code{\link{setComparedDataSets}};
  \code{\link{setDataSets}};
  \code{\link{setLegendSettings}};
  \code{\link{setMainDataSet}};
  \code{\link{setPanelsSettings}}; \code{\link{setPanels}};
  \code{\link{setPeriodSelector}};
  \code{\link{setStockEventsSettings}};
  \code{\link{setValueAxesSettings}}

Other AmStockChart setters: \code{\link{addComparedDataSet}};
  \code{\link{addDataSet}}; \code{\link{addPanel}};
  \code{\link{setCategoryAxesSettings}};
  \code{\link{setChartCursorSettings}};
  \code{\link{setChartScrollbarSettings}};
  \code{\link{setComparedDataSets}};
  \code{\link{setDataSets}};
  \code{\link{setLegendSettings}};
  \code{\link{setMainDataSet}};
  \code{\link{setPanelsSettings}}; \code{\link{setPanels}};
  \code{\link{setPeriodSelector}};
  \code{\link{setStockEventsSettings}};
  \code{\link{setValueAxesSettings}}
}

