% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockPanel.R
\docType{methods}
\name{setDrawOnAxis,StockPanel-method}
\alias{setDrawOnAxis,StockPanel-method}
\title{SETTER}
\usage{
\S4method{setDrawOnAxis}{StockPanel}(.Object, valueAxis = NULL, ...)
}
\description{
SETTER
}
\examples{
stockPanel() \%>>\% setDrawOnAxis( valueAxis() )
}

