% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R, R/ChartScrollbar.R
\docType{methods}
\name{setGraph}
\alias{setGraph}
\title{Setter for graph}
\usage{
\S4method{setGraph}{AmChart}(.Object, graph = NULL, ...)

\S4method{setGraph}{ChartScrollbar}(.Object, graph = NULL, ...)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmChart}}.}

\item{graph}{(optionnal) Object of class \code{\linkS4class{AmGraph}}.}

\item{...}{Properties of AmGraph
See \url{http://docs.amcharts.com/3/javascriptcharts/AmGraph}}

\item{\code{.Object}:}{Object of class \code{\linkS4class{AmChart}} and type \code{gantt}.}

\item{\code{graph}:}{Object of class \code{\linkS4class{AmGraph}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.

The updated object of class \code{\linkS4class{ChartScrollbar}}.
}
\description{
Setter for graph

Setter for graph
}
\details{
Method to use in case of AmChart of type \code{gantt}.
For other type see \code{\link{setGraphs}} or \code{\link{addGraph}}.

Method to use in case of AmChart of type \code{gantt}.
For other type see \code{\link{setGraphs}} or \code{\link{addGraph}}.
}
\examples{
library(pipeR)
amGanttChart() \%>>\% setGraph()
library(pipeR)
chartScrollbar() \%>>\% setGraph(test = 1)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{ChartScrollbar}} S4 class

Other AmChart methods: \code{\link{addArrow}};
  \code{\link{addAxe}}; \code{\link{addGraph}};
  \code{\link{addGuide}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}}; \code{\link{addListener}};
  \code{\link{addSegment}}; \code{\link{addSubData}};
  \code{\link{addTitle}}; \code{\link{addTrendLine}};
  \code{\link{addValueAxes}};
  \code{\link{initialize,AmChart-method}};
  \code{\link{initialize,AmStockChart-method}};
  \code{\link{setAllLabels}}; \code{\link{setArrows}};
  \code{\link{setAxes}}; \code{\link{setBalloon}};
  \code{\link{setCategoryAxis}};
  \code{\link{setCategoryField}};
  \code{\link{setChartCursor}};
  \code{\link{setChartScrollbar}};
  \code{\link{setCreditsPosition}};
  \code{\link{setDataProvider}},
  \code{\link{setDataProvider}}; \code{\link{setExport}};
  \code{\link{setGraphs}}; \code{\link{setGuides}};
  \code{\link{setLegend}};
  \code{\link{setSubChartProperties}};
  \code{\link{setTitles}}; \code{\link{setTrendLines}};
  \code{\link{setValueAxes}}; \code{\link{setValueAxis}}

Other AmChart setters: \code{\link{addArrow}};
  \code{\link{addAxe}}; \code{\link{addGraph}};
  \code{\link{addGuide}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}}; \code{\link{addListener}};
  \code{\link{addTitle}}; \code{\link{addTrendLine}};
  \code{\link{addValueAxes}}; \code{\link{setAllLabels}};
  \code{\link{setArrows}}; \code{\link{setAxes}};
  \code{\link{setBalloon}}; \code{\link{setCategoryAxis}};
  \code{\link{setCategoryField}};
  \code{\link{setChartCursor}};
  \code{\link{setChartScrollbar}};
  \code{\link{setCreditsPosition}};
  \code{\link{setDataProvider}},
  \code{\link{setDataProvider}}; \code{\link{setExport}};
  \code{\link{setGraphs}}; \code{\link{setGuides}};
  \code{\link{setLegend}};
  \code{\link{setSubChartProperties}};
  \code{\link{setTitles}}; \code{\link{setTrendLines}};
  \code{\link{setValueAxes}}; \code{\link{setValueAxis}}
}

