% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setGuides}
\alias{setGuides}
\alias{setGuides,AmChart,list-method}
\title{Setter}
\usage{
setGuides(.Object, guides)

\S4method{setGuides}{AmChart,list}(.Object, guides)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{guides}{Object of class \code{list}.
Each element must be an object of class \code{\linkS4class{Guide}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,guides = list}: 
}}
\examples{
library(pipeR)
guides <- list(guide(fillAlpha = .1), guide(fillAlpha = .5))
amChart() \%>>\% setGuides(guides)
amChart(guides = guides)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{Guide}} S4 class
}

