% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setGuides}
\alias{setGuides}
\alias{setGuides,AmChart,list-method}
\title{Setter}
\usage{
setGuides(.Object, guides)

\S4method{setGuides}{AmChart,list}(.Object, guides)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{guides}{Object of class \code{list}.
Each element must be an object of class \code{\linkS4class{Guide}}.}
}
\value{
The updated object of class \code{\linkS4class{AmChart}}.
}
\description{
Setter
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AmChart,guides = list}: 
}}
\examples{
library(pipeR)
guides <- list(guide(fillAlpha = .1), guide(fillAlpha = .5))
amChart() \%>>\% setGuides(guides)
amChart(guides = guides)
}
\seealso{
\code{\linkS4class{AmChart}} S4 class

\code{\linkS4class{Guide}} S4 class

Other AmChart methods: \code{\link{addArrow}},
  \code{\link{addArrow,AmChart-method}};
  \code{\link{addAxe}},
  \code{\link{addAxe,AmChart-method}};
  \code{\link{addGraph}},
  \code{\link{addGraph,AmChart-method}};
  \code{\link{addGuide,AmChart-method}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}},
  \code{\link{addLabel,AmChart-method}};
  \code{\link{addListener,AmObject,character,character-method}};
  \code{\link{addSegment}},
  \code{\link{addSegment,AmChart,numeric-method}};
  \code{\link{addSubData}},
  \code{\link{addSubData,AmChart,numeric,data.frame-method}};
  \code{\link{addTitle}},
  \code{\link{addTitle,AmChart-method}};
  \code{\link{addTrendLine}},
  \code{\link{addTrendLine,AmChart-method}};
  \code{\link{addValueAxes}},
  \code{\link{addValueAxes,AmChart-method}};
  \code{\link{setAllLabels}},
  \code{\link{setAllLabels,AmChart,list-method}};
  \code{\link{setArrows}},
  \code{\link{setArrows,AmChart-method}};
  \code{\link{setAxes}},
  \code{\link{setAxes,AmChart,list-method}};
  \code{\link{setBalloon}},
  \code{\link{setBalloon,AmChart-method}};
  \code{\link{setCategoryAxis}},
  \code{\link{setCategoryAxis,AmChart-method}};
  \code{\link{setCategoryField}},
  \code{\link{setCategoryField,AmChart,character-method}};
  \code{\link{setChartCursor}},
  \code{\link{setChartCursor,AmChart-method}};
  \code{\link{setChartScrollbar}},
  \code{\link{setChartScrollbar,AmChart-method}};
  \code{\link{setCreditsPosition}},
  \code{\link{setCreditsPosition,AmChart,character-method}};
  \code{\link{setDataProvider,AmChart,data.frame-method}},
  \code{\link{setDataProvider,DataSet,data.frame-method}};
  \code{\link{setExport}},
  \code{\link{setExport,AmChart-method}};
  \code{\link{setGraph,AmChart-method}},
  \code{\link{setGraph,ChartScrollbar-method}};
  \code{\link{setGraphs}},
  \code{\link{setGraphs,AmChart,list-method}};
  \code{\link{setLegend}},
  \code{\link{setLegend,AmChart-method}};
  \code{\link{setSubChartProperties}},
  \code{\link{setSubChartProperties,AmChart-method}};
  \code{\link{setTitles}},
  \code{\link{setTitles,AmChart,list-method}};
  \code{\link{setTrendLines}},
  \code{\link{setTrendLines,AmChart,list-method}};
  \code{\link{setValueAxes}},
  \code{\link{setValueAxes,AmChart,list-method}};
  \code{\link{setValueAxis,AmChart-method}}

Other AmChart setters: \code{\link{addArrow}},
  \code{\link{addArrow,AmChart-method}};
  \code{\link{addAxe}},
  \code{\link{addAxe,AmChart-method}};
  \code{\link{addGraph}},
  \code{\link{addGraph,AmChart-method}};
  \code{\link{addGuide,AmChart-method}},
  \code{\link{addGuide,AxisBase-method}};
  \code{\link{addLabel}},
  \code{\link{addLabel,AmChart-method}};
  \code{\link{addListener,AmObject,character,character-method}};
  \code{\link{addTitle}},
  \code{\link{addTitle,AmChart-method}};
  \code{\link{addTrendLine}},
  \code{\link{addTrendLine,AmChart-method}};
  \code{\link{addValueAxes}},
  \code{\link{addValueAxes,AmChart-method}};
  \code{\link{setAllLabels}},
  \code{\link{setAllLabels,AmChart,list-method}};
  \code{\link{setArrows}},
  \code{\link{setArrows,AmChart-method}};
  \code{\link{setAxes}},
  \code{\link{setAxes,AmChart,list-method}};
  \code{\link{setBalloon}},
  \code{\link{setBalloon,AmChart-method}};
  \code{\link{setCategoryAxis}},
  \code{\link{setCategoryAxis,AmChart-method}};
  \code{\link{setCategoryField}},
  \code{\link{setCategoryField,AmChart,character-method}};
  \code{\link{setChartCursor}},
  \code{\link{setChartCursor,AmChart-method}};
  \code{\link{setChartScrollbar}},
  \code{\link{setChartScrollbar,AmChart-method}};
  \code{\link{setCreditsPosition}},
  \code{\link{setCreditsPosition,AmChart,character-method}};
  \code{\link{setDataProvider,AmChart,data.frame-method}},
  \code{\link{setDataProvider,DataSet,data.frame-method}};
  \code{\link{setExport}},
  \code{\link{setExport,AmChart-method}};
  \code{\link{setGraph,AmChart-method}},
  \code{\link{setGraph,ChartScrollbar-method}};
  \code{\link{setGraphs}},
  \code{\link{setGraphs,AmChart,list-method}};
  \code{\link{setLegend}},
  \code{\link{setLegend,AmChart-method}};
  \code{\link{setSubChartProperties}},
  \code{\link{setSubChartProperties,AmChart-method}};
  \code{\link{setTitles}},
  \code{\link{setTitles,AmChart,list-method}};
  \code{\link{setTrendLines}},
  \code{\link{setTrendLines,AmChart,list-method}};
  \code{\link{setValueAxes}},
  \code{\link{setValueAxes,AmChart,list-method}};
  \code{\link{setValueAxis,AmChart-method}}
}

