% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{setMainDataSet,AmStockChart-method}
\alias{setMainDataSet,AmStockChart-method}
\title{Setter}
\usage{
\S4method{setMainDataSet}{AmStockChart}(.Object, dataSet = NULL, ...)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{dataSet}{(optionnal) Object of class \code{\linkS4class{DataSet}}.
Data set selected as main.}

\item{...}{Properties of the \code{\linkS4class{DataSet}} to add.}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter
}
\examples{
library(pipeR)
amStockChart() \%>>\% setMainDataSet(showInCompare = TRUE)
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

\code{\linkS4class{DataSet}} S4 class

Other AmStockChart methods: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setDataSetSelector,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setPanels,AmStockChart,list-method}};
  \code{\link{setPanelsSettings,AmStockChart-method}};
  \code{\link{setPeriodSelector,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}};
  \code{\link{setValueAxesSettings,AmStockChart-method}}

Other AmStockChart setters: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setDataSetSelector,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setPanels,AmStockChart,list-method}};
  \code{\link{setPanelsSettings,AmStockChart-method}};
  \code{\link{setPeriodSelector,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}};
  \code{\link{setValueAxesSettings,AmStockChart-method}}
}

