% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartCursor.R
\docType{methods}
\name{setOneBalloonOnly,ChartCursor,logical-method}
\alias{setOneBalloonOnly,ChartCursor,logical-method}
\title{SETTER}
\usage{
\S4method{setOneBalloonOnly}{ChartCursor,logical}(.Object, oneBalloonOnly)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{ChartCursor}}}

\item{oneBalloonOnly}{Object of class \code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}
}
\value{
Updated .Object
}
\description{
SETTER
}
\examples{
library(pipeR)
chartCursor() \%>>\% setOneBalloonOnly(TRUE)
}

