% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{setPanels}
\alias{setPanels}
\title{Setter}
\usage{
\S4method{setPanels}{AmStockChart,list}(.Object, panels)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{panels}{Object of class \code{\linkS4class{list}}.
Each element of this list must be a stockPanel}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter
}
\examples{
library(pipeR)
amStockChart() \%>>\% setPanels(list(stockPanel(compared = TRUE), stockPanel(compared = TRUE)))
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

\code{\linkS4class{StockPanel}} S4 class

Other AmStockChart methods: \code{\link{addComparedDataSet}};
  \code{\link{addDataSet}}; \code{\link{addPanel}};
  \code{\link{setCategoryAxesSettings}};
  \code{\link{setChartCursorSettings}};
  \code{\link{setChartScrollbarSettings}};
  \code{\link{setComparedDataSets}};
  \code{\link{setDataSetSelector}};
  \code{\link{setDataSets}};
  \code{\link{setLegendSettings}};
  \code{\link{setMainDataSet}};
  \code{\link{setPanelsSettings}};
  \code{\link{setPeriodSelector}};
  \code{\link{setStockEventsSettings}};
  \code{\link{setValueAxesSettings}}

Other AmStockChart setters: \code{\link{addComparedDataSet}};
  \code{\link{addDataSet}}; \code{\link{addPanel}};
  \code{\link{setCategoryAxesSettings}};
  \code{\link{setChartCursorSettings}};
  \code{\link{setChartScrollbarSettings}};
  \code{\link{setComparedDataSets}};
  \code{\link{setDataSetSelector}};
  \code{\link{setDataSets}};
  \code{\link{setLegendSettings}};
  \code{\link{setMainDataSet}};
  \code{\link{setPanelsSettings}};
  \code{\link{setPeriodSelector}};
  \code{\link{setStockEventsSettings}};
  \code{\link{setValueAxesSettings}}
}

