% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{setPeriodSelector,AmStockChart-method}
\alias{setPeriodSelector,AmStockChart-method}
\title{Setter for PeriodSelector}
\usage{
\S4method{setPeriodSelector}{AmStockChart}(.Object, periodSelector = NULL,
  ...)
}
\arguments{
\item{.Object}{Object of class \code{\linkS4class{AmStockChart}}.}

\item{periodSelector}{Object of class \code{\linkS4class{PeriodSelector}}.}

\item{...}{Properties of the class \code{\linkS4class{PeriodSelector}}.}
}
\value{
The updated object of class \code{\linkS4class{AmStockChart}}.
}
\description{
Setter for PeriodSelector
}
\details{
Use this methode in case of an AmStockChart.
}
\examples{
library(pipeR)
amStockChart() \%>>\% setPeriodSelector(dateFormat = "DD-MM-YYYY")
}
\seealso{
\code{\linkS4class{AmStockChart}} S4 class

Other AmStockChart methods: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setDataSetSelector,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setMainDataSet,AmStockChart-method}};
  \code{\link{setPanels,AmStockChart,list-method}};
  \code{\link{setPanelsSettings,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}};
  \code{\link{setValueAxesSettings,AmStockChart-method}}

Other AmStockChart setters: \code{\link{addDataSet,AmStockChart-method}};
  \code{\link{addPanel,AmStockChart-method}};
  \code{\link{setDataSetSelector,AmStockChart-method}};
  \code{\link{setLegendSettings,AmStockChart-method}};
  \code{\link{setMainDataSet,AmStockChart-method}};
  \code{\link{setPanels,AmStockChart,list-method}};
  \code{\link{setPanelsSettings,AmStockChart-method}};
  \code{\link{setStockEventsSettings,AmStockChart-method}};
  \code{\link{setValueAxesSettings,AmStockChart-method}}
}

