% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{setStockEvents,DataSet,list-method}
\alias{setStockEvents,DataSet,list-method}
\title{SETTER}
\usage{
\S4method{setStockEvents}{DataSet,list}(.Object, stockEvents)
}
\description{
SETTER
}

