% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockEvent.R
\docType{methods}
\name{setStockGraph,StockEvent-method}
\alias{setStockGraph,StockEvent-method}
\title{SETTER}
\usage{
\S4method{setStockGraph}{StockEvent}(.Object, stockGraph = NULL, ...)
}
\arguments{
\item{stockGraph}{Object of class \code{list}.
Containing properties of stockGraph.
This is the graph on which event will be displayed.
You can use a reference to the stock graph object or id of the graph.}

\item{...}{}
}
\description{
SETTER
}
\examples{
library(pipeR)
stockEvent() \%>>\% setStockGraph(stockGraph = stockGraph(balloonText = "balloonText"))
stockEvent() \%>>\% setStockGraph(stockGraph = "stockGraph1")
}

