% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockPanel.R
\docType{methods}
\name{setStockLegend,StockPanel-method}
\alias{setStockLegend,StockPanel-method}
\title{SETTER}
\usage{
\S4method{setStockLegend}{StockPanel}(.Object, stockLegend = NULL, ...)
}
\description{
SETTER
}
\examples{
stockPanel() \%>>\% setStockLegend( stockLegend() )
}

