% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{setText,Label,character-method}
\alias{setText,Label,character-method}
\title{SETTER}
\usage{
\S4method{setText}{Label,character}(.Object, text)
}
\description{
SETTER
}
\examples{
library(pipeR)
label() \%>>\% setText("Bonjour")
}

