% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart_getSet.R
\docType{methods}
\name{setType,AmChart,character-method}
\alias{setType,AmChart,character-method}
\title{Setter for Type}
\usage{
\S4method{setType}{AmChart,character}(.Object, type)
}
\arguments{
\item{.Object}{object of class \code{\linkS4class{AmChart}}.}

\item{type}{object of class \code{\linkS4class{character}}.}
}
\description{
Setter for Type
}
\examples{
setType(.Object = amChart(), type = "pie")
# equivalent to
amPieChart()
}

