% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartScrollbar.R
\docType{methods}
\name{setUpdateOnReleaseOnly,ChartScrollbar,logical-method}
\alias{setUpdateOnReleaseOnly,ChartScrollbar,logical-method}
\title{SETTER}
\usage{
\S4method{setUpdateOnReleaseOnly}{ChartScrollbar,logical}(.Object,
  updateOnReleaseOnly)
}
\arguments{
\item{.Object}{}

\item{updateOnReleaseOnly}{Object of class \code{logical}.
Specifies if the chart should be updated while dragging/resizing the scrollbar
or only at the moment when user releases mouse button.}
}
\description{
SETTER
}
\examples{
library(pipeR)
chartScrollbar() \%>>\% setUpdateOnReleaseOnly(TRUE)
}

