% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmLegend.R
\docType{methods}
\name{setUseGraphSettings,AmLegend,logical-method}
\alias{setUseGraphSettings,AmLegend,logical-method}
\title{SETTER}
\usage{
\S4method{setUseGraphSettings}{AmLegend,logical}(.Object, useGraphSettings)
}
\arguments{
\item{.Object}{}

\item{useGraphSettings}{Object of class \code{logical}.
If this is set to TRUE, border color instead of background color will be changed when
user rolls-over the slice, graph, etc.}
}
\description{
SETTER
}
\examples{
library(pipeR)
amLegend() \%>>\% setUseGraphSettings(TRUE)
}

