% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrendLine.R
\docType{methods}
\name{setValueAxis,TrendLine-method}
\alias{setValueAxis,TrendLine-method}
\title{SETTER}
\usage{
\S4method{setValueAxis}{TrendLine}(.Object, valueAxis = NULL, ...)
}
\description{
SETTER
}
\examples{
library(pipeR)
setValueAxis(trendLine(), valueAxis(title = "Hello !", axisTitleOffset = 12))
setValueAxis(trendLine(), title = "Hello !", axisTitleOffset = 12)
}

