% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChartCursor.R
\docType{methods}
\name{setValueLineAxis,ChartCursor-method}
\alias{setValueLineAxis,ChartCursor-method}
\title{SETTER}
\usage{
\S4method{setValueLineAxis}{ChartCursor}(.Object, valueLineAxis = NULL, ...)
}
\arguments{
\item{.Object:}{\code{\linkS4class{ChartCursor}}.}

\item{valueLineAxis:}{object of class \code{\linkS4class{ValueAxis}}.}
}
\description{
SETTER
}
\examples{
library(pipeR)
chartCursor() \%>>\% setValueLineAxis( title = "Hello !", axisTitleOffset = 12 )
}

