% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.R
\docType{methods}
\name{setX,Label,numeric-method}
\alias{setX,Label,numeric-method}
\title{SETTER}
\usage{
\S4method{setX}{Label,numeric}(.Object, x)
}
\description{
SETTER
}
\examples{
label() \%>>\% setX(16)
}

