% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmChart.R
\docType{methods}
\name{show,AmChart-method}
\alias{show,AmChart-method}
\title{Visualize AmChart with show}
\usage{
\S4method{show}{AmChart}(object)
}
\arguments{
\item{object}{\linkS4class{AmChart}}
}
\description{
Visualize AmChart with show
}
\examples{
pipeR::pipeline(
  rAmCharts::amPieChart(valueField = "value", titleField = "key"),
  rAmCharts::setDataProvider(data.frame(key = c("FR", "US"), value = c(20,10))),
  rAmCharts::setExport(position = "bottom-left")
)
}
\seealso{
Other Visualizations: \code{\link{print,AmObject-method}};
  \code{\link{show,AmObject-method}};
  \code{\link{show,AmStockChart-method}}
}

