% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmChart.R, R/class_AmStockChart.R
\docType{methods}
\name{show,AmChart-method}
\alias{show,AmChart-method}
\title{Visualize AmChart with show}
\usage{
\S4method{show}{AmChart}(object)

\S4method{show}{AmChart}(object)
}
\arguments{
\item{object}{\linkS4class{AmStockChart}}
}
\description{
Display the object in the console

Display the object in the console
}
\examples{
pipeR::pipeline(
  amPieChart(valueField = "value", titleField = "key"),
  setDataProvider(data.frame(key = c("FR", "US"), value = c(20,10))),
  setExport(position = "bottom-left")
)

}

