% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_AmCharts.R
\docType{methods}
\name{show,AmChart-method}
\alias{show,AmChart-method}
\title{Visualize AmStockChart with show}
\usage{
\S4method{show}{AmChart}(object)
}
\arguments{
\item{object}{\linkS4class{AmChart}.}
}
\value{
If the object has a valid type, it will plot the chart.
If not the method will trigger the method 'print'.
}
\description{
Display the object in the console.
}

