% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmObject.R
\docType{methods}
\name{show,AmObject-method}
\alias{show,AmObject-method}
\title{Visualize with show}
\usage{
\S4method{show}{AmObject}(object)
}
\arguments{
\item{object}{\linkS4class{AmObject}}
}
\examples{
pipeR::pipeline(
  rAmCharts::amPieChart(valueField = "value", titleField = "key", creditsPosition = "top-right",
                        backgroundColor = "#7870E8"),
  rAmCharts::setDataProvider(data.frame(key = c("FR", "US"), value = c(20,10))),
  rAmCharts::setExport(position = "bottom-left")
)
}
\seealso{
Other Visualizations: \code{\link{print,AmObject-method}},
  \code{\link{show,AmChart-method}},
  \code{\link{show,AmStockChart-method}}
}

