% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmObject.R
\docType{methods}
\name{show,AmObject-method}
\alias{show,AmObject-method}
\title{Visualize with show}
\usage{
\S4method{show}{AmObject}(object)
}
\arguments{
\item{object}{\linkS4class{AmObject}}
}
\description{
Visualize with show
}
\examples{
show(amSerialChart())
}
\seealso{
Other Visualizations: \code{\link{print,AmObject-method}}
}

