% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmObject.R
\docType{methods}
\name{show,AmObject-method}
\alias{show,AmObject-method}
\title{Visualize with show}
\usage{
\S4method{show}{AmObject}(object)
}
\arguments{
\item{object}{\linkS4class{AmObject}.}
}
\description{
Display the object in the console.
}
\examples{
library(pipeR)
amPieChart(valueField = "value", titleField = "key", backgroundColor = "#7870E8") \%>>\%
  setDataProvider(data.frame(key = c("FR", "US"), value = c(20,10))) \%>>\%
  setExport(position = "bottom-left")
}

