% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmStockChart.R
\docType{methods}
\name{show,AmStockChart-method}
\alias{show,AmStockChart-method}
\title{Visualize AmStockChart with show}
\usage{
\S4method{show}{AmStockChart}(object)
}
\arguments{
\item{object}{\linkS4class{AmStockChart}}
}
\description{
Display the object in the console
}

