% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StockEvent.R
\docType{class}
\name{StockEvent-class}
\alias{StockEvent-class}
\title{StockEvent class}
\description{
StockEvent class
}
\section{Slots}{

\describe{
\item{\code{\code{stockGraph}:}}{Object of class \code{list}.
Containing properties of stockGraph.
This is the graph on which event will be displayed.
You can use a reference to the stock graph object or id of the graph.}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

