% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_StockEvent.R
\docType{class}
\name{StockEvent-class}
\alias{StockEvent-class}
\title{StockEvent class}
\description{
StockEvent is object which holds information about event (bullet).
Values from StockEventsSettings will be used if not set.
Stock event bullet's size depends on it's graphs fontSize.
When user rolls-over, clicks or rolls-out of the event bullet, AmStockChart dispatches events.
}
\details{
Run \code{api("StockEvent")} for more information.
}
\section{Slots}{

\describe{
\item{\code{stockGraph}}{\linkS4class{AmGraph} containing properties of stockGraph.
This is the graph on which event will be displayed.
You can use a reference to the stock graph object or id of the graph.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

