% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockEvent.R
\name{stockEvent}
\alias{stockEvent}
\title{#’ Constructor.}
\usage{
stockEvent(backgroundAlpha = 1, stockGraph, ...)
}
\arguments{
\item{\code{...}:}{{Properties of StockEvent.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/StockEvent}}}}
}
\value{
An \code{\linkS4class{StockEvent}} object
}
\description{
#’ Constructor.
}
\examples{
stockEvent()
}

