% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AmLegend.R
\name{stockLegend}
\alias{stockLegend}
\title{Constructor for StockLegend.}
\usage{
stockLegend(useGraphSettings, valueTextComparing = "[[percents.value]]\%",
  ...)
}
\arguments{
\item{useGraphSettings}{Legend markers can mirror graph's settings,
displaying a line and a real bullet as in the graph itself.
Set this property to true if you want to enable this feature.}

\item{valueTextComparing}{\code{character}}

\item{...}{Properties of AmLegend.
See \url{http://docs.amcharts.com/3/javascriptstockchart/StockLegend}}
}
\value{
An \code{\linkS4class{AmLegend}} object
}
\description{
Constructor for StockLegend.
}
\examples{
stockLegend(useGraphSettings = TRUE)
}

