% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AmLegend.R
\name{stockLegend}
\alias{stockLegend}
\title{Constructor for StockLegend.}
\usage{
stockLegend(useGraphSettings, valueTextComparing = "[[percents.value]]\%", ...)
}
\arguments{
\item{useGraphSettings}{\code{logical}
Legend markers can mirror graph's settings,
displaying a line and a real bullet as in the graph itself.
Set this property to TRUE if you want to enable this feature.}

\item{valueTextComparing}{\code{character}}

\item{...}{Properties of AmLegend.
See \url{http://docs.amcharts.com/3/javascriptstockchart/StockLegend}}
}
\value{
An \linkS4class{AmLegend} object
}
\description{
This method is used for \linkS4class{AmStockChart}.
}
\examples{
stockLegend(useGraphSettings = TRUE)

}
