% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StockPanel.R
\docType{class}
\name{StockPanel-class}
\alias{StockPanel-class}
\title{StockPanel class}
\description{
StockPanel class
}
\section{Fields}{

\describe{
\item{\code{drawOnAxis}}{\code{list}. Containing properties of ValueAxis.
Specifies on which value axis user can draw trend lines.
Set drawingIconsEnabled to true if you want drawing icons to be visible.
First value axis will be used if not set here.
You can use a reference to the value axis object or id of value axis.}

\item{\code{stockGraphs}}{\code{list}.
Each element must be have been created with stockGraph(*)}

\item{\code{stockLegend}}{\code{list}.
Each element must be have been created with stockLegend(*)}

\item{\code{allLabels}}{\code{list} of \linkS4class{Label} properties.}

\item{\code{arrows}}{\code{list} of \linkS4class{GaugeArrow} properties.}

\item{\code{axes}}{\code{list} of \linkS4class{GaugeAxis} properties.}

\item{\code{balloon}}{\code{list} of an \linkS4class{AmBalloon} properties.
Creates the balloons (tooltips) of the chart,
It follows the mouse cursor when you roll-over the data items.
The framework generates the instances automatically you only need to adjust
the appearance to your needs.}

\item{\code{categoryAxis}}{\code{list} of a \linkS4class{CategoryAxis} properties.
Read-only. Chart creates category axis itself.
If you want to change some properties,
you should get this axis from the chart and set properties to this object.}

\item{\code{categoryField}}{\code{character}.
Category field name tells the chart the name of the field in your dataProvider object
which will be used for category axis values.}

\item{\code{ChartCursor}}{\code{list} of a \linkS4class{ChartCursor} properties.}

\item{\code{ChartScrollbar}}{\code{list} of a \linkS4class{ChartScrollbar} properties.}

\item{\code{creditsPosition}}{\code{character},
specifying position of link to amCharts site.
Allowed values are: top-left, top-right, bottom-left and bottom-right.}

\item{\code{dataProvider}}{\code{list}, containing the data.
Use providing method toList* to convert a \code{data.frame}.}

\item{\code{graphs}}{\code{list}.  List of AmGraphs properties
See \linkS4class{AmGraph}.
Creates the visualization of the data in following types: line, column, step line,
smoothed line, olhc and candlestick.}

\item{\code{graph}}{\code{list} of \linkS4class{AmGraph} properties.
Graph of a Gantt chart. Gant chart actually creates multiple graphs (separate for each segment).
Properties of this graph are passed to each of the created graphs
- this allows you to control the look of segments.}

\item{\code{guides}}{\code{list} of Guides properties.
See \linkS4class{Guide}.
Instead of adding guides to the axes, you can push all of them to this array.
In case guide has category or date defined, it will automatically will be assigned to the category axis.
Otherwise to first value axis, unless you specify a different valueAxes for the guide.}

\item{\code{legend}}{\code{list} of an \linkS4class{AmLegend} properties.}

\item{\code{segmentsField}}{\code{character}.}

\item{\code{subChartProperties}}{\code{list}.}

\item{\code{theme}}{\code{character}.
Theme of a chart. Config files of themes can be found in amcharts/themes/ folder.}

\item{\code{titles}}{\code{list} of Titles properties.
See \linkS4class{Title}.}

\item{\code{trendLines}}{\code{list} of \linkS4class{TrendLine}.
You can add trend lines to a chart using this list or access already existing trend lines.}

\item{\code{type}}{\code{character}.
Possible types are: serial, pie, radar,
(types xy, radar, funnel, gauge, map, stock. are in development).}

\item{\code{valueAxes}}{Object of class \code{list}. List of ValueAxes' properties.
See \linkS4class{ValueAxis}.
Chart creates one value axis automatically,
so if you need only one value axis, you don't need to create it.}

\item{\code{valueAxis}}{\code{list} of Value axis properties for Gantt chart.
Set it's type to "date" if your data is date or time based.
In case of Value axis for a Gantt chart. Set it's type to "date" if your data is date or time based.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string. See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties for the class.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

