% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Title.R
\docType{class}
\name{Title-class}
\alias{Title-class}
\title{Title class}
\description{
Title class
}
\section{Slots}{

\describe{
\item{\code{text}}{\code{character}.
Text of a title.}

\item{\code{size}}{\code{numeric}.
Text size of a title.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

