% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Title.R
\docType{methods}
\name{initialize,Title-method}
\alias{amTitle}
\alias{initialize,Title-method}
\alias{setSize}
\alias{setSize,Title,numeric-method}
\alias{setText,Title,character-method}
\alias{title}
\title{Initializes A Title}
\usage{
\S4method{initialize}{Title}(.Object, text, size, ...)

title(text, size, ...)

amTitle(text, size, ...)

\S4method{setText}{Title,character}(.Object, text)

setSize(.Object, size)

\S4method{setSize}{Title,numeric}(.Object, size)
}
\arguments{
\item{.Object}{\linkS4class{Title}}

\item{text}{\code{character}, title text.}

\item{size}{\code{numeric}, title size.}

\item{...}{other properties of Title.}
}
\value{
(updated) \linkS4class{Title}
}
\description{
Uses the constructor to create the object
or update an existing one with the setters.
}
\examples{
new("Title", size = 10)
rAmCharts:::title(text = "Main", size = 10)
rAmCharts:::title(text = "Main", bold = TRUE)
amTitle(text = "Main", size = 10)
amTitle(text = "Main", bold = TRUE)
setText(.Object = amTitle(), text = "Bonjour")
setSize(amTitle(), 16)
}

