% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Title.R
\name{title}
\alias{title}
\title{Constructor for an AmGraph}
\usage{
title(text, size, ...)
}
\arguments{
\item{...}{Properties of Title.
See \url{http://docs.amcharts.com/3/javascriptcharts/Title}}
}
\value{
An \code{\linkS4class{Title}} object
}
\description{
Constructor for an AmGraph
}
\examples{
new("Title", text = "bonjour")
title(text = "bonjour")
}

