% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Title.R
\name{title}
\alias{title}
\title{#’ Constructor.}
\usage{
title(text, size, ...)
}
\arguments{
\item{\code{...}:}{{Properties of Title.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/Title}}}}
}
\value{
An \code{\linkS4class{Title}} object
}
\description{
#’ Constructor.
}
\examples{
new("Title", text = "bonjour")
title(text = "bonjour")
}

