% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrendLine.R
\docType{class}
\name{TrendLine-class}
\alias{TrendLine-class}
\title{TrendLine class}
\description{
TrendLine class
}
\section{Slots}{

\describe{
\item{\code{finalValue}}{\code{numeric}.
Value at which trend line should end.}

\item{\code{finalXValue}}{\code{numeric}.
Used by XY chart only. X value at which trend line should end.}

\item{\code{initialValue}}{\code{numeric}.
Value from which trend line should start.}

\item{\code{initialXValue}}{\code{numeric}.
Used by XY chart only. X value from which trend line should start.}

\item{\code{valueAxis}}{\linkS4class{ValueAxis}.
Value axis of the trend line. Will use first value axis of the chart if not set any.
You can use a reference to the value axis object or id of value axis.}

\item{\code{valueAxisX}}{\linkS4class{ValueAxis}.
Used by XY chart only. X axis of trend line.
Will use first X axis of the chart if not set any.
You can use a reference to the value axis object or id of value axis.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list},
containing other avalaible properties non coded in the package yet.}

\item{\code{value}}{\code{numeric}.}
}}
\author{
DataKnowledge
}

