% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ValueAxis.R
\docType{class}
\name{ValueAxis-class}
\alias{ValueAxis-class}
\title{ValueAxis class}
\description{
ValueAxis class
}
\section{Slots}{

\describe{
\item{\code{\code{title}:}}{{ Object of class \code{character}. Title of the axis..}}
}}
\section{Slots}{

}
\author{
DataKnowledge
}

