% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ValueAxis.R
\docType{class}
\name{ValueAxis-class}
\alias{ValueAxis-class}
\title{ValueAxis class}
\description{
Extension for ValueAxis to create an axis for amSerialChart, amRadarChart,
amXYChart charts, multiple can be assigned.
Gets automatically populated, one for amSerialChart and two for amXYChart charts,
if none has been specified.
}
\details{
Run \code{api("ValueAxis")} for more information and all avalaible properties.
}
\section{Slots}{

\describe{
\item{\code{title}}{\code{character}. Title of the axis.}

\item{\code{guides}}{\code{list}.}

\item{\code{listeners}}{\code{list} containining the listeners to add to the object.
The list must be named as in the official API. Each element must be a character string.
See examples for details.}

\item{\code{otherProperties}}{\code{list}
containing other avalaible properties not yet coded in the package.}

\item{\code{value}}{\code{numeric}.
Guides belonging to this axis. Use addGuide method}
}}
\author{
DataKnowledge
}

