% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ValueAxis.R
\name{valueAxis}
\alias{valueAxis}
\title{#’ Constructor.}
\usage{
valueAxis(title, ...)
}
\arguments{
\item{\code{...}:}{{Properties of ValueAxis.
See \code{\url{http://docs.amcharts.com/3/javascriptcharts/ValueAxis}}}}
}
\value{
An \code{\linkS4class{ValueAxis}} object
}
\description{
#’ Constructor.
}
\examples{
valueAxis(title = "Hello !", axisTitleOffset = 12)
}

